### 2024-03-26 #6542 @editorialbot
[PRE REVIEW]: REHO: A Decision Support Tool for Renewable Energy Communities
**Submitting author:** @DorsanL (<a href="http://orcid.org/0009-0008-7309-5537">Dorsan Lepour</a>)
**Repository:** https://github.com/IPESE/REHO
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.2
