### 2024-03-26 8a561ec
Line only in model, Tr in both, both working + write results df_Grid without GWP
