### 2024-03-26 e03a6b8
cleaned write_results
