+++ reho/paths.py
@@ -59,6 +59,7 @@ path_to_configurations = os.path.join(os.getcwd(), 'configurations')
 path_sia = os.path.join(path_to_sia, 'sia2024_rooms_sia380_1.csv')
 path_norms = os.path.join(path_to_sia, 'sia2024_data.xlsx')
 
+
 def path_handler(path_given):
     """To handle the path to csv file, absolute path or not"""
 
