+++ reho/model/preprocessing/EV_profile_generator.py
@@ -4,7 +4,7 @@ import pandas as pd
 import numpy as np
 
 
-def generate_EV_plugged_out_profiles_district(cluster):
+def generate_EV_plugged_out_profiles_district(cluster, timestamp_data):
     """
     Computes hourly electric vehicle (EV) profiles for each typical day considering weekdays and weekends. Data are taken from
     `UK Department for Transport 2013 <https://www.gov.uk/government/collections/energy-and-environment-statistics#publications>`_
@@ -43,7 +43,7 @@ def generate_EV_plugged_out_profiles_district(cluster):
                                skiprows=1)
         timestamp = pd.DataFrame(timestamp, columns=("Day", "Frequency", "Weekday"))
     else:
-        df = pd.read_csv(os.path.join(path_to_clustering, 'timestamp_' + File_ID + '.dat'), delimiter='\t')
+        df = timestamp_data #pd.read_csv(os.path.join(path_to_clustering, 'timestamp_' + File_ID + '.dat'), delimiter='\t')
         timestamp = df.fillna(1)  # only weekdays
 
     # Federal Office of Statistic, Comportement de la population en matiere de transports, 2015
