### 2024-03-27 63bb452
lausanne script
