+++ scripts/lausanne/1a_lausanne_capex.py
@@ -1,0 +1,32 @@
+from reho.model.reho import *
+
+
+if __name__ == '__main__':
+
+    # Set building parameters
+    reader = QBuildingsReader()
+    qbuildings_data = reader.read_csv(buildings_filename='../lausanne/QBuildings/Lausanne_sectors.csv', nb_buildings=2)
+
+    # Select weather data
+    cluster = {'Location': 'Pully', 'Attributes': ['I', 'T', 'W'], 'Periods': 10, 'PeriodDuration': 24}
+
+    # Set scenario
+    scenario = dict()
+    scenario['Objective'] = 'TOTEX'
+    scenario['name'] = 'totex'
+    scenario['exclude_units'] = ['Battery', 'NG_Cogeneration']
+    scenario['enforce_units'] = []
+
+    # Initialize available units and grids
+    grids = infrastructure.initialize_grids()
+    units = infrastructure.initialize_units(scenario, grids)
+
+    # Set method options
+    method = {'building-scale': True}
+
+    # Run optimization
+    reho = reho(qbuildings_data=qbuildings_data, units=units, grids=grids, cluster=cluster, scenario=scenario, method=method, solver="gurobi")
+    reho.single_optimization()
+
+    # Save results
+    reho.save_results(format=['xlsx', 'pickle'], filename='1a')
