### 2024-03-27 b97ae70
update lausanne script
