### 2024-03-28 #6542 @DorsanL
Dear @kyleniemeyer,

Thank you for launching the pre-review process ! 🙏

### Editor’s choice

* As most appropriate editors, I was obviously thinking of [fraukewiese](https://github.com/fraukewiese) and [AdamRJensen](https://github.com/adamrjensen/).
* [boisgera](https://github.com/boisgera) through its expertise in optimization, Python and software engineering could also be considered?
* [sjpfenninger](https://github.com/sjpfenninger) and [timtroendle](https://github.com/timtroendle) certainly have the relevant expertise, but they are now listed as emeritus editors…

### Reviewers choice

After examining the list, it seems to me that the people listed below would be an appropriate choice:
* [samuelduchesne](https://github.com/samuelduchesne)
* [nmstreethran](https://github.com/nmstreethran)
* [willu47](https://github.com/willu47)
* [nick-gorman](https://github.com/nick-gorman)
* [olejandro](https://github.com/olejandro)
* [EwaGomez](https://github.com/EwaGomez)
* [FlorianK13](https://github.com/FlorianK13)
* [BenWinchester](https://github.com/BenWinchester)
* [FWuellhorst](https://github.com/FWuellhorst)
* [jtock](https://github.com/jtock)
* [GregorBecker](https://github.com/GregorBecker)
* [mfleschutz](https://github.com/mfleschutz)

### Related JOSS papers

Below is a list of recent JOSS publications (along with their reviewers) dealing with subjects similar to REHO:

* [The Spreadsheet Energy System Model Generator (SESMG): A tool for the optimization of urban energy systems](https://joss.theoj.org/papers/10.21105/joss.05519)
     * [nick-gorman](https://github.com/nick-gorman)
     * [willu47](https://github.com/willu47)

* [CLOVER: A modelling framework for sustainable community-scale energy systems](https://joss.theoj.org/papers/10.21105/joss.04799)
     * [EwaGomez](https://github.com/EwaGomez)
     * [fneum](https://github.com/fneum) _(not in the reviewer list anymore)_

* [CESAR-P: A dynamic urban building energy simulation tool](https://joss.theoj.org/papers/10.21105/joss.04261)
     * [olejandro](https://github.com/olejandro)
     * [noah80](https://github.com/noah80)

* [PyHeatDemand - Processing Tool for Heat Demand Data](https://joss.theoj.org/papers/10.21105/joss.06275)
     * [nmstreethran](https://github.com/nmstreethran)
     * [samuelduchesne](https://github.com/samuelduchesne)

* [RHEIA: Robust design optimization of renewable Hydrogen and dErIved energy cArrier systems](https://joss.theoj.org/papers/10.21105/joss.04370)
     * [andr1976](https://github.com/andr1976) _(not in the reviewer list anymore)_
     * [ClaraBuettner](https://github.com/ClaraBuettner) _(not in the reviewer list anymore)_

* [otoole: OSeMOSYS Tools for Energy Work](https://joss.theoj.org/papers/10.21105/joss.05511)
     * [olejandro](https://github.com/olejandro)
     * [fneum](https://github.com/fneum) _(not in the reviewer list anymore)_

___________

Thank you for your attention to our review, and we look forward to hearing from you! 👌
