+++ reho/model/postprocessing/KPIs.py
@@ -309,8 +309,8 @@ def postcompute_average_emission(df_annual, df_annual_net, df_profiles, df_profi
     df_el_net = df_profiles_net.xs('Electricity', level=0)
 
     File_ID = WD.get_cluster_file_ID(cluster)
-    res_profile = emissions.return_typical_emission_profiles(df_Time, File_ID, timestamp_file,'method 1', emissions_matrix)
-    res_av = emissions.find_average_value('CH', 'method 1',emissions_matrix)
+    res_profile = emissions.return_typical_emission_profiles(df_Time, File_ID, 'method 1', timestamp_file, emissions_matrix)
+    res_av = emissions.find_average_value('CH', 'method 1', emissions_matrix)
     s_RES_dy = pd.Series(dtype='float')
     s_RES_av = pd.Series(dtype='float')
 
