+++ reho/model/postprocessing/write_results.py
@@ -1,5 +1,6 @@
 import pandas as pd
 import numpy as np
+import logging
 
 __doc__ = """
 *Extracts the results from the AMPL model and converts it to Python dictionary and pandas dataframes.*
@@ -60,7 +61,7 @@ def get_df_Results_from_SP(ampl, scenario, method, buildings_data, filter=True):
 
         df_Performance = pd.concat([df_Performance, df_Epsilon], axis=1)
         df_Performance.index.names = ['Hub']
-        print(df_Performance)
+        logging.info(df_Performance)
 
         return df_Performance.sort_index()
 
@@ -130,7 +131,7 @@ def get_df_Results_from_SP(ampl, scenario, method, buildings_data, filter=True):
         df_Unit = pd.concat([df1, df2, df3, df4, df5], axis=1)
         df_Unit.index.names = ['Unit']
         df_Unit = df_Unit.sort_index()
-        print(df_Unit)
+        logging.info(df_Unit)
 
         # Unit_t
         df1 = get_variable_in_pandas(df, 'Units_demand')
@@ -355,7 +356,7 @@ def get_df_Results_from_SP(ampl, scenario, method, buildings_data, filter=True):
             try:
                 parameters_record[p] = ampl.getData(p).toPandas()
             except:
-                print(p)
+                logging.info(p)
 
     if filter:
         for df_name, df in df_Results.items():
@@ -493,7 +494,7 @@ def get_df_Results_from_MP(ampl, binary=False, method=None, district=None, read_
         df_Unit = pd.concat([df_Unit, df_DHN], axis=0)
     df_Unit.index.names = ['Unit']
     df_Results["df_Unit"] = df_Unit.sort_index()
-    print(df_Results["df_Unit"])
+    logging.info(df_Results["df_Unit"])
 
     # Unit_t
     df1 = get_variable_in_pandas(df, 'Units_demand')
