### 2024-04-03 e26b3ab
add method save_all_df (if False save only essential results)
