+++ reho/model/reho.py
@@ -701,35 +701,34 @@ class reho(district_decomposition):
         df = self.get_final_SPs_results(MP_selection, 'df_Grid_t')
         df = df.droplevel(['Scn_ID', 'Pareto_ID', 'Iter', 'FeasibleSolution', 'house'])
         df = df.sort_index(level='Hub')
-        uncontrollable_load = df.groupby(["Layer", "Period", "Time"]).sum()["Uncontrollable_load"]
 
         h_op = df_Time.dp
         h_op.iloc[-2:] = 1
         df_network = last_results["df_District_t"].copy()
-        df_network[["Network_supply", "Network_demand"]] = df_network[["Network_supply", "Network_demand"]].divide(h_op,
-                                                                                                                   axis=0,
-                                                                                                                   level='Period')
-        df_network["Uncontrollable_load"] = uncontrollable_load
-        df_network = pd.concat([df_network], keys=['Network'], names=['Hub']).reorder_levels(
-            ['Layer', 'Hub', 'Period', 'Time'])
-        df_network = df_network.rename(
-            columns={"Cost_demand_network": "Cost_demand", "Cost_supply_network": "Cost_supply",
-                     "Network_demand": "Grid_demand", "Network_supply": "Grid_supply"})
+        df_network[["Network_supply", "Network_demand"]] = df_network[["Network_supply", "Network_demand"]].divide(h_op, axis=0, level='Period')
+
+        if self.method["save_all_df"]:
+            df_network["Uncontrollable_load"] = df.groupby(["Layer", "Period", "Time"]).sum()["Uncontrollable_load"]
+
+        df_network = pd.concat([df_network], keys=['Network'], names=['Hub']).reorder_levels(['Layer', 'Hub', 'Period', 'Time'])
+        df_network = df_network.rename(columns={"Cost_demand_network": "Cost_demand",
+                                                "Cost_supply_network": "Cost_supply",
+                                                "Network_demand": "Grid_demand",
+                                                "Network_supply": "Grid_supply"})
+        columns = ["Cost_demand", "Cost_supply"]
+        if self.method["save_all_df"]:
+            columns = columns + ["GWP_demand", "GWP_supply"]
+
         for h in self.buildings_data.keys():
-            for column in ["Cost_demand", "Cost_supply", "GWP_demand", "GWP_supply"]:
+            for column in columns:
                 df.loc[pd.IndexSlice[:, h, :, :], column] = df_network[column].values
+
         df_Grid_t = pd.concat([df, df_network])
 
         # df_Unit
         df_Unit = self.get_final_MP_results(Pareto_ID=Pareto_ID, Scn_ID=Scn_ID)
         df_Unit = df_Unit.droplevel(['FeasibleSolution', 'Hub'])
         df_Unit = df_Unit.sort_index(level='Unit')
-
-        # df_Unit_t
-        df_Unit_t = self.get_final_SPs_results(MP_selection, 'df_Unit_t')
-        df_Unit_t = df_Unit_t.droplevel(['Scn_ID', 'Pareto_ID', 'Iter', 'FeasibleSolution', 'house'])
-        df_district_units = last_results["df_Unit_t"]
-        df_Unit_t = pd.concat([df_Unit_t, df_district_units])
 
         # df_Annuals
         df = self.get_final_SPs_results(MP_selection, 'df_Annuals')
@@ -738,8 +737,7 @@ class reho(district_decomposition):
         df = df.sort_index(level='Layer')
[…]
