+++ reho/model/postprocessing/write_results.py
@@ -155,36 +155,25 @@ def get_df_Results_from_SP(ampl, scenario, method, buildings_data, filter=True):
     def set_df_grid(df, ampl):
         # Grid_t
         df1 = get_variable_in_pandas(df, 'Grid_demand')
-        df2 = get_variable_in_pandas(df, 'Grid_supply')
-        df_cs = get_parameter_in_pandas(ampl, 'Cost_supply', multi_index=True)
-        df_cs = df_cs.reorder_levels((1, 0, 2, 3))
-        df_cd = get_parameter_in_pandas(ampl, 'Cost_demand', multi_index=True)
-        df_cd = df_cd.reorder_levels((1, 0, 2, 3))
-        df_electricity = get_parameter_in_pandas(ampl, 'Domestic_electricity', multi_index=True)
+        hub = df1.index.levels[1].values[0]
+        df1 = df1.droplevel(1)
+        df2 = get_variable_in_pandas(df, 'Grid_supply').droplevel(1)
+
+        df_cs = get_parameter_in_pandas(ampl, 'Cost_supply', multi_index=True).droplevel(0)
+        df_cd = get_parameter_in_pandas(ampl, 'Cost_demand', multi_index=True).droplevel(0)
+
+        df_electricity = get_parameter_in_pandas(ampl, 'Domestic_electricity', multi_index=True).droplevel(0)
         df_electricity.columns = ['Uncontrollable_load']
         df_electricity = df_electricity.set_index([pd.Index(["Electricity"]*df_electricity.index.size), df_electricity.index])
-        df12 = pd.concat([df1, df2, df_cs, df_cd, df_electricity], axis=1)
 
-        df3 = get_variable_in_pandas(df, 'Network_demand')
-        df4 = get_variable_in_pandas(df, 'Network_supply')
-        df_cs = get_parameter_in_pandas(ampl, 'Cost_supply_network', multi_index=True)
-        df_cd = get_parameter_in_pandas(ampl, 'Cost_demand_network', multi_index=True)
         df_es = get_parameter_in_pandas(ampl, 'GWP_supply', multi_index=True)
         df_ed = get_parameter_in_pandas(ampl, 'GWP_demand', multi_index=True)
 
-        # Rename Network to Grid and therefore add Level 'Hub' = Network
-        df3.columns = ['Grid_demand']
-        df4.columns = ['Grid_supply']
-        df_cs.columns = ['Cost_supply']
-        df_cd.columns = ['Cost_demand']
-
-        df34 = pd.concat([df3, df4, df_cs, df_cd, df_es, df_ed], axis=1)
-
-        df34['Hub'] = 'Network'
-        df34.set_index('Hub', append=True, inplace=True)
-        df34 = df34.reorder_levels((0, 3, 1, 2))
+        df12 = pd.concat([df1, df2, df_cs, df_cd, df_es, df_ed, df_electricity], axis=1)
+        df12["Hub"] = hub
+        df12.set_index('Hub', append=True, inplace=True)
+        df_Grid_t = df12.reorder_levels((0, 3, 1, 2))
         # combine Network & Grid dataframe
-        df_Grid_t = pd.concat([df12, df34], sort=True)
         df_Grid_t.index.names = ['Layer', 'Hub', 'Period', 'Time']
 
         return df_Grid_t.sort_index()
@@ -339,13 +328,23 @@ def get_df_Results_from_SP(ampl, scenario, method, buildings_data, filter=True):
     df.columns = ["Varname", "Value"]
     df_Results["df_Performance"] = set_df_performance(df, ampl, scenario)
[…]
