### 2024-04-03 860101f
add method disable prints
