### 2024-04-03 af06f6e
remove function get_variable and get object from ampl directly
