+++ reho/model/postprocessing/write_results.py
@@ -8,46 +8,47 @@ __doc__ = """
 
 def get_df_Results_from_SP(ampl, scenario, method, buildings_data, filter=True):
 
-    def set_df_performance(df, ampl, scenario):
-        df1 = get_variable_in_pandas(df, 'Costs_House_op')  # without the comfort penalty costs
+    def set_df_performance(ampl, scenario):
+        df1 = get_parameter_in_pandas(ampl, 'Costs_House_op')  # without the comfort penalty costs
         df1 = df1.rename(columns={'Costs_House_op': 'Costs_op'})
 
-        df2 = get_variable_in_pandas(df, 'Costs_House_inv')
+        df2 = get_parameter_in_pandas(ampl, 'Costs_House_inv')
         df2 = df2.rename(columns={'Costs_House_inv': 'Costs_inv'})
         tau = ampl.getParameter('tau').getValues().toList()
         df2['Costs_inv'] = df2['Costs_inv'] * tau[0]
         df2['ANN_factor'] = tau[0]
-        df2['Costs_grid_connection'] = get_variable_in_pandas(df, 'Costs_grid_connection_House').groupby(level=1).sum()  # yearly cost for grid connection
 
-        df3 = get_variable_in_pandas(df, 'Costs_House_rep')
+        df2['Costs_grid_connection'] = get_parameter_in_pandas(ampl, 'Costs_grid_connection_House', multi_index=True).groupby(level=1).sum()  # yearly cost for grid connection
+
+        df3 = get_parameter_in_pandas(ampl, 'Costs_House_rep')
         df3['Costs_House_rep'] = df3['Costs_House_rep'] * tau[0]
         df3 = df3.rename(columns={'Costs_House_rep': 'Costs_rep'})
 
-        df4 = get_variable_in_pandas(df, 'Costs_House_cft')
+        df4 = get_parameter_in_pandas(ampl, 'Costs_House_cft')
         df4 = df4.rename(columns={'Costs_House_cft': 'Costs_ft'})
 
-        df5 = get_variable_in_pandas(df, 'GWP_house_op')
+        df5 = get_parameter_in_pandas(ampl, 'GWP_house_op')
         df5 = df5.rename(columns={'GWP_house_op': 'GWP_op'})
 
-        df6 = get_variable_in_pandas(df, 'GWP_house_constr')
+        df6 = get_parameter_in_pandas(ampl, 'GWP_house_constr')
         df6 = df6.rename(columns={'GWP_house_constr': 'GWP_constr'})
 
-        df71 = get_parameter_in_pandas(ampl, 'EMOO_CAPEX', multi_index=False)
-        df72 = get_parameter_in_pandas(ampl, 'EMOO_OPEX', multi_index=False)
-        df73 = get_parameter_in_pandas(ampl, 'EMOO_TOTEX', multi_index=False)
-        df75 = get_parameter_in_pandas(ampl, 'EMOO_GWP', multi_index=False)
-        df76 = get_parameter_in_pandas(ampl, 'EMOO_grid', multi_index=False)
+        df71 = get_parameter_in_pandas(ampl, 'EMOO_CAPEX')
+        df72 = get_parameter_in_pandas(ampl, 'EMOO_OPEX')
+        df73 = get_parameter_in_pandas(ampl, 'EMOO_TOTEX')
+        df75 = get_parameter_in_pandas(ampl, 'EMOO_GWP')
+        df76 = get_parameter_in_pandas(ampl, 'EMOO_grid')
 
-        df_N1 = get_variable_in_pandas(df, 'Costs_op')  # without the comfort penalty costs
-        df_N2 = get_variable_in_pandas(df, 'Costs_inv')
[…]
