### 2024-04-04 45f9661
store infrastructure object from SPs
