### 2024-04-04 90f6dde
initialize infrastructure_SP and add condition reading dual values EMOO constraints
