+++ reho/model/district_decomposition.py
@@ -61,7 +61,10 @@ class district_decomposition:
             self.qbuildings_data = qbuildings_data
         self.buildings_data = qbuildings_data['buildings_data']
         self.ERA = sum([self.buildings_data[house]['ERA'] for house in self.buildings_data.keys()])
+
         self.infrastructure = infrastructure.infrastructure(qbuildings_data, units, grids)
+        self.infrastructure_SP = dict()
+        self.build_infrastructure_SP()
 
         self.csv_data = dict()
         self.csv_data["irradiation"] = pd.read_csv(path_to_irradiation, index_col=[0])
@@ -132,8 +135,6 @@ class district_decomposition:
 
         self.results_SP = dict()
         self.results_MP = dict()
-
-        self.infrastructure_SP = dict()
 
         self.solver_attributes_SP = pd.DataFrame()
         self.solver_attributes_MP = pd.DataFrame()
@@ -231,17 +232,16 @@ class district_decomposition:
                 results = {h: self.pool.apply_async(self.SP_initiation_execution, args=(scenario, Scn_ID, Pareto_ID, h, epsilon_init, beta)) for h in self.infrastructure.houses}
 
                 # sometimes, python goes to fast and extract the results before calculating them. This step makes python wait finishing the calculations
-                while len(results[list(self.buildings_data.keys())[-1]].get()) != 3:
+                while len(results[list(self.buildings_data.keys())[-1]].get()) != 2:
                     time.sleep(1)
 
                 # the memory to write and share results is not parallel -> results have to be stored outside calculation
                 for h in self.infrastructure.houses:
-                    (df_Results, attr, infrastructure_SP) = results[h].get()
+                    (df_Results, attr) = results[h].get()
                     self.add_df_Results_SP(Scn_ID, Pareto_ID, self.iter, h, df_Results, attr)
-                    self.infrastructure_SP[h] = infrastructure_SP
             else:
                 for id, h in enumerate(self.infrastructure.houses):
-                    df_Results, attr, infrastructure_SP = self.SP_initiation_execution(scenario, Scn_ID=Scn_ID, Pareto_ID=Pareto_ID, h=h, epsilon_init=epsilon_init, beta=beta)
+                    df_Results, attr = self.SP_initiation_execution(scenario, Scn_ID=Scn_ID, Pareto_ID=Pareto_ID, h=h, epsilon_init=epsilon_init, beta=beta)
                     self.add_df_Results_SP(Scn_ID, Pareto_ID, self.iter, h, df_Results, attr)
 
             self.feasible_solutions += 1  # after each 'round' of SP execution the number of feasible solutions increase
@@ -271,7 +271,7 @@ class district_decomposition:
             print('INITIATE HOUSE: ' + h)
 
         # find district structure and parameter for one single building
-        buildings_data_SP, parameters_SP, infrastructure_SP = self.__split_parameter_sets_per_building(h)
+        buildings_data_SP, parameters_SP = self.__split_parameter_sets_per_building(h)
 
         # epsilon constraints on districts may lead to infeasibilities on building level -> apply them in MP only
         if epsilon_init is not None and self.method['building-scale']:
@@ -289,9 +289,9 @@ class district_decomposition:
             parameters_SP['beta_duals'] = beta_list
 
         if self.method['use_facades'] or self.method['use_pv_orientation']:
-            REHO = compact_optimization(infrastructure_SP, buildings_data_SP, parameters_SP, self.set_indexed, self.cluster, scenario, self.method, self.solver, self.qbuildings_data, csv_data=self.csv_data)
[…]
