+++ scripts/examples/3h_Electricity_prices.py
@@ -21,7 +21,7 @@ if __name__ == '__main__':
 
     # Initialize available units and grids
     # you can use prices coming from public databases
-    prices = electricity_prices.get_prices_from_elcom(canton=reader, category='H4')
+    prices = electricity_prices.get_prices_from_elcom_by_city(city=reader, category='H4')
     grids = infrastructure.initialize_grids({'Electricity': {"Cost_supply_cst": prices['finalcosts'][0], "Cost_demand_cst": 0.16},
                                              'NaturalGas': {"Cost_supply_cst": 0.15},
                                              })
