+++ reho/model/preprocessing/sia_parser.py
@@ -6,7 +6,7 @@ __doc__ = """
 *Collects data from "SIA" Swiss Norms, which are used to distinguish between eight different building types in their usage and behavior.*
 """
 
-def read_sia2024_rooms_sia380_1(digit, sia_file):
+def read_sia2024_rooms_sia380_1(digit, df_sia):
 
     dict_affiliation2digit = {'collective housing': 'I',
                               'individual housing': 'II',
@@ -22,9 +22,9 @@ def read_sia2024_rooms_sia380_1(digit, sia_file):
                               'covered swimming-pool': 'XII',
                               'other': 'XIII'}
 
-    sia_file = sia_file.rename(columns=dict_affiliation2digit)
+    df_sia = df_sia.rename(columns=dict_affiliation2digit)
 
-    return sia_file[digit]
+    return df_sia[digit]
 
 
 def read_sia_2024_profiles(status, df):
