+++ reho/paths.py
@@ -42,6 +42,8 @@ path_to_qbuildings = os.path.join(path_to_data, 'QBuildings')
 
 # SIA
 path_to_sia = os.path.join(path_to_data, 'SIA')
+path_to_sia_equivalence = os.path.join(path_to_sia, 'sia2024_rooms_sia380_1.csv')
+path_to_sia_norms = os.path.join(path_to_sia, 'sia2024_data.xlsx')
 
 # skydome
 path_to_skydome = os.path.join(path_to_data, 'skydome')
@@ -55,9 +57,7 @@ path_to_weather = os.path.join(path_to_data, 'weather')
 # scripts specific paths
 path_to_clustering = os.path.join(os.getcwd(), 'data', 'clustering')
 path_to_configurations = os.path.join(os.getcwd(), 'configurations')
-#path_to_timestamp = os.path.join(path_to_clustering, 'timestamp_'+ File_ID +'.dat')
-path_sia = os.path.join(path_to_sia, 'sia2024_rooms_sia380_1.csv')
-path_norms = os.path.join(path_to_sia, 'sia2024_data.xlsx')
+
 
 
 def path_handler(path_given):
