+++ documentation/conf.py
@@ -42,8 +42,10 @@ bibtex_reference_sorting = None
 # -- Options for HTML output -------------------------------------------------
 
 html_theme = 'pydata_sphinx_theme'
+
 html_sidebars = {
-  "**": []
+  "_autosummary": ["sidebar-nav-bs"],
+  "sections/*": []
 }
 
 html_theme_options = {

[…]
