### 2024-04-09 20b1ebc
debug examples 0, 5a and 3i
