+++ reho/model/postprocessing/write_results.py
@@ -151,30 +151,51 @@ def get_df_Results_from_SP(ampl, scenario, method, buildings_data, filter=True):
 
         return df_Unit, df_Unit_t
 
-    def set_df_grid(ampl):
+    def set_df_grid(ampl, method):
         # Grid_t
         df1 = get_parameter_in_pandas(ampl, 'Grid_demand', multi_index=True)
-        hub = df1.index.levels[1].values[0]
-        df1 = df1.droplevel(1)
-        df2 = get_parameter_in_pandas(ampl, 'Grid_supply', multi_index=True).droplevel(1)
+        df2 = get_parameter_in_pandas(ampl, 'Grid_supply', multi_index=True)
 
-        df_cs = get_parameter_in_pandas(ampl, 'Cost_supply', multi_index=True).droplevel(0)
-        df_cd = get_parameter_in_pandas(ampl, 'Cost_demand', multi_index=True).droplevel(0)
+        df_cs = get_parameter_in_pandas(ampl, 'Cost_supply', multi_index=True)
+        df_cs = df_cs.reorder_levels((1, 0, 2, 3))
+        df_cd = get_parameter_in_pandas(ampl, 'Cost_demand', multi_index=True)
+        df_cd = df_cd.reorder_levels((1, 0, 2, 3))
 
-        df_electricity = get_parameter_in_pandas(ampl, 'Domestic_electricity', multi_index=True).droplevel(0)
+        df_electricity = get_parameter_in_pandas(ampl, 'Domestic_electricity', multi_index=True)
         df_electricity.columns = ['Uncontrollable_load']
         df_electricity = df_electricity.set_index([pd.Index(["Electricity"]*df_electricity.index.size), df_electricity.index])
 
         df_es = get_parameter_in_pandas(ampl, 'GWP_supply', multi_index=True)
         df_ed = get_parameter_in_pandas(ampl, 'GWP_demand', multi_index=True)
+        df_em = pd.concat([df_es, df_ed], axis=1)
+        df_em_tot = pd.DataFrame()
+        for bui in df1.index.get_level_values(1).unique():
+            df = df_em
+            df["Hub"] = bui
+            df_em_tot = pd.concat([df_em_tot, df])
+        df_em_tot.set_index('Hub', append=True, inplace=True)
+        df_em_tot = df_em_tot.reorder_levels((0, 3, 1, 2))
 
-        df12 = pd.concat([df1, df2, df_cs, df_cd, df_es, df_ed, df_electricity], axis=1)
-        df12["Hub"] = hub
-        df12.set_index('Hub', append=True, inplace=True)
-        df_Grid_t = df12.reorder_levels((0, 3, 1, 2))
-        # combine Network & Grid dataframe
+        df_Grid_t = pd.concat([df1, df2, df_cs, df_cd, df_em_tot, df_electricity], axis=1)
+
+        if not method["district-scale"] or not method["district-scale"]:
+            df3 = get_parameter_in_pandas(ampl, 'Network_demand', multi_index=True)
+            df3.columns = ['Grid_demand']
+            df4 = get_parameter_in_pandas(ampl, 'Network_supply', multi_index=True)
+            df4.columns = ['Grid_supply']
+            df_cs_net = get_parameter_in_pandas(ampl, 'Cost_supply_network', multi_index=True)
+            df_cs_net.columns = ['Cost_supply']
+            df_cd_net = get_parameter_in_pandas(ampl, 'Cost_demand_network', multi_index=True)
+            df_cd_net.columns = ['Cost_demand']
+
+            df_electricity_net = df_electricity.groupby(level=(0, 2, 3)).sum()
[…]
