+++ documentation/sections/5_Getting_started.rst
@@ -406,60 +406,12 @@ Set method options
 You can use different methodology options in REHO, specified in the :code:`method` dictionary.
 The methods available are listed in :ref:`tbl-methods`.
 
-.. table:: List of the available methods in REHO
-    :name: tbl-methods
-
-    +-------------------------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+------------------+
-    | Method name                   | Description                                                                                                                                                                                                                                                                                                                 | Default behavior |
-    +===============================+=============================================================================================================================================================================================================================================================================================================================+==================+
-    |                                                                                                                                                                                                          *Solar methods*                                                                                                                                                       |
-    +-------------------------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+------------------+
-    | use_facades                   | Allows to consider the facades for PV panels installation                                                                                                                                                                                                                                                                   |       False      |
-    +-------------------------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+------------------+
-    | use_pv_orientation            | Considers the orientation for the solar potential estimation, including a shadow model from neighbor buildings                                                                                                                                                                                                              |       False      |
-    +-------------------------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+------------------+
-    |                                                                                                                                                                                                       *Optimization methods*                                                                                                                                                   |
-    +-------------------------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+------------------+
-    | building-scale                | Optimizes by considering than each building is an independent system                                                                                                                                                                                                                                                        |       False      |
-    +-------------------------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+------------------+
-    | district-scale                | Optimizes by allowing exchanges between buildings and the use of district units                                                                                                                                                                                                                                             |       False      |
-    +-------------------------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+------------------+
-    | parallel_computation          | Allows to solve sub-problems in parallel                                                                                                                                                                                                                                                                                    |       True       |
-    +-------------------------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+------------------+
-    | switch_off_second_objective   |                                                                                                                                                                                                                                                                                                                             |       False      |
-    +-------------------------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+------------------+
-    |                                                                                                                                                                                                             *Profiles*                                                                                                                                                         |
-    +-------------------------------+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+------------------+
-    | include_stochasticity         | Includes variability among SIA typical consumption profiles                                                                                                                                                                                                                                                                 |       False      |
[…]
