### 2024-04-09 ded1f8a
check path AMPL
