### 2024-04-09 247b825
dynamic mode_speed and transportunits parametrisation
