### 2024-04-10 21a0dd4
problem on new pathway objective
