### 2024-04-10 17bfde6
add a testing of examples
