### 2024-04-10 035386f
check AMPL path + add test_examples
