### 2024-04-10 fbd5fbd
check AMPL path + add test_examples
