+++ requirements.txt
@@ -1,6 +1,6 @@
 amplpy>=0.12.0,<0.13.0
---index-url https://pypi.ampl.com ampl_module_base
---index-url https://pypi.ampl.com ampl_module_highs
+--extra-index-url https://pypi.ampl.com ampl_module_base
+--extra-index-url https://pypi.ampl.com ampl_module_highs
 pandas>=1.5.3,<2.0.0
 openpyxl>=3.1.2,<4.0.0
 numpy>=1.23.4,<2.0.0
@@ -16,4 +16,5 @@ plotly>=5.10,<6.0.0
 kaleido>=0.2.1,<1.0.0
 python-dotenv>=1.0
 requests>=2.0.0,<3.0.0
-coloredlogs
+coloredlogs>=15.0
+pytest==8.*
