+++ reho/model/preprocessing/skydome.py
@@ -40,14 +40,14 @@ def convert_results_txt_to_csv(load_timesteps):
     print(df)
 
 
-def skydome_to_df(csv_data):
+def skydome_to_df(location_data):
     """ reads two txt files: one containing the area and one the position of the center point of the 145 patches,
      which define the skydome. Calculates basic additional values and returns all data in one single df
 
     :return:   df
     """
-    df_area = csv_data["df_area"]  # area of patches
-    df_cenpts = csv_data["df_cenpts"]  #  # location of centre points
+    df_area = location_data["df_Area"]  # area of patches
+    df_cenpts = location_data["df_Cenpts"]  # location of centre points
 
     df_dome = pd.DataFrame()
     df_dome['Area'] = df_area[0]
@@ -68,17 +68,17 @@ def skydome_to_df(csv_data):
     return df_dome
 
 
-def irradiation_to_df(ampl, df_IRR, df_time):
+def irradiation_to_df(ampl, df_irradiation, df_time):
     """reads Irradiation values of all 145 for the timesteps given in the csv file.
      Converts them to float and returns them as df"""
 
     #change column name from string to int
-    df_IRR.columns = df_IRR.columns.astype(int)
+    df_irradiation.columns = df_irradiation.columns.astype(int)
     #change values from string to float
-    df_IRR = df_IRR.infer_objects()
+    df_irradiation = df_irradiation.infer_objects()
 
     # parse index as datetime
-    df_IRR.index = pd.to_datetime(df_IRR.index)
+    df_irradiation.index = pd.to_datetime(df_irradiation.index)
 
     # get relevant cluster information
     PeriodDuration  = ampl.getParameter('TimeEnd').getValues().toPandas()
@@ -92,7 +92,7 @@ def irradiation_to_df(ampl, df_IRR, df_time):
         end = PeriodDuration.xs(p+1).TimeEnd  # ampl starts at 1
 
         date2 = date1 + timedelta(hours=int(end)-1)
-        df_period = df_IRR.loc[date1:date2]
+        df_period = df_irradiation.loc[date1:date2]
 
         for t in np.arange(1, int(end)+1):  # ampl starts at 1
             list_timesteps.append((p +1, t))  # create ample index
@@ -102,27 +102,25 @@ def irradiation_to_df(ampl, df_IRR, df_time):
     idx = pd.MultiIndex.from_tuples(list_timesteps)
 
     # marry index and data
-    df_IRR = df_p.set_index(idx)
+    df_irradiation = df_p.set_index(idx)
 
-    df = df_IRR.stack()
+    df = df_irradiation.stack()
     df = df.reorder_levels([2, 0, 1])
     df = pd.DataFrame(df)
     df = df.rename(columns = {0 : 'Irr'})
     return df
 
 
-def irradiation_to_df_general(irradiation):
+def irradiation_to_df_general(df_irradiation):
     """reads Irradiation values of all 145 for the timesteps given in the csv file.
      Converts them to float and returns them as df"""
 
-    df_IRR = irradiation #pd.read_csv(irradiation_csv, index_col=[0])
-
     #change column name from string to int
-    df_IRR.columns = df_IRR.columns.astype(int)
+    df_irradiation.columns = df_irradiation.columns.astype(int)
     # change values from string to float
-    df_IRR = df_IRR.infer_objects()
+    df_irradiation = df_irradiation.infer_objects()
 
-    return df_IRR
+    return df_irradiation
 
 
[…]
