+++ reho/model/reho.py
@@ -151,9 +151,11 @@ class reho(district_decomposition):
             ampl, exitcode = self.execute_dantzig_wolfe_decomposition(scenario, Scn_ID, Pareto_ID=1)
         else:
             if self.method['use_facades'] or self.method['use_pv_orientation']:
-                REHO = compact_optimization(self.infrastructure, self.buildings_data, self.parameters, self.set_indexed, self.cluster, scenario, self.method, self.solver, self.qbuildings_data, csv_data=self.csv_data)
+                REHO = compact_optimization(self.infrastructure, self.buildings_data, self.location_data, self.parameters, self.set_indexed, self.cluster,
+                                            scenario, self.method, self.solver, self.qbuildings_data)
             else:
-                REHO = compact_optimization(self.infrastructure, self.buildings_data, self.parameters, self.set_indexed, self.cluster, scenario, self.method, self.solver, csv_data=self.csv_data)
+                REHO = compact_optimization(self.infrastructure, self.buildings_data, self.location_data, self.parameters, self.set_indexed, self.cluster,
+                                            scenario, self.method, self.solver)
             ampl, exitcode = REHO.solve_model()
 
         scenario = {'Objective': objective1}
@@ -182,9 +184,11 @@ class reho(district_decomposition):
             ampl, exitcode = self.execute_dantzig_wolfe_decomposition(scenario, Scn_ID, Pareto_ID=Pareto_ID)
         else:
             if self.method['use_facades'] or self.method['use_pv_orientation']:
-                REHO = compact_optimization(self.infrastructure, self.buildings_data, self.parameters, self.set_indexed, self.cluster, scenario, self.method, self.solver, self.qbuildings_data, csv_data=self.csv_data)
+                REHO = compact_optimization(self.infrastructure, self.buildings_data, self.location_data, self.parameters, self.set_indexed, self.cluster,
+                                            scenario, self.method, self.solver, self.qbuildings_data)
             else:
-                REHO = compact_optimization(self.infrastructure, self.buildings_data, self.parameters, self.set_indexed, self.cluster, scenario, self.method, self.solver, csv_data=self.csv_data)
+                REHO = compact_optimization(self.infrastructure, self.buildings_data, self.location_data, self.parameters, self.set_indexed, self.cluster,
+                                            scenario, self.method, self.solver)
             ampl, exitcode = REHO.solve_model()
 
         scenario = {'Objective': objective2}
@@ -236,11 +240,11 @@ class reho(district_decomposition):
                 ampl, exitcode = self.execute_dantzig_wolfe_decomposition(scenario, Scn_ID, Pareto_ID=nParetoIT, epsilon_init=epsilon_init)
             else:
                 if self.method['use_facades'] or self.method['use_pv_orientation']:
-                    REHO = compact_optimization(self.infrastructure, self.buildings_data, self.parameters, self.set_indexed,
-                                                self.cluster, scenario, self.method, self.solver, self.qbuildings_data, csv_data=self.csv_data)
+                    REHO = compact_optimization(self.infrastructure, self.buildings_data, self.location_data, self.parameters, self.set_indexed,
+                                                self.cluster, scenario, self.method, self.solver, self.qbuildings_data)
                 else:
-                    REHO = compact_optimization(self.infrastructure, self.buildings_data, self.parameters, self.set_indexed,
-                                                self.cluster, scenario, self.method, self.solver, csv_data=self.csv_data)
+                    REHO = compact_optimization(self.infrastructure, self.buildings_data, self.location_data, self.parameters, self.set_indexed,
+                                                self.cluster, scenario, self.method, self.solver)
[…]
