+++ reho/model/preprocessing/sia_parser.py
@@ -6,7 +6,7 @@ __doc__ = """
 *Collects data from "SIA" Swiss Norms, which are used to distinguish between eight different building types in their usage and behavior.*
 """
 
-def read_sia2024_rooms_sia380_1(digit, df_sia):
+def read_sia2024_rooms_sia380_1(digit, df_SIA_380):
 
     dict_affiliation2digit = {'collective housing': 'I',
                               'individual housing': 'II',
@@ -22,9 +22,9 @@ def read_sia2024_rooms_sia380_1(digit, df_sia):
                               'covered swimming-pool': 'XII',
                               'other': 'XIII'}
 
-    df_sia = df_sia.rename(columns=dict_affiliation2digit)
+    df_SIA_380 = df_SIA_380.rename(columns=dict_affiliation2digit)
 
-    return df_sia[digit]
+    return df_SIA_380[digit]
 
 
 def read_sia_2024_profiles(status, df):
