+++ reho/model/district_decomposition.py
@@ -1,6 +1,7 @@
 import reho.model.infrastructure as infrastructure
 from reho.model.compact_optimization import *
 import reho.model.postprocessing.write_results as WR
+from reho.model.preprocessing.location_data import *
 from itertools import groupby
 import warnings
 import time
@@ -52,28 +53,19 @@ class district_decomposition:
         # methods
         self.method = initialize_default_methods(method)
         self.logger = logging.getLogger(__name__)
-        if not method['disable_print']:
+        if method['print_logs']:
             coloredlogs.install(level=logging.INFO, logger=self.logger, isatty=True,
                                 fmt="%(message)s", stream=sys.stdout)
 
-        # District attributes / used also in REHO
+        # infrastructure
         if method['use_facades'] or method['use_pv_orientation']:
             self.qbuildings_data = qbuildings_data
         self.buildings_data = qbuildings_data['buildings_data']
         self.ERA = sum([self.buildings_data[house]['ERA'] for house in self.buildings_data.keys()])
-
+        
         self.infrastructure = infrastructure.infrastructure(qbuildings_data, units, grids)
         self.infrastructure_SP = dict()
         self.build_infrastructure_SP()
-
-        self.csv_data = dict()
-        self.csv_data["irradiation"] = pd.read_csv(path_to_irradiation, index_col=[0])
-        self.csv_data["df_area"] = pd.read_csv(path_to_areas, header=None)
-        self.csv_data["df_cenpts"] = pd.read_csv(path_to_cenpts, header=None)
-        self.csv_data["df_sia"] = pd.read_csv(path_to_sia_equivalence, sep=';', index_col=[0], header=[0])
-        self.csv_data["emissions_matrix"] = pd.read_csv(path_to_emissions_matrix, index_col = [0,1,2])
-        self.csv_data["sia2024_data"] = pd.read_excel(path_to_sia_norms, sheet_name=['profiles', 'calculs', 'data'], engine='openpyxl',
-                           index_col=[0], skiprows=[0, 2, 3, 4], header=[0])
 
         if cluster is None:
             self.cluster = {'Location': 'Geneva', 'Attributes': ['I', 'T', 'W'], 'Periods': 10, 'PeriodDuration': 24}
@@ -81,20 +73,28 @@ class district_decomposition:
             self.cluster = cluster
         self.File_ID = WD.get_cluster_file_ID(self.cluster)
 
-        path_to_timestamp = os.path.join(path_to_clustering, 'timestamp_' + self.File_ID + '.dat')
-        path_to_westfacades_irr = os.path.join(path_to_clustering, 'westfacades_irr_' + self.File_ID + '.txt')
-        self.csv_data["timestamp"] = pd.read_csv(path_to_timestamp, delimiter='\t', parse_dates=[0])
-        self.csv_data["westfacades_irr"] = pd.read_csv(path_to_westfacades_irr, header=None)[0].values
+        # load SIA norms
+        sia_data = dict()
+        sia_data["df_SIA_380"] = pd.read_csv(path_to_sia_equivalence, sep=';', index_col=[0], header=[0])
+        sia_data["df_SIA_2024"] = pd.read_excel(path_to_sia_norms, sheet_name=['profiles', 'calculs', 'data'], engine='openpyxl', index_col=[0],
+                                                     skiprows=[0, 2, 3, 4], header=[0])
+
+        # retrieve location data
+        self.location_data = return_location_data(cluster)
 
         if parameters is None:
             self.parameters = {}
         else:
             self.parameters = parameters
 
-        # Heat gains from electricity and people, domestic hot water demand, domestic electricity demand
+        # build end use demands profile
[…]
