+++ reho/model/preprocessing/buildings_profiles.py
@@ -1,43 +1,12 @@
-import numpy as np
-import pandas as pd
 from datetime import timedelta
-from reho.model.preprocessing.weather import get_cluster_file_ID
 from reho.model.preprocessing.sia_parser import *
 from reho.model.preprocessing.QBuildings import *
 
 __doc__ = """
-*Generates the buildings profiles : 1) Domestic hot water (DHW), 2) Domestic electricity, 3) Heat gains, 4) Solar gains.*
+Generates the buildings profiles : 1) Domestic hot water (DHW), 2) Domestic electricity, 3) Heat gains, 4) Solar gains.*
 """
 
-def annual_to_typical(cluster, annual_file, timestamp_data, typical_file=None):
-    """
-    From an annual profile, with 8760 values, extracts the values corresponding to the typical days.
-    """
-
-    # Get which days are the typical ones
-    typical_days = pd.Series([i.strftime("%m/%d/%Y") for i in timestamp_data.Date]).values
-    df_annual = file_reader(annual_file)
-    t1 = pd.to_datetime('1/1/2005', dayfirst=True, infer_datetime_format=True)
-
-    # hour 1 is between 0:00 - 1:00 and is indexed with starting hour so 0:00
-    for h in df_annual.index.values:
-        df_annual.loc[h, 'h'] = t1 + timedelta(hours=(int(h) - 1))
-
-    df_annual = df_annual.set_index('h')
-    df_typical = pd.DataFrame()
-    for i, td in enumerate(typical_days):
-        df_typical = pd.concat([df_typical, df_annual.loc[td]], sort=True)
-    periods = list(range(1, len(typical_days) + 1))
-    hours = list(range(1, cluster['PeriodDuration'] + 1))
-    df_typical = df_typical.reset_index(drop=True)
-    df_typical = df_typical.set_index(pd.MultiIndex.from_product([periods, hours], names=['Period', 'Hour']))
-    if typical_file:
-        df_typical.to_csv(typical_file)
-
-    return df_typical
-
-
-def profile_reference_temperature(parameters_to_ampl, cluster):
+def reference_temperature_profile(parameters_to_ampl, cluster):
     # TODO: time dependent indoor temperature f.e. lower at night
 
     total_timesteps = cluster['Periods'] * cluster['PeriodDuration'] + 2
@@ -49,8 +18,56 @@ def profile_reference_temperature(parameters_to_ampl, cluster):
 
     return np_temperature
 
+def solar_gains_profile(buildings_data, sia_data, location_data):
+    """
+    Computes the solar heat gains from the irradiance.
 
-def build_eud_profiles(buildings_data, cluster, df_sia, df_SIA, df_timestamp,
+    It uses a typical irradiation file and uses `calc_orientation_profiles` to obtain the irradiation on the west
+    facades. Additionally, the solar gains depends on the facades and on a window fraction (obtained from SIA 2024).
+
+    Parameters
+    ----------
+    ampl : AMPL
+        The AMPL object created in the reho.model.reho.reho class.
+    buildings_data : dict
+        Dictionary of buildings data from QBuildingsReader class.
+    File_ID : str
+        File ID of the clustering results, used to know the periods and period duration.
+
+    Returns
+    -------
+    A Numpy array of shape (242,) with the solar gains for each timesteps.
+    """
+    
+    irr_west = location_data["df_Westfacades_irr"]
+
+    g = np.repeat(0.5, len(irr_west))  # g-value SIA 2024
+    g[irr_west > 0.2] = 0.1  # assumption that if irradiation exceeds 200 W/m2, we use sunblinds
+
+    np_gains = np.array([])
+    for b in buildings_data:
[…]
