### 2024-04-11 8dd3214
debug issues with EMOO SPs
