+++ reho/plotting/plotting.py
@@ -1812,58 +1812,110 @@ def plot_pathway(results, scn_id=None, EMOO_list=[], y_span=[], label='EN_short'
         raise SystemExit('Please enter the list of years used in the pathway optimization')
 
     # Plot
-    fig,ax=plt.subplots(figsize=(12,7))
-    ax.set_xlabel('Year',fontsize=20)
-    ax.set_xlim((min(y_span), max(y_span)))
-    ax.set_title("Investment Pathway", fontsize=35)
-    ax.tick_params(axis='x', labelsize=15)
+    #fig, (ax1, ax2) = plt.subplots(1, 2, sharex=False, figsize=(20,7))
+    fig, ax1=plt.subplots()
+
+    # ax1.set_xlabel('Year',fontsize=20)
+    # ax1.set_xlim((min(y_span), max(y_span)))
+    ax1.set_title("Investment Pathway", fontsize=35)
+    ax1.tick_params(axis='x', labelsize=15)
 
     obj=[]
+    obj2=[]
+    add_line=False
     if objective=='CAPEX':
         obj_label = 'Cumulative capital investment [CHF/m2/y]'
         for i in range(len(EMOO_list)):
             obj=obj+[results[scn_id][i]['df_KPIs']['capex_m2']['Network']]
-    lns1 = ax.plot(y_span, np.cumsum(obj), label=objective, color='C0', linewidth=2)
+    elif objective=='TOTEX':
+        obj_label = 'Investment [CHF/m2/y]'
+        for i in range(len(EMOO_list)):
+            obj = obj + [results[scn_id][i]['df_KPIs']['capex_m2']['Network']]
+            obj2= obj2 + [results[scn_id][i]['df_KPIs']['opex_m2']['Network']]
+        objective='Cumulative CAPEX'
+        add_line=True
+        lns4 = ax1.plot(y_span, obj2, label='OPEX', color='C0', linewidth=2, linestyle='--')
+
+    lns1 = ax1.plot(y_span, np.cumsum(obj), label=objective, color='C0', linewidth=2)
+
+
     # Left y-axis params
-    ax.set_ylabel(obj_label, color='C0', fontsize=20)
-    ax.tick_params(axis='y', labelcolor='C0', labelsize=15)
-    ax.set_ylim(bottom=0)
+    ax1.set_ylabel(obj_label, color='C0', fontsize=20)
+    ax1.tick_params(axis='y', labelcolor='C0', labelsize=15)
+    ax1.set_ylim(bottom=0)
 
     # Creating a second y-axis
-    ax2 = ax.twinx()
+    ax11 = ax1.twinx()
     constr=[]
     if constraint=='GWP':
         constr_label='GWP [kgCO2-eq/m2]'
         for i in range(len(EMOO_list)):
             constr=constr+[results[scn_id][i]['df_KPIs']['gwp_tot_m2']['Network']]
+
+    elif constraint == 'elec_export':
+        constr_label='Electricity Exportation [MWh/m2/y]'
+        for i in range(len(EMOO_list)):
+            constr=constr+[(results[scn_id][i]['df_Annuals'].loc['Electricity'].loc['Network']['Demand_MWh']-results[scn_id][i]['df_Annuals'].loc['Electricity'].loc['Network']['Supply_MWh'])/results[scn_id][i]['df_Buildings']['ERA'].sum()]
+
     # Plotting the second set of data on the second y-axis
-    lns2 = ax2.plot(y_span, constr, label=constraint, color='C1', linewidth=2)
-    lns3 = ax2.plot(y_span, EMOO_list, label='EMOO', color='C1', linestyle='--', linewidth=2)
[…]
