+++ reho/model/preprocessing/local_data.py
@@ -1,0 +1,45 @@
+from reho.model.preprocessing.QBuildings import *
+from reho.model.preprocessing.weather import get_cluster_file_ID
+
+
+def return_local_data(cluster):
+    """
+    Reads .csv, .txt and .xlsx files to retrieve the local data (carbon tag, solar irradiation) corresponding to the buildings' area (cluster).
+
+    Parameters
+    ----------
+    cluster : dict
+        Define location district, number of periods, and number of timesteps.
+
+    Returns
+    -------
+    local_data : dict
+        A dictionary containing the following pandas dataframe: df_Emissions, df_Irradiation, df_Area, df_Cenpts, df_Timestamp, df_Westfacades_irr.
+    """
+
+    local_data = dict()
+
+    # Carbon emissions
+    local_data["df_Emissions"] = file_reader(path_to_emissions, index_col=[0, 1, 2])
+
+    # Solar
+    # TODO: These files are specific to Rolle.
+    local_data["df_Irradiation"] = pd.read_csv(path_to_irradiation, index_col=[0])
+    local_data["df_Area"] = pd.read_csv(path_to_areas, header=None)
+    local_data["df_Cenpts"] = pd.read_csv(path_to_cenpts, header=None)
+
+    File_ID = get_cluster_file_ID(cluster)
+    path_to_timestamp = os.path.join(path_to_clustering, 'timestamp_' + File_ID + '.dat')
+    local_data["df_Timestamp"] = pd.read_csv(path_to_timestamp, delimiter='\t', parse_dates=[0])
+
+    path_to_westfacades_irr = os.path.join(path_to_clustering, 'westfacades_irr_' + File_ID + '.txt')
+    # check if irradiation already exists:
+    if not os.path.exists(path_to_westfacades_irr):
+        local_data["df_Timestamp"].Date = pd.to_datetime(local_data["df_Timestamp"]['Date'], format="%m/%d/%Y/%H")
+        frequency_dict = pd.Series(local_data["df_Timestamp"].Frequency.values, index=local_data["df_Timestamp"].Date).to_dict()
+        frequency_dict['PeriodDuration'] = {p + 1: cluster['PeriodDuration'] for p in range(cluster['Periods'])}
+        df_annual, irr_west = SKD.calc_orientation_profiles(270, 90, 0, local_data, frequency_dict)
+        np.savetxt(path_to_westfacades_irr, irr_west)
+    local_data["df_Westfacades_irr"] = pd.read_csv(path_to_westfacades_irr, header=None)[0].values
+
+    return local_data
