### 2024-04-16 eee4fb4
documentation update (automodules with :members:)
