+++ reho/model/postprocessing/__init__.py
@@ -1,3 +1,3 @@
 __doc__ = """
-*Directory where the output of the optimization from the AMPL model is extracted and processed to give a REHO results dictionary.*
+Directory where the output of the optimization from the AMPL model is extracted and processed to give a REHO results dictionary.
 """
