+++ reho/plotting/plotting.py
@@ -14,7 +14,7 @@ from reho.model.compact_optimization import *
 
 
 __doc__ = """
-*Contains ready-to-use representations for results generated by REHO.*
+Contains ready-to-use representations for results generated by REHO.
 """
 # ---------------------------------------------------------------------------------------------
 # Definition of colors and labels
