+++ reho/model/preprocessing/__init__.py
@@ -1,3 +1,3 @@
 __doc__ = """
-*Directory where the data are prepared as input for the AMPL optimization.*
+Directory where the data are prepared as input for the AMPL optimization.
 """
