+++ reho/model/preprocessing/sia_parser.py
@@ -1,9 +1,8 @@
 import pandas as pd
-from reho.paths import *
 import numpy as np
 
 __doc__ = """
-*Collects data from "SIA" Swiss Norms, which are used to distinguish between eight different building types in their usage and behavior.*
+Collects data from the `SIA Swiss norms <https://www.sia.ch/fr/services/sia-norm/>`_ , which are used to distinguish between eight different building types in their usage and behavior.
 """
 
 def read_sia2024_rooms_sia380_1(digit, df_SIA_380):
