+++ reho/model/reho.py
@@ -151,10 +151,10 @@ class reho(district_decomposition):
             ampl, exitcode = self.execute_dantzig_wolfe_decomposition(scenario, Scn_ID, Pareto_ID=1)
         else:
             if self.method['use_facades'] or self.method['use_pv_orientation']:
-                REHO = compact_optimization(self.infrastructure, self.buildings_data, self.location_data, self.parameters, self.set_indexed, self.cluster,
+                REHO = compact_optimization(self.infrastructure, self.buildings_data, self.local_data, self.parameters, self.set_indexed, self.cluster,
                                             scenario, self.method, self.solver, self.qbuildings_data)
             else:
-                REHO = compact_optimization(self.infrastructure, self.buildings_data, self.location_data, self.parameters, self.set_indexed, self.cluster,
+                REHO = compact_optimization(self.infrastructure, self.buildings_data, self.local_data, self.parameters, self.set_indexed, self.cluster,
                                             scenario, self.method, self.solver)
             ampl, exitcode = REHO.solve_model()
 
@@ -184,10 +184,10 @@ class reho(district_decomposition):
             ampl, exitcode = self.execute_dantzig_wolfe_decomposition(scenario, Scn_ID, Pareto_ID=Pareto_ID)
         else:
             if self.method['use_facades'] or self.method['use_pv_orientation']:
-                REHO = compact_optimization(self.infrastructure, self.buildings_data, self.location_data, self.parameters, self.set_indexed, self.cluster,
+                REHO = compact_optimization(self.infrastructure, self.buildings_data, self.local_data, self.parameters, self.set_indexed, self.cluster,
                                             scenario, self.method, self.solver, self.qbuildings_data)
             else:
-                REHO = compact_optimization(self.infrastructure, self.buildings_data, self.location_data, self.parameters, self.set_indexed, self.cluster,
+                REHO = compact_optimization(self.infrastructure, self.buildings_data, self.local_data, self.parameters, self.set_indexed, self.cluster,
                                             scenario, self.method, self.solver)
             ampl, exitcode = REHO.solve_model()
 
@@ -240,10 +240,10 @@ class reho(district_decomposition):
                 ampl, exitcode = self.execute_dantzig_wolfe_decomposition(scenario, Scn_ID, Pareto_ID=nParetoIT, epsilon_init=epsilon_init)
             else:
                 if self.method['use_facades'] or self.method['use_pv_orientation']:
-                    REHO = compact_optimization(self.infrastructure, self.buildings_data, self.location_data, self.parameters, self.set_indexed,
+                    REHO = compact_optimization(self.infrastructure, self.buildings_data, self.local_data, self.parameters, self.set_indexed,
                                                 self.cluster, scenario, self.method, self.solver, self.qbuildings_data)
                 else:
-                    REHO = compact_optimization(self.infrastructure, self.buildings_data, self.location_data, self.parameters, self.set_indexed,
+                    REHO = compact_optimization(self.infrastructure, self.buildings_data, self.local_data, self.parameters, self.set_indexed,
                                                 self.cluster, scenario, self.method, self.solver)
                 ampl, exitcode = REHO.solve_model()
 
@@ -283,10 +283,10 @@ class reho(district_decomposition):
                     ampl, exitcode = self.execute_dantzig_wolfe_decomposition(scenario, Scn_ID, Pareto_ID=nParetoIT, epsilon_init=epsilon_init)
                 else:
                     if self.method['use_facades'] or self.method['use_pv_orientation']:
-                        REHO = compact_optimization(self.infrastructure, self.buildings_data, self.location_data, self.parameters,
+                        REHO = compact_optimization(self.infrastructure, self.buildings_data, self.local_data, self.parameters,
[…]
