+++ documentation/sections/4_Package_structure.rst
@@ -39,9 +39,9 @@ REHO exploits the benefits of two programming languages:
 
 Directory for data-related files.
 
-- **electricity/**
+- **elcom/**
 - **emissions/**
-- **parameters/**
+- **infrastructure/**
 - **QBuildings/**
 - **SIA/**
 - **skydome/**
@@ -74,18 +74,25 @@ Core of the optimization model (model objectives, constraints, modelling equatio
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 .. automodule:: reho.model.postprocessing.building_scale_network_builder
+    :members:
 
 `KPIs.py`
 ~~~~~~~~~~~~~~~~~~~~~~~
 
 .. automodule:: reho.model.postprocessing.KPIs
+    :members:
+
+`sensitivity_analysis.py`
+~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+.. automodule:: reho.model.postprocessing.sensitivity_analysis
+    :members:
 
 `write_results.py`
 ~~~~~~~~~~~~~~~~~~~~~~~
 
 .. automodule:: reho.model.postprocessing.write_results
-
-
+    :members:
 
 **preprocessing/**
 ------------------------
@@ -96,22 +103,18 @@ Core of the optimization model (model objectives, constraints, modelling equatio
 ~~~~~~~~~~~~~~~~~~~~~~~
 
 .. automodule:: reho.model.preprocessing.buildings_profiles
-    :members: eud_profiles, solar_gains_profile
+    :members:
 
 
 `clustering.py`
-~~~~~~~~~~~~~~~~~~~~~~~
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-.. autoclass:: reho.model.preprocessing.clustering
-
-.. note::
-    The meteo file provided are only the ones from the 6 meteo archetypes. Should we link every other location to these
-    typical meteos if no meteo file is provided? This could be easily done as the transformer in QBuildings have
-    the meteo_cluster.
+.. automodule:: reho.model.preprocessing.clustering
+    :members:
 
 
 `electricity_prices.py`
-~~~~~~~~~~~~~~~~~~~~~~~
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 .. automodule:: reho.model.preprocessing.electricity_prices
     :members: get_prices_from_elcom_by_canton, get_prices_from_elcom_by_city, get_injection_prices, get_electricity_prices
@@ -121,37 +124,36 @@ Core of the optimization model (model objectives, constraints, modelling equatio
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 .. automodule:: reho.model.preprocessing.emissions_parser
-
+    :members:
 
 `EV_profile_generator.py`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 .. automodule:: reho.model.preprocessing.EV_profile_generator
-    :members: generate_EV_plugged_out_profiles_district
+    :members:
 
-`location_data.py`
+`local_data.py`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-.. automodule:: reho.model.preprocessing.location_data
+.. automodule:: reho.model.preprocessing.local_data
+    :members:
 
 `QBuildings.py`
 ~~~~~~~~~~~~~~~~~~~~~~~
 
[…]

[…]
