+++ requirements.txt
@@ -16,7 +16,6 @@ plotly>=5.10,<6.0.0
 kaleido>=0.2.1,<1.0.0
 python-dotenv>=1.0
 requests>=2.0.0,<3.0.0
-coloredlogs>=15.0
-pytest==8.*
-SALib
-qmcpy
+coloredlogs>=>=15.0.1
+SALib>=1.4.8
+qmcpy>=1.4.5
