+++ reho/plotting/plotting.py
@@ -51,16 +51,16 @@ def handle_zero_rows(df):
     return df.loc[~is_zero_row]
 
 
-def prepare_dfs(df_eco, indexed_on='Scn_ID', neg=False, premium_version=False,
+def prepare_dfs(df_Economics, indexed_on='Scn_ID', neg=False, premium_version=None,
                 additional_costs={}, scaling_factor=1):
     """
     This function prepares the dataframes that will be needed for the plot_performance and plot_actors
     """
-    df_eco = df_eco.xs('Network', level='Hub', axis=0)
-    df_eco = df_eco.groupby(level=indexed_on, sort=False).sum() * scaling_factor
-    indexes = df_eco.index.tolist()
+    df_Economics = df_Economics.xs('Network', level='Hub', axis=0)
+    df_Economics = df_Economics.groupby(level=indexed_on, sort=False).sum() * scaling_factor
+    indexes = df_Economics.index.tolist()
 
-    data_capex = df_eco.xs('investment', level='Category', axis=1).transpose()
+    data_capex = df_Economics.xs('investment', level='Category', axis=1).transpose()
     data_capex.index.names = ['Unit']
 
     if 'isolation' in additional_costs:
@@ -68,7 +68,7 @@ def prepare_dfs(df_eco, indexed_on='Scn_ID', neg=False, premium_version=False,
 
     data_capex = data_capex.reset_index().merge(layout, left_on="Unit", right_on='Name').set_index("Unit").fillna(0)
 
-    data_opex = df_eco.xs('operation', level='Category', axis=1).transpose()
+    data_opex = df_Economics.xs('operation', level='Category', axis=1).transpose()
     indices = data_opex.index.get_level_values(0)
     new_indices = []
     [new_indices.append(tuple(idx.split("_", 1))) for idx in indices]
@@ -79,8 +79,8 @@ def prepare_dfs(df_eco, indexed_on='Scn_ID', neg=False, premium_version=False,
             new_indices[i] = ('revenues', 'Electrical_grid_feed_in')
     data_opex.index = pd.MultiIndex.from_tuples(new_indices, names=['type', 'Layer'])
 
-    if premium_version:
-        data_opex.loc[('avoided', 'solar_premium'), :] = data_opex.loc[('avoided', 'PV_SC')] * (0.279 - 0.1645) / 0.279
+    if premium_version is not None:
+        data_opex.loc[('avoided', 'solar_premium'), :] = data_opex.loc[('avoided', 'PV_SC')] * (premium_version[0] - premium_version[1]) / premium_version[0]
         data_opex.loc[('revenues', 'solar_value'), :] = data_opex.loc[('revenues', 'Electrical_grid_feed_in')] + \
                                                         data_opex.loc[('avoided', 'PV_SC')] - data_opex.loc[
                                                             ('avoided', 'solar_premium')]
@@ -93,7 +93,8 @@ def prepare_dfs(df_eco, indexed_on='Scn_ID', neg=False, premium_version=False,
         data_opex.loc[('costs', 'Gasoline'), :] = additional_costs['mobility']
 
     if 'ict' in additional_costs:
-        data_opex.loc[('costs', 'ict'), :] = additional_costs['ict']
+        data_opex.loc[('costs', 'Data'), :] = additional_costs['ict']
+        data_opex.loc[('revenues', 'Data'), :] = 0
 
     if neg:
         indices = data_opex.index.get_level_values(0)
@@ -171,7 +172,7 @@ def plot_performance(results, plot='costs', indexed_on='Scn_ID', label='EN_long'
[…]
