### 2024-04-20 130d5cc
working log : charging station, outside charging and load
