+++ scripts/solene/0_Suurstoffidata.py
@@ -7,3 +7,4 @@ from reho.model.reho import *
 if __name__ == '__main__':
     filepath = "data/mobility/Daten_März_2023.sav"
     df, meta = pyreadstat.read_sav(filepath)
+    a = 1
