### 2024-04-22 27a9fbf
bug : EV_E_charged_outside extracting
