+++ reho/model/postprocessing/write_results.py
@@ -515,14 +515,17 @@ def get_df_Results_from_MP(ampl, binary=False, method=None, district=None, read_
             df6 = get_parameter_in_pandas(ampl, 'EV_V2V', multi_index=True)
             df6 = pd.concat([df6], keys=['Electricity'], names=['Layer'])
             df7 = get_parameter_in_pandas(ampl, 'EV_E_charged_outside', multi_index=True)
+            # df7 = df7.unstack(level = 0 ).droplevel(0,axis = 1)
+            # df7 = pd.concat([df7], keys=['Electricity'], names=['Layer'])
             df_Unit_t = pd.concat([df_Unit_t, df4, df5, df6,df7], axis=1)
     df_Unit_t.index.names = ['Layer', 'Unit', 'Period', 'Time']
 
-    units_districts = district.UnitsOfDistrict
-    district_l_u = []
-    for l, units in district.UnitsOfLayer.items():
-        [district_l_u.append((l, unit)) for unit in units if unit in units_districts]
-    df_Unit_t = df_Unit_t.reset_index(level=['Period', 'Time']).loc[district_l_u, :]
+    # this bit does not let me export EV_E_charged_outside properly :
+    # units_districts = district.UnitsOfDistrict
+    # district_l_u = []
+    # for l, units in district.UnitsOfLayer.items():
+    #     [district_l_u.append((l, unit)) for unit in units if unit in units_districts]
+    # df_Unit_t = df_Unit_t.reset_index(level=['Period', 'Time']).loc[district_l_u, :]
     df_Results["df_Unit_t"] = df_Unit_t.reset_index().set_index(['Layer', 'Unit', 'Period', 'Time']).sort_index()
 
     # df_lca
