+++ scripts/solene/iterative_scenario.py
@@ -1,0 +1,236 @@
+from reho.model.reho import *
+import datetime
+
+
+def run_district(transformer,price = pd.DataFrame(),externalload = pd.DataFrame()):
+    # Set building parameters
+    reader = QBuildingsReader()
+    reader.establish_connection('Suisse')
+    qbuildings_data = reader.read_db(transformer=transformer, nb_buildings=2)
+
+    # Select weather data
+    cluster = {'Location': 'Geneva', 'Attributes': ['I', 'T', 'W'], 'Periods': 10, 'PeriodDuration': 24}
+
+    # Set scenario
+    scenario = dict()
+    scenario['Objective'] = 'TOTEX'
+    scenario['EMOO'] = {}
+    scenario['specific'] = []
+    scenario['name'] = 'totex'
+    scenario['exclude_units'] = ['NG_Cogeneration']
+    scenario['enforce_units'] = ['EV_district']
+
+    # Initialize available units and grids
+    grids = infrastructure.initialize_grids({'Electricity': {},
+                                             'NaturalGas': {},
+                                             'FossilFuel': {},
+                                             'Mobility': {},
+                                             })
+    units = infrastructure.initialize_units(scenario, grids, district_data=True)
+
+    parameters = {'Population': 9}
+
+    # Set method options
+    method = {'building-scale': True}
+    rehorun = reho(qbuildings_data=qbuildings_data, units=units, grids=grids, cluster=cluster, scenario=scenario,
+                method=method, parameters=parameters, solver="gurobi")
+
+    # Set specific parameters
+    # reho.parameters['TransformerCapacity'] = np.array([1e6, 1e6, 0, 1e6])  # TODO : robustesse of mobility Network
+    rehorun.parameters['Mode_Speed'] = {'Bike2_district': 20}
+    # rehorun.parameters['DailyDist'] = 9
+    if not price.empty :
+        rehorun.parameters['outside_charging_price'] = price
+    if not externalload.empty :
+        rehorun.parameters['charging_externalload'] = externalload
+
+    # Other params
+    # Mobility demand profile can be changed in the csv file data/mobility/dailyprofiles.csv
+    # (imported through the function generate_mobility_profiles)
+
+    # Run optimization
+    rehorun.single_optimization()
+
+    # Save results
+    rehorun.save_results(format=['xlsx', 'pickle'], filename='Mob')
+
+    # Mobility Formatting of results
+    date = datetime.datetime.now().strftime("%d_%H%M")
+    result_file_path = 'results/Mob_totex.xlsx'
+    df_Unit_t = pd.read_excel(result_file_path, sheet_name="df_Unit_t",
+                              index_col=[0, 1, 2, 3])  # refaire sans passer par le xlsx
+    df_Grid_t = pd.read_excel(result_file_path, sheet_name="df_Grid_t",
+                              index_col=[0, 1, 2, 3])
+    # df_Unit_t.index.name = ['Layer','Unit','Period','Time']
+    df_Unit_t = df_Unit_t[df_Unit_t.index.get_level_values("Layer") == "Mobility"]
+    df_Grid_t = df_Grid_t[df_Grid_t.index.get_level_values("Layer") == "Mobility"]
+    df_dd = df_Grid_t[df_Grid_t.index.get_level_values("Hub") == "Network"]
+    # df_dd.index = df_dd.index.droplevel('Hub')
+    df_dd.reset_index("Hub", inplace=True)
+
+    df_mobility = df_Unit_t[['Units_demand', 'Units_supply']].unstack(level='Unit')
+    df_mobility['Domestic_energy'] = df_dd['Domestic_energy']
+    df_mobility.sort_index(inplace=True)
+    df_mobility.to_excel(f"results/3f_mobility{date}.xlsx")
[…]
