### 2024-04-23 abf7ae8
add combined option in plot_performance
