+++ reho/plotting/plotting.py
@@ -12,7 +12,6 @@ from matplotlib.legend_handler import HandlerTuple
 import locale, calendar
 from reho.model.compact_optimization import *
 
-
 __doc__ = """
 Contains ready-to-use representations for results generated by REHO.
 """
@@ -80,7 +79,8 @@ def prepare_dfs(df_Economics, indexed_on='Scn_ID', neg=False, premium_version=No
     data_opex.index = pd.MultiIndex.from_tuples(new_indices, names=['type', 'Layer'])
 
     if premium_version is not None:
-        data_opex.loc[('avoided', 'solar_premium'), :] = data_opex.loc[('avoided', 'PV_SC')] * (premium_version[0] - premium_version[1]) / premium_version[0]
+        data_opex.loc[('avoided', 'solar_premium'), :] = data_opex.loc[('avoided', 'PV_SC')] * (
+                premium_version[0] - premium_version[1]) / premium_version[0]
         data_opex.loc[('revenues', 'solar_value'), :] = data_opex.loc[('revenues', 'Electrical_grid_feed_in')] + \
                                                         data_opex.loc[('avoided', 'PV_SC')] - data_opex.loc[
                                                             ('avoided', 'solar_premium')]
@@ -136,36 +136,47 @@ def remove_building_from_index(df):
 
 
 def plot_performance(results, plot='costs', indexed_on='Scn_ID', label='EN_long', add_annotation=True, per_m2=False,
-                     additional_costs={}, filename=None,
+                     additional_costs={}, scc=0.177, filename=None,
                      export_format='html', scaling_factor=1, return_df=False):
     """
     Plot performance based on REHO results.
 
-    :param results: Dictionary of REHO results.
-    :type results: dict
-    :param plot: Choose among 'costs' and 'gwp'.
-    :type plot: str
-    :param label: Indicates the language to use for the plot. Pick among 'FR_long', 'FR_short', 'EN_long', 'EN_short'.
-    :type label: str
-    :param indexed_on: Whether the results should be grouped on Scn_ID or Pareto_ID.
-    :type indexed_on: str
-    :param add_annotation: Adds the numerical values along the bar plots.
-    :type add_annotation: bool
-    :param per_m2: Set to True to obtain the results divided by the total ERA.
-    :type per_m2: bool
-    :param additional_costs: Dictionary of additional costs to include (choose between 'isolation', 'mobility', and 'ict') and scaling values.
-    :type additional_costs: dict
-    :param filename: Name of the file to be saved.
-    :type filename: str
-    :param export_format: Can be either 'html', 'png', or 'plotly_plot'.
-    :type export_format: str
-    :param scaling_factor: Scaling factor for the plot if a linear assumption is made.
-    :type scaling_factor: int or float
-    :param return_df: A dataframe can be returned for further post-processing or reporting purposes.
-    :type return_df: bool
+    Parameters
+    ----------
+    results: dict
+        Dictionary of REHO results.
+    plot: str
+        Choose among those three possibilities:
 
-    :return: (Optional) A dataframe for further post-processing or reporting purposes.
-    :rtype: pd.DataFrame or None
+        - 'costs' for the economic performance indicators,
+        - 'gwp' for the global warming potential indicators,
+        - 'combined' for a combination of the two indicators, where the emissions are converted into costs using the
+          ``scc`` parameter.
+    label: str
+        Indicates the language to use for the plot. Choose among 'FR_long', 'FR_short', 'EN_long', 'EN_short'.
+    indexed_on: str
[…]
