### 2024-04-23 c19c739
add combined option in plot_expenses (former plot_actors)
