+++ scripts/template/plot.py
@@ -6,6 +6,7 @@ results = pd.read_pickle('results/my_case_study.pickle')
 # Performance plot : Costs and Global Warming Potential
 plotting.plot_performance(results, plot='costs', indexed_on='Scn_ID', filename="figures/performance_costs").show()
 plotting.plot_performance(results, plot='gwp', indexed_on='Scn_ID', filename="figures/performance_gwp").show()
+plotting.plot_performance(results, plot='combined', indexed_on='Scn_ID', filename="figures/performance_combined").show()
 
 # Sankey diagram
 plotting.plot_sankey(results['totex'][0], label='EN_long', color='ColorPastel').show()
