+++ reho/plotting/plotting.py
@@ -51,10 +51,20 @@ def handle_zero_rows(df):
     return df.loc[~is_zero_row]
 
 
+def custom_round(value, decimal):
+    if decimal == 0:
+        rounded_value = int(round(value))
+    elif decimal == 1:
+        rounded_value = round(value, 1)
+    else:
+        raise ValueError("decimal argument must be 0 or 1")
+    return rounded_value
+
+
 def prepare_dfs(df_Economics, indexed_on='Scn_ID', neg=False, premium_version=None,
                 additional_data={}, scaling_factor=1):
     """
-    This function prepares the dataframes that will be needed for the plot_performance and plot_actors
+    This function prepares the dataframes that will be needed for the plot_performance and plot_expenses
     """
     df_Economics = df_Economics.xs('Network', level='Hub', axis=0)
     df_Economics = df_Economics.groupby(level=indexed_on, sort=False).sum() * scaling_factor
@@ -188,6 +198,7 @@ def plot_performance(results, plot='costs', indexed_on='Scn_ID', label='EN_long'
 
     change_data = pd.DataFrame()
     change_data.index = ['x_axis_1', 'x_axis_2', 'y_axis', 'keyword', 'total', 'unites', 'scc_legend']
+    decimal = 0
     lang = re.split('_', label)[0]
 
     if plot == 'costs':
@@ -198,6 +209,7 @@ def plot_performance(results, plot='costs', indexed_on='Scn_ID', label='EN_long'
             df_costs = df_costs / era.sum()
             change_data.loc['y_axis']['FR'] = "Coûts [CHF/m2/an]"
             change_data.loc['y_axis']['EN'] = "Costs [CHF/m2/y]"
+            decimal = 1
         indexes, data_capacities, data_resources = prepare_dfs(df_costs, indexed_on, neg=True,
                                                      additional_data=additional_costs,
                                                      scaling_factor=scaling_factor)
@@ -215,8 +227,9 @@ def plot_performance(results, plot='costs', indexed_on='Scn_ID', label='EN_long'
         df_impact = df_Economics.xs('impact', level='Perf_type')
         if per_m2:
             df_impact = df_impact / era.sum()
-            change_data.loc['y_axis']['FR'] = "GWP [kgCO2/m2/an]"
-            change_data.loc['y_axis']['EN'] = "GWP [kgCO2/m2/y]"
+            change_data.loc['y_axis']['FR'] = "Émissions [kgCO2/m2/an]"
+            change_data.loc['y_axis']['EN'] = "Emissions [kgCO2/m2/y]"
+            decimal = 1
         indexes, data_capacities, data_resources = prepare_dfs(df_impact, indexed_on, neg=True,
                                                     additional_data=additional_gwp,
                                                     scaling_factor=scaling_factor)
@@ -239,7 +252,7 @@ def plot_performance(results, plot='costs', indexed_on='Scn_ID', label='EN_long'
             df_impact = df_impact / era.sum()
             change_data.loc['y_axis']['FR'] = "Coûts [CHF/m2/an]"
             change_data.loc['y_axis']['EN'] = "Costs [CHF/m2/y]"
-
+            decimal = 1
         indexes, data_capacities, data_resources = prepare_dfs(df_costs, indexed_on, neg=True,
                                                                additional_data=additional_costs,
                                                                scaling_factor=scaling_factor)
@@ -255,10 +268,10 @@ def plot_performance(results, plot='costs', indexed_on='Scn_ID', label='EN_long'
 
         showlegend = True
 
-    sum_resources = data_resources[indexes].sum(axis=0).astype(int).reset_index(drop=True)
[…]
