+++ scripts/solene/iterative_scenario.py
@@ -1,86 +1,35 @@
 from reho.model.reho import *
 import datetime
 
+def compute_iterative_parameters(variables,parameters):
+    # for 2 districts:
+    if len(variables.keys()) == 2:
+        for d1,d2 in zip(variables.keys(),reversed(variables.keys())):
+            parameters[d1] = {    "outside_charging_price" : variables[d2]['pi'].rename("outside_charging_price"),
+                            "charging_externalload"  : variables[d2]["externaldemand"].stack().to_frame(name = "charging_externalload")
+                        }
 
-def run_district(transformer,price = pd.DataFrame(),externalload = pd.DataFrame()):
-    # Set building parameters
-    reader = QBuildingsReader()
-    reader.establish_connection('Suisse')
-    qbuildings_data = reader.read_db(transformer=transformer, nb_buildings=2)
+def check_convergence(deltas,df_delta,variables):
+    
+    # Compute Delta
+    df_demand = pd.DataFrame()
+    df_load = pd.DataFrame()
+    for k in variables.keys():
+        df = variables[k]['externaldemand']
+        df.columns.name = "Activity"
+        df_demand[k] = df.stack()
+       
+        df = variables[k]['externalload']
+        df.columns.name = "Activity"
+        df_load[k] = df.stack()
+    
+    df_delta[f"demand{i}"] = df_demand.sum(axis=1)
+    df_delta[f"load{i}"] = df_load.sum(axis=1)
 
-    # Select weather data
-    cluster = {'Location': 'Geneva', 'Attributes': ['I', 'T', 'W'], 'Periods': 10, 'PeriodDuration': 24}
-
-    # Set scenario
-    scenario = dict()
-    scenario['Objective'] = 'TOTEX'
-    scenario['EMOO'] = {}
-    scenario['specific'] = []
-    scenario['name'] = 'totex'
-    scenario['exclude_units'] = ['NG_Cogeneration']
-    scenario['enforce_units'] = ['EV_district']
-
-    # Initialize available units and grids
-    grids = infrastructure.initialize_grids({'Electricity': {},
-                                             'NaturalGas': {},
-                                             'FossilFuel': {},
-                                             'Mobility': {},
-                                             })
-    units = infrastructure.initialize_units(scenario, grids, district_data=True)
-
-    parameters = {'Population': 9}
-
-    # Set method options
-    method = {'building-scale': True}
-    rehorun = reho(qbuildings_data=qbuildings_data, units=units, grids=grids, cluster=cluster, scenario=scenario,
-                method=method, parameters=parameters, solver="gurobi")
-
-    # Set specific parameters
-    # reho.parameters['TransformerCapacity'] = np.array([1e6, 1e6, 0, 1e6])  # TODO : robustesse of mobility Network
-    rehorun.parameters['Mode_Speed'] = {'Bike2_district': 20}
-    # rehorun.parameters['DailyDist'] = 9
-    if not price.empty :
-        rehorun.parameters['outside_charging_price'] = price
-    if not externalload.empty :
-        rehorun.parameters['charging_externalload'] = externalload
-
-    # Other params
-    # Mobility demand profile can be changed in the csv file data/mobility/dailyprofiles.csv
-    # (imported through the function generate_mobility_profiles)
-
-    # Run optimization
-    rehorun.single_optimization()
-
-    # Save results
-    rehorun.save_results(format=['xlsx', 'pickle'], filename='Mob')
-
-    # Mobility Formatting of results
-    date = datetime.datetime.now().strftime("%d_%H%M")
[…]
