+++ reho/model/reho.py
@@ -926,3 +926,60 @@ class reho(district_decomposition):
 
                     writer.close()
                     print('Results are saved in ' + result_file_path)
+
+
+    def save_mobility_results(self,filename = 'mobility_results'):
+        """"
+        Applies some data shuffling on the reho.results and saves the resulting dataframes in an excel file
+        Parameters
+        ----------
+        filename : str, optional
+            Base name of the file to be saved. The extension will be added based on the format.
+            Default is 'mobility_results'.
+        Returns
+        -------
+        None
+        """
+        try:
+            os.makedirs('results')
+        except OSError:
+            if not os.path.isdir('results'):
+                raise
+
+        # reho results
+        results = self.results
+
+        for Scn_ID in list(results.keys()):
+            for Pareto_ID in list(results[Scn_ID].keys()):
+
+                # access results and datawrangling
+                df_Unit_t = results[Scn_ID][Pareto_ID]['df_Unit_t']
+                df_Grid_t = results[Scn_ID][Pareto_ID]['df_Grid_t']
+
+                df_Unit_t_pkm = df_Unit_t[df_Unit_t.index.get_level_values("Layer") == "Mobility"]
+                df_Grid_t_pkm = df_Grid_t[(df_Grid_t.index.get_level_values("Layer") == "Mobility") & (df_Grid_t.index.get_level_values("Hub") == "Network") ]
+                df_Grid_t_pkm.reset_index("Hub", inplace=True)
+
+                df_pkm = df_Unit_t_pkm[['Units_supply']].unstack(level='Unit')
+                df_pkm['PT'] = df_Grid_t_pkm['Grid_supply']
+                df_pkm['Domestic_energy'] = df_Grid_t_pkm['Domestic_energy']
+                
+
+                df_EV = df_Unit_t[(df_Unit_t.index.get_level_values("Layer") == "Electricity") & (df_Unit_t.index.get_level_values("Unit").str.contains('EV')) ] # TODO more properly than .str.contains('EV') ? 
+
+
+                # saving
+                if Pareto_ID == 0:
+                    result_file_path = 'results/' + str(filename) + '_' + str(Scn_ID) + '.xlsx'
+                else:
+                    result_file_path = 'results/' + str(filename) + '_' + str(Scn_ID) + str(Pareto_ID) + '.xlsx'
+                
+                writer = pd.ExcelWriter(result_file_path)
+                df_pkm.to_excel(writer, sheet_name='df_pkm')
+                df_EV.to_excel(writer, sheet_name='df_EV')
+
+                writer.close()
+                print('Results are saved in ' + result_file_path)
+
+
+                
