+++ reho/plotting/plotting.py
@@ -1775,7 +1775,7 @@ def unit_monthly_plot(results, to_plot, label='EN_short', save_path="", filename
 
     return fig
 
-def plot_pathway(results, scn_id=None, EMOO_list=[], y_span=[], label='EN_short', save_path="", filename=None, export_format='html', objective=None, constraint=None):
+def plot_pathway(results, method='district-scale',scn_id=None, EMOO_list=[], y_span=[], label='EN_short', save_path="", filename=None, export_format='html', objective=None, constraint=None):
     """
         Generate a line plot showing the investment pathway for the respective Objective function and EMOO selected
 
@@ -1811,20 +1811,25 @@ def plot_pathway(results, scn_id=None, EMOO_list=[], y_span=[], label='EN_short'
     if y_span==[]:
         raise SystemExit('Please enter the list of years used in the pathway optimization')
 
-    # Plot
-    #fig, (ax1, ax2) = plt.subplots(1, 2, sharex=False, figsize=(20,7))
-    fig, ax1=plt.subplots()
+    ###################################
+    ##### First plot: EMOO pathway ####
+    ###################################
 
-    # ax1.set_xlabel('Year',fontsize=20)
-    # ax1.set_xlim((min(y_span), max(y_span)))
+    # Initialize plot
+    fig, ax1=plt.subplots(figsize=(12,7))
+
+    # Set title
     ax1.set_title("Investment Pathway", fontsize=35)
+
+    # Set x-axis parameters
     ax1.tick_params(axis='x', labelsize=15)
 
+    # Retrieve and plot results
     obj=[]
     obj2=[]
     add_line=False
-    # if -1 in results[scn_id].keys():
     if objective=='CAPEX':
+        objective = 'Cumulative CAPEX'
         obj_label = 'Cumulative capital investment [CHF/m2/y]'
         for i in range(len(EMOO_list)):
             obj=obj+[results[scn_id][i]['df_KPIs']['capex_m2']['Network']]
@@ -1833,20 +1838,22 @@ def plot_pathway(results, scn_id=None, EMOO_list=[], y_span=[], label='EN_short'
         for i in range(len(EMOO_list)):
             obj = obj + [results[scn_id][i]['df_KPIs']['capex_m2']['Network']]
             obj2= obj2 + [results[scn_id][i]['df_KPIs']['opex_m2']['Network']]
+        obj_old = objective
         objective='Cumulative CAPEX'
+
         add_line=True
         lns4 = ax1.plot(y_span, obj2, label='OPEX', color='C0', linewidth=2, linestyle='--')
-
     lns1 = ax1.plot(y_span, np.cumsum(obj), label=objective, color='C0', linewidth=2)
-
 
     # Left y-axis params
     ax1.set_ylabel(obj_label, color='C0', fontsize=20)
     ax1.tick_params(axis='y', labelcolor='C0', labelsize=15)
     ax1.set_ylim(bottom=0)
 
-    # Creating a second y-axis
+    # Creat a second y-axis for EMOO values
     ax11 = ax1.twinx()
+
+    # Retrieve data for second axis
     constr=[]
     if constraint=='GWP':
         constr_label='GWP [kgCO2-eq/m2]'
@@ -1855,19 +1862,30 @@ def plot_pathway(results, scn_id=None, EMOO_list=[], y_span=[], label='EN_short'
 
     elif constraint == 'elec_export':
         constr_label='Electricity Exportation [MWh/m2/y]'
-        for i in range(len(EMOO_list)):
[…]
