### 2024-04-29 #6542 @willu47
I do not have an AMPL license and therefore cannot test the package. I also think it is a bit disingenuous of the authors to call their package "fully open source", yet wrap it around a commercial dependency.  Can AMPL be substituted for the open source GNUMathProg, for example?
