### 2024-04-29 #6542 @mbarzegary
@willu47 thank you for raising this important point. I believe this is a serious issue that we missed. Even the community edition of AMPL seems to be quite limiting. I will check it with the AEiC team.
