+++ reho/model/preprocessing/skydome.py
@@ -10,6 +10,7 @@ __doc__ = """
 Used for PV orientation.
 """
 
+
 def convert_results_txt_to_csv(load_timesteps):
     """ load one txt file for each hour/timestep in "load_timesteps".
     txt file contains oriented irradiation of each 145 skypatches. Combines all hours to one df
@@ -32,7 +33,7 @@ def convert_results_txt_to_csv(load_timesteps):
 
     # hour 1 is between 0:00 - 1:00 and is indexed with starting hour so 0:00
     for h in df.index.values:
-        df.loc[h, 'time'] = t1 + timedelta(hours=(int(h)-1))
+        df.loc[h, 'time'] = t1 + timedelta(hours=(int(h) - 1))
 
     df = df.set_index('time')
     output_file = os.path.join(path_to_skydome, 'total_irradiation.csv')
@@ -41,10 +42,17 @@ def convert_results_txt_to_csv(load_timesteps):
 
 
 def skydome_to_df(local_data):
-    """ reads two txt files: one containing the area and one the position of the center point of the 145 patches,
-     which define the skydome. Calculates basic additional values and returns all data in one single df
+    """
+    Reads two txt files: one containing the area and one the position of the center point of the 145 patches,
+    which define the skydome. Calculates basic additional values and returns all data in one single df
 
-    :return:   df
+    The irradiation dome is defined as a composition of patches, defined by their area and their centroids. Those
+    patches do not depend on the location. The sky is always decomposed the same way, as done by
+    `Lady Bug tool <https://www.ladybug.tools/ladybug/docs/index.html>`_.
+
+    Returns
+    -------
+       pd.DataFrame
     """
     df_area = local_data["df_Area"]  # area of patches
     df_cenpts = local_data["df_Cenpts"]  # location of centre points
@@ -58,12 +66,12 @@ def skydome_to_df(local_data):
     # Add basic caluclations
     df_dome['XY'] = df_dome[['X', 'Y']].apply(f_sqrt, axis=1)
     df_dome['XYZ'] = df_dome[['XY', 'Z']].apply(f_sqrt, axis=1)
-    df_dome['Sin_e'] = round(df_dome['Z']/df_dome['XYZ'],4)
-    df_dome['Cos_e'] = round(df_dome['XY']/df_dome['XYZ'],4)
-    df_dome['Cos_a'] = round(df_dome['Y']/df_dome['XY'],4)
-    df_dome['Sin_a'] = round(df_dome['X']/df_dome['XY'],4)
+    df_dome['Sin_e'] = round(df_dome['Z'] / df_dome['XYZ'], 4)
+    df_dome['Cos_e'] = round(df_dome['XY'] / df_dome['XYZ'], 4)
+    df_dome['Cos_a'] = round(df_dome['Y'] / df_dome['XY'], 4)
+    df_dome['Sin_a'] = round(df_dome['X'] / df_dome['XY'], 4)
     df_dome['azimuth'] = df_dome[['X', 'Y']].apply(f_atan, axis=1)
-    df_dome['elavation'] = df_dome[['Z', 'XY']].apply(f_atan, axis=1)
+    df_dome['elevation'] = df_dome[['Z', 'XY']].apply(f_atan, axis=1)
 
     return df_dome
 
@@ -72,16 +80,16 @@ def irradiation_to_df(ampl, df_irradiation, df_time):
     """reads Irradiation values of all 145 for the timesteps given in the csv file.
      Converts them to float and returns them as df"""
 
-    #change column name from string to int
+    # change column name from string to int
     df_irradiation.columns = df_irradiation.columns.astype(int)
[…]
