### 2024-04-29 47dec01
remove weather files from repository
