+++ scripts/examples/1a_building-scale_totex.py
@@ -9,7 +9,7 @@ if __name__ == '__main__':
     qbuildings_data = reader.read_db(transformer=3658, nb_buildings=2)
 
     # Select weather data
-    cluster = {'Location': 'Geneva', 'Attributes': ['I', 'T', 'W'], 'Periods': 10, 'PeriodDuration': 24}
+    cluster = {'Location': 'Geneva', 'Attributes': ['T', 'I', 'W'], 'Periods': 10, 'PeriodDuration': 24}
 
     # Set scenario
     scenario = dict()
