+++ scripts/template/run.py
@@ -1,32 +1,34 @@
 from reho.model.reho import *
 
 
-# Set building parameters
-reader = QBuildingsReader()
-qbuildings_data = reader.read_csv(buildings_filename='data/buildings.csv', nb_buildings=1)
+if __name__ == '__main__':
 
-# Select weather data
-cluster = {'Location': 'Geneva', 'Attributes': ['I', 'T', 'W'], 'Periods': 10, 'PeriodDuration': 24}
+    # Set building parameters
+    reader = QBuildingsReader()
+    qbuildings_data = reader.read_csv(buildings_filename='data/buildings.csv', nb_buildings=1)
 
-# Set scenario
-scenario = dict()
-scenario['Objective'] = 'TOTEX'
-scenario['EMOO'] = {}
-scenario['specific'] = []
-scenario['name'] = 'totex'
-scenario['exclude_units'] = ['ThermalSolar', 'NG_Cogeneration']
-scenario['enforce_units'] = []
+    # Select weather data
+    cluster = {'custom_weather': 'data/profiles/Sion.csv', 'Location': 'Sion', 'Attributes': ['T', 'I', 'W'], 'Periods': 10, 'PeriodDuration': 24}
 
-# Initialize available units and grids
-grids = infrastructure.initialize_grids()
-units = infrastructure.initialize_units(scenario, grids)
+    # Set scenario
+    scenario = dict()
+    scenario['Objective'] = 'TOTEX'
+    scenario['EMOO'] = {}
+    scenario['specific'] = []
+    scenario['name'] = 'totex'
+    scenario['exclude_units'] = ['ThermalSolar', 'NG_Cogeneration']
+    scenario['enforce_units'] = []
 
-# Set method options
-method = {'building-scale': True}
+    # Initialize available units and grids
+    grids = infrastructure.initialize_grids()
+    units = infrastructure.initialize_units(scenario, grids)
 
-# Run optimization
-reho = reho(qbuildings_data=qbuildings_data, units=units, grids=grids, cluster=cluster, scenario=scenario, method=method, solver="highs")
-reho.single_optimization()
+    # Set method options
+    method = {'building-scale': True}
 
-# Save results
-reho.save_results(format=['xlsx', 'pickle'], filename='my_case_study')
+    # Run optimization
+    reho = reho(qbuildings_data=qbuildings_data, units=units, grids=grids, cluster=cluster, scenario=scenario, method=method, solver="highs")
+    reho.single_optimization()
+
+    # Save results
+    reho.save_results(format=['xlsx', 'pickle'], filename='my_case_study')
