+++ reho/model/preprocessing/QBuildings.py
@@ -170,7 +170,7 @@ class QBuildingsReader:
 
         return qbuildings
 
-    def read_db(self, transformer=None, nb_buildings=None, egid=None, to_csv=False, return_location=False):
+    def read_db(self, transformer=None, nb_buildings=None, egid=None, to_csv=False):
         """
         Reads the database and extract from it the buildings required, by the LV transformer's ID.
 
@@ -188,8 +188,6 @@ class QBuildingsReader:
             To specify a list of buildings to optimize with their EGIDs
         to_csv : bool
             To export the data into csv
-        return_location : bool
-            To obtain the corresponding meteo cluster, in the returned dictionary under the key ``Location``
 
         Returns
         -------
@@ -294,9 +292,6 @@ class QBuildingsReader:
 
         if qbuildings["buildings_data"] == {}:
             raise print("Empty building data")
-
-        if return_location:
-            qbuildings['Location'] = translate_meteo_to_period_cluster(self.data['transformers']['meteo'][0])
 
         return qbuildings
 
@@ -540,16 +535,6 @@ def translate_roofs_to_REHO(df_roofs):
     return df_roofs
 
 
-def translate_meteo_to_period_cluster(location):
-    dict_meteo = {'Zermatt': 'Disentis',
-                  'Geneva': 'Geneva',
-                  'Berne': 'Bern-Liebefeld',
-                  'CHDF': 'Piotta',
-                  'Zurich': 'Zuerich-SMA',
-                  'Gruyères': 'Lugano'}  # TODO set real cluster
-    return dict_meteo[location]
-
-
 def get_roofs(self, buildings):
     selected_roofs = []
     for i, building in buildings.iterrows():
@@ -585,7 +570,7 @@ def calculate_id_building_shadows(df_angles, id_building, local_data):
         df = df_angles.loc[(df_angles['cosa2'] > 0)].copy()
         # filter buildings which are more than 180 degree apart from patch with az
         df.loc[:, 'tanba'] = df.tanb * df.cosa2
-        # calculate tan(beta) for all buildings. Assumption: dxy is shortest distance and buildings infinite wide
+        # calculate tan(beta) for all buildings. Assumption: dxy is the shortest distance and buildings infinite wide
 
         max_tanba = df['tanba'].max()  # get max obscurance tan(beta, alpha)
         max_b = math.degrees(math.atan(max_tanba))  # get angle
@@ -616,7 +601,7 @@ def neighbourhood_angles(buildings, facades):
         # exclude current building to avoid division with zero
         facades_build = facades[facades.id_building == id_building]  # facades of building
         for f in facades_build.index:
-            df_c = pd.DataFrame(index=df_district.index)  # df for calculating values for each facades
+            df_c = pd.DataFrame(index=df_district.index)  # df for calculating values for each facade
             df_c['dx'] = df_district.x.values - facades_build.loc[f]['CX']
             df_c['dy'] = df_district.y.values - facades_build.loc[f]['CY']
             df_c['dxy'] = (df_c.dx * df_c.dx + df_c.dy * df_c.dy) ** 0.5
