### 2024-04-29 #6542 @JosephLoustau
Dear all,

Thank you for your interest for our work !

Using REHO indeed requires an AMPL license, but as stated in our documentation ([Getting started - AMPL](https://reho.readthedocs.io/en/main/sections/5_Getting_started.html#ampl)) :

> "The AMPL Community Edition offers a free, full-powered AMPL license for personal, academic, and commercial-prototyping use." (cf [ampl website](https://ampl.com/ce/)).

In this way there is no limitation on the number of variables and constraints, and REHO is fully compatible with this license used in combination with an open-source solver such as HiGHS. As such, REHO is effectively unsuitable for a commercial application, but operates just fine as research software.

We hope this clears up the doubts you have about the matter, and remain at your disposal to provide any guidance you may require.
