### 2024-04-30 #6542 @mbarzegary
@JosephLoustau thank you for clarifying this.

I asked the AEiC team, and this is a summary:

> This case comes up occasionally. To publish the paper in JOSS, the reviewers certainly need to be able to test/run the code fully, so the authors need to either make their code compatible with openly available tools or find reviewers who have access to what it needs.

@willu47 according the explanation above and the compatibility of the code with the AMPL community edition (and HiGHS solver), do you think you are able to review this submission?
