### 2024-04-30 #6542 @mbarzegary
@nmstreethran thanks for your reply. Can you please confirm that you have no issue with the raised point (working with AMPL community edition)?
