### 2024-05-02 #6542 @willu47
Hi @mbarzegary - given the guidance of the AEiC team I can review the project.  I will be able to install the AMPL community edition and HIGHs.

I am available to conduct the review from the 20th May 2024.
