+++ scripts/solene/iterative_scenario.py
@@ -11,20 +11,29 @@ def compute_iterative_parameters(variables,parameters):
                             "charging_externalload"  : variables[d2]["externaldemand"].stack().to_frame(name = "charging_externalload").reorder_levels([2,0,1])
                         }
     else:
-        # temporary weights
+        # temporary weights (for loads => loads should be replaced by district activity index, prices proxy to be found)
         weights = pd.DataFrame(index=variables.keys(),columns=['leisure','work','travel']).fillna(1/len(variables.keys()))
-
-        df_load = pd.DataFrame(columns = variables.keys(),index=variables[list(variables.keys())[0]]['externalload'].stack().to_frame().reorder_levels([2,0,1]).index).fillna(0)
+        df_load = pd.DataFrame(columns = variables.keys(),index=    variables[list(variables.keys())[0]]['externalload'].stack().to_frame().reorder_levels([2,0,1]).index).fillna(0)
+        df_prices = pd.DataFrame(columns = variables.keys(),index=  variables[list(variables.keys())[0]]['pi'].index).fillna(0)
         for d in variables.keys(): # TODO : a refaire avec pas des boucles un jour :/
             for i in variables.keys():
                 if i != d:
                     w = weights.loc[i]/(1-weights.loc[d])
-                    p = variables[d]['externaldemand'].mul(w)
-                    p.columns.name = 'Activity'
-                    p = p.stack().to_frame(name = i).reorder_levels([2,0,1])
-                    df_load[[i]] = df_load[[i]]  + p
+                    #loads
+                    l = variables[d]['externaldemand'].mul(w)
+                    l.columns.name = 'Activity'
+                    l = l.stack().to_frame(name = i).reorder_levels([2,0,1])
+                    df_load[[i]] = df_load[[i]]  + l
 
-            parameters[d] = {"charging_externalload" : df_load[[d]].rename(columns={d :"charging_externalload"})}
+                    #prices
+                    wp = 1/(len(variables.keys()) - 1) # temporary weight
+                    p = variables[d]['pi'].mul(wp)
+                    p.name = i
+                    df_prices[i] = df_prices[i].add(p)
+
+
+            parameters[d] = {"charging_externalload" : df_load[[d]].rename(columns={d :"charging_externalload"}),
+                             "outside_charging_price" : df_prices[[d]].rename(columns={d :"outside_charging_price"})}
 
 
 
@@ -54,7 +63,7 @@ def check_convergence(deltas,df_delta,variables):
     # Check no_improvement criteria
     count = 0
     for n in range(len(deltas) - 1, -1, -1):
-        t = abs((deltas[i] - deltas[i-1])/deltas[i])
+        t = abs((deltas[n] - deltas[n-1])/deltas[n])
         if t < termination_threshold:
             count += 1
         else:
@@ -70,7 +79,7 @@ def check_convergence(deltas,df_delta,variables):
 
 if __name__ == '__main__':
 
-    districts = [3658,3112]
+    districts = [3658,3112,277]
 
     # Initialization of scenarios - Generic parameters
     ## Set building parameters
@@ -111,7 +120,10 @@ if __name__ == '__main__':
         vars()['reho_' + str(transformer)] = reho(qbuildings_data=qbuildings_data, units=units, grids=grids, cluster=cluster, scenario=scenario,
                     method=method, parameters=parameters, solver="baron")
 
-
+    # if you want to add non-zero params  ["outside_charging_price","charging_externalload"] for the init scenario
+    PARAM_INIT = False
+    with open('data/mobility/parameters_iteration.pickle', 'rb') as handle:
+        parameters_init = pickle.load(handle)
 
 
     # Run optimization
@@ -127,9 +139,13 @@ if __name__ == '__main__':
[…]
