### 2024-05-03 #6542 @nmstreethran
> @nmstreethran thanks for your reply. Can you please confirm that you have no issue with the raised point (working with AMPL community edition)

Yes, that's not an issue for me @mbarzegary
