### 2024-05-03 acba5f3
EV model
