+++ reho/model/reho.py
@@ -655,8 +655,14 @@ class reho(district_decomposition):
         self.method['building-scale'] = method
         self.initialize_optimization_tracking_attributes()
 
-    def pathway(self,EMOO_list=[400],EMOO_type="GWP",existing_init=None):
+    def pathway(self,EMOO_list=[400],EMOO_type="GWP",existing_init=None,EV=[],EV_battery=[]):
+
         if existing_init is None:
+            if EV != []:
+                self.parameters["n_vehicles"] = EV[0]
+            if EV_battery != []:
+                self.parameters["Units_Ext_district"] = np.array(
+                    [EV_battery[1] if "Battery_district" in u else 0 for u in self.infrastructure.UnitsOfDistrict])
             Scn_ID=self.scenario["name"]
             if 'EMOO' not in self.scenario.keys():
                 self.scenario['EMOO'] ={'EMOO_'+EMOO_type:EMOO_list[0]}
@@ -664,12 +670,18 @@ class reho(district_decomposition):
                 self.scenario['EMOO']['EMOO_'+EMOO_type]=EMOO_list[0]
             self.single_optimization(Pareto_ID=0)
         else:
+            if EV != []:
+                self.parameters["n_vehicles"] = EV[0]
+            if EV_battery != []:
+                self.parameters["Units_Ext_district"] = np.array(
+                    [EV_battery[0] if "Battery_district" in u else 0 for u in self.infrastructure.UnitsOfDistrict])
             if 'Units' in existing_init.keys():
                 existing_units=existing_init['Units']
                 self.parameters["Units_Ext"] = np.array([existing_units[existing_units.index.map(lambda x: h in x)].loc[
                                                              [s for s in self.infrastructure.Units if h in s]][
                                                              'Units_Mult'].to_list() for id, h in
                                                          enumerate(self.infrastructure.houses)])
+                self.parameters["Units_Ext_district"]=existing_units[existing_units.index.str.contains('district')]['Units_Mult'].values
             if 'Transformer' in existing_init.keys():
                 self.parameters["Transformer_Ext"]=existing_init['Transformer']
             if 'Lines' in existing_init.keys():
@@ -681,11 +693,20 @@ class reho(district_decomposition):
                 self.scenario['EMOO'] ={'EMOO_'+EMOO_type:EMOO_list[0]}
             else:
                 self.scenario['EMOO']['EMOO_'+EMOO_type]=EMOO_list[0]
+            if EV != []:
+                self.parameters["n_vehicles"] = EV[0]
+            if EV_battery != []:
+                Additional_battery = np.array([EV_battery[1] if "Battery_district" in u else 0 for u in
+                                               self.infrastructure.UnitsOfDistrict]) - np.array(
+                    [EV_battery[0] if "Battery_district" in u else 0 for u in self.infrastructure.UnitsOfDistrict])
+            else:
+                Additional_battery = np.array([0 for u in self.infrastructure.UnitsOfDistrict])
             existing_units = self.results[Scn_ID][-1]['df_Unit'][['Units_Mult']]
             self.parameters["Units_Ext"] = np.array([existing_units[existing_units.index.map(lambda x: h in x)].loc[
                                                          [s for s in self.infrastructure.Units if h in s]][
                                                          'Units_Mult'].to_list() for id, h in
[…]
