### 2024-05-06 #6542 @mbarzegary
Hi @hgandhi2411 👋
Considering your expertise, would you be willing to review this submission for JOSS? As you may know, we carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Paper: https://raw.githubusercontent.com/openjournals/joss-papers/joss.06542/joss.06542/10.21105.joss.06542.pdf
Software: https://github.com/IPESE/REHO
