### 2024-05-08 6c6d4e0
iterative scenario runs with 3 districts and 1 activity (work)
