+++ scripts/solene/iterative_scenario.py
@@ -3,37 +3,72 @@ import pandas as pd
 from reho.model.reho import *
 import datetime
 
+def rho_param(ext_districts,S,activities = ["work","leisure","travel"]):
+    """
+    This function is used to calculate the parameter rho from the share of activities S(a) in each districts.
+    To be updated when I have the values S.
+    """
+    rho = pd.DataFrame(index=ext_districts,columns=activities).fillna(1/len(ext_districts))
+    rho = rho.stack().to_frame(name= "share_district_activity").reorder_levels([1,0])
+    rho.loc['travel'] = 0 # additionnal precaution
+    return rho
+
+
+
+
 def compute_iterative_parameters(variables,parameters):
-    # for 2 districts:
-    if len(variables.keys()) == 2:
-        for d1,d2 in zip(variables.keys(),reversed(variables.keys())):
-            parameters[d1] = {    "outside_charging_price" : variables[d2]['pi'].rename("outside_charging_price"),
-                            "charging_externalload"  : variables[d2]["externaldemand"].stack().to_frame(name = "charging_externalload").reorder_levels([2,0,1])
-                        }
-    else:
-        # temporary weights (for loads => loads should be replaced by district activity index, prices proxy to be found)
-        weights = pd.DataFrame(index=variables.keys(),columns=['leisure','work','travel']).fillna(1/len(variables.keys()))
-        df_load = pd.DataFrame(columns = variables.keys(),index=    variables[list(variables.keys())[0]]['externalload'].stack().to_frame().reorder_levels([2,0,1]).index).fillna(0)
-        df_prices = pd.DataFrame(columns = variables.keys(),index=  variables[list(variables.keys())[0]]['pi'].index).fillna(0)
-        for d in variables.keys(): # TODO : a refaire avec pas des boucles un jour :/
-            for i in variables.keys():
-                if i != d:
-                    w = weights.loc[i]/(1-weights.loc[d])
-                    #loads
-                    l = variables[d]['externaldemand'].mul(w)
-                    l.columns.name = 'Activity'
-                    l = l.stack().to_frame(name = i).reorder_levels([2,0,1])
-                    df_load[[i]] = df_load[[i]]  + l
+    df_load = pd.DataFrame()
+    df_prices = pd.DataFrame()
+    for d in variables.keys():
+        df_load = pd.concat([df_load,variables[d]['externaldemand']])
+        price = variables[d]['pi'].to_frame().copy()
+        price['district'] = d
+        price = price.set_index('district',append = True).reorder_levels([2,0,1])
+        df_prices = pd.concat([df_prices,price])
 
-                    #prices
-                    wp = 1/(len(variables.keys()) - 1) # temporary weight
-                    p = variables[d]['pi'].mul(wp)
-                    p.name = i
-                    df_prices[i] = df_prices[i].add(p)
+    df_load = df_load.groupby(["district" ,"Period", "Time"]).agg('sum').stack()
+    df_load = df_load.unstack(level='district').reorder_levels([2,0,1])
+    
+    for d in variables.keys():
+            parameters[d] = {   "charging_externalload"     : df_load[[str(d)]].rename(columns={str(d) :"charging_externalload"}),
+                                "outside_charging_price"    : df_prices[df_prices.index.get_level_values(level="district") != d].rename(columns = {'pi' : 'outside_charging_price'}),
+                                "externalload_sellingprice" : variables[d]['pi'].rename("externalload_sellingprice")}
 
 
-            parameters[d] = {"charging_externalload" : df_load[[d]].rename(columns={d :"charging_externalload"}),
[…]
