+++ scripts/solene/mobility_scenario.py
@@ -36,7 +36,6 @@ if __name__ == '__main__':
 
     # Set specific parameters
     # reho.parameters['TransformerCapacity'] = np.array([1e6, 1e6, 0, 1e6])  # TODO : robustesse of mobility Network
-    reho.parameters['Population'] = 9
     reho.parameters['Mode_Speed'] = {'Bike2_district': 20}
     # reho.parameters['DailyDist'] = 9
 
@@ -52,26 +51,6 @@ if __name__ == '__main__':
     date = datetime.datetime.now().strftime("%d_%H%M")
     reho.save_results(format=['xlsx', 'pickle'], filename='Mob')
     reho.save_mobility_results(filename = f"3f_mobility{date}")
-
-    # # Mobility Formatting of results 
-    # date = datetime.datetime.now().strftime("%d_%H%M")
-    # result_file_path = 'results/Mob_totex.xlsx'
-    # df_Unit_t = pd.read_excel(result_file_path, sheet_name="df_Unit_t",
-    #                           index_col=[0, 1, 2, 3])  # refaire sans passer par le xlsx
-    # df_Grid_t = pd.read_excel(result_file_path, sheet_name="df_Grid_t",
-    #                           index_col=[0, 1, 2, 3])
-    # # df_Unit_t.index.name = ['Layer','Unit','Period','Time']
-    # df_Unit_t = df_Unit_t[df_Unit_t.index.get_level_values("Layer") == "Mobility"]
-    # df_Grid_t = df_Grid_t[df_Grid_t.index.get_level_values("Layer") == "Mobility"]
-    # df_dd = df_Grid_t[df_Grid_t.index.get_level_values("Hub") == "Network"]
-    # # df_dd.index = df_dd.index.droplevel('Hub')
-    # df_dd.reset_index("Hub", inplace=True)
-
-    # df_mobility = df_Unit_t[['Units_demand', 'Units_supply']].unstack(level='Unit')
-    # df_mobility['Domestic_energy'] = df_dd['Domestic_energy']
-    # df_mobility.sort_index(inplace = True)
-    # df_mobility.to_excel(f"results/3f_mobility{date}.xlsx")
-    # print(f"Results are saved in 3f_mobility{date}")
 
     # # getting parameters for iteration
     print(reho.results_MP["totex"][0][0]["df_Dual_t"]["pi"].xs("Electricity"))
