### 2024-05-10 3f2ce0f
debug calculation beta duals and adapt plots EVs
