### 2024-05-13 IPESE/REHO#17 @hgandhi2411
[joss-review] Fix installation instructions in README
README installation instructions direct the users to the "User guide" section of documentation for detailed instructions. However, these instructions are under "Getting Started". Please consolidate the differences between the github instructions and documentation.
I would advise providing either complete installation steps in the README or directing users to the documentation page. The way it stands currently, the steps provided in README are incomplete.
