### 2024-05-13 IPESE/REHO#18 @hgandhi2411
[joss-review]: scripts/template/run.py fails
### Is there an existing issue for this?

- [ ] I have searched the existing issues

### What happened?

Used the pip installation as instructed in reho [documentation](https://reho.readthedocs.io/en/main/sections/5_Getting_started.html). Two issues:
1. To check proper installation, I still have to clone the repo
2. /scripts/template/run.py fails to recognize the `.env` file (copied the example.env file and added ampl.lic path as well as my ampl key) placed in the root directory and fails to find another file `Sion-hour.dat`
Attaching the complete error trace here.

### Version

1.0.2 (Stable)

### Relevant log output

```shell
AMPL_PATH is not defined. Please include a .env file at the project root (e.g., AMPL_PATH='C:/AMPL')
/opt/anaconda3/envs/reho-review/lib/python3.10/site-packages/reho/model/preprocessing/QBuildings.py:426: UserWarning: Geometry column does not contain geometry.
  new_buildings_data[REHO_index] = df_buildings[key]
Traceback (most recent call last):
  File "/Users/hetagandhi/Documents/REHO/scripts/template/run.py", line 30, in <module>
    reho = reho(qbuildings_data=qbuildings_data, units=units, grids=grids, cluster=cluster, scenario=scenario, method=method, solver="highs")
  File "/opt/anaconda3/envs/reho-review/lib/python3.10/site-packages/reho/model/reho.py", line 28, in __init__
    super().__init__(qbuildings_data, units, grids, parameters, set_indexed, cluster, method, solver, DW_params)
  File "/opt/anaconda3/envs/reho-review/lib/python3.10/site-packages/reho/model/district_decomposition.py", line 65, in __init__
    self.File_ID = WD.get_cluster_file_ID(self.cluster)
  File "/opt/anaconda3/envs/reho-review/lib/python3.10/site-packages/reho/model/preprocessing/weather.py", line 70, in get_cluster_file_ID
    df = read_hourly_dat(cluster['Location'])
  File "/opt/anaconda3/envs/reho-review/lib/python3.10/site-packages/reho/model/preprocessing/weather.py", line 85, in read_hourly_dat
    df = np.loadtxt(os.path.join(path_to_weather, 'hour', location + '-hour.dat'), unpack=True, skiprows=1)
  File "/opt/anaconda3/envs/reho-review/lib/python3.10/site-packages/numpy/lib/npyio.py", line 1373, in loadtxt
    arr = _read(fname, dtype=dtype, comment=comment, delimiter=delimiter,
  File "/opt/anaconda3/envs/reho-review/lib/python3.10/site-packages/numpy/lib/npyio.py", line 992, in _read
    fh = np.lib._datasource.open(fname, 'rt', encoding=encoding)
  File "/opt/anaconda3/envs/reho-review/lib/python3.10/site-packages/numpy/lib/_datasource.py", line 193, in open
    return ds.open(path, mode, encoding=encoding, newline=newline)
  File "/opt/anaconda3/envs/reho-review/lib/python3.10/site-packages/numpy/lib/_datasource.py", line 533, in open
    raise FileNotFoundError(f"{path} not found.")
FileNotFoundError: /opt/anaconda3/envs/reho-review/lib/python3.10/site-packages/reho/data/weather/hour/Sion-hour.dat not found.
(reho-review) hetagandhi@Abhisheks-Air template % ls ../../reho/data/weather/
```

### Anything else?

_No response_
