+++ reho/plotting/plotting.py
@@ -796,7 +796,7 @@ def plot_pareto_units(results, objectives=["CAPEX", "OPEX"], label='EN_long', co
     if objectives[1] == "OPEX":
         obj_y = "Costs [CHF/m$^2$yr]"
     elif objectives[1] == "GWP":
-        obj_y = "GWP [kgkgCO2/m$^2$yr]"
+        obj_y = "GWP [kgCO2/m$^2$yr]"
 
     plt.xlabel(obj_x)
     plt.ylabel(obj_y)
@@ -1518,12 +1518,13 @@ def plot_EVs(results, era, label='EN_long', color='ColorPastel'):
     return plt
 
 
-def plot_load_duration_curve(results, ids, save_fig=False, label='EN_long', color='ColorPastel'):
+def plot_load_duration_curve(results, ids, save_fig=False, label='EN_long', color='ColorPastel',scenarios=[]):
     fig, ax = plt.subplots(figsize=(9, 6))
-    axx = ax.twinx()
+
     linstyles = ["-", ":"]
     idx = list(range(1, 8761))
     col = 0.2
+    iter=0
     for res in results:
         for id in ids:
             profile = dict()
@@ -1535,28 +1536,32 @@ def plot_load_duration_curve(results, ids, save_fig=False, label='EN_long', colo
                 expanded_profile = np.repeat(profile[id].xs(i, level="Period").values, int(res[0][id]["df_Time"].dp[i]))
                 periods_profile = np.concatenate((periods_profile, expanded_profile))
             profile[id] = np.sort(periods_profile)
-            ax.plot(idx, profile[id][::-1], linstyles[0], color=[1.0 - col, 0.1, col], label="EV " + str(id - 1) + "0%")
+            ax.plot(idx, profile[id][::-1], linstyles[0],  label="District " +scenarios[iter],color='C'+str(iter))
+
+            iter=iter+1
             col = col + 0.3
     ax.set_ylabel('transformer exchange [kW]', fontsize=19)
     ax.set_xlabel('time [hours]', fontsize=19)
-    ax.legend(loc="upper right", fontsize=17)
-    ax.hlines(570, -20, 12000, linewidth=1.0, linestyle="--", color=layout.loc['Electrical_grid', color])
-    ax.hlines(-570, -20, 12000, linewidth=1.0, linestyle="--", color=layout.loc['Electrical_grid', color])
-    ax.text(5400, 600, 'transformer capacity', color=layout.loc['Electrical_grid', color], fontsize=18)
-    ax.text(3700, 380, 'grid export', color="grey", fontsize=18)
-    ax.text(3700, -520, 'grid import', color="grey", fontsize=18)
-    ax.set_xlim([-2, 8800])
-    ax.set_ylim([-1000, 2000])
+    ax.legend(loc="lower left", fontsize=17)
+    ax.grid()
+    ax.set_title('Load duration curve of each district',fontsize=25)
+    # ax.hlines(570, -20, 12000, linewidth=1.0, linestyle="--", color=layout.loc['Electrical_grid', color])
+    # ax.hlines(-570, -20, 12000, linewidth=1.0, linestyle="--", color=layout.loc['Electrical_grid', color])
+    # ax.text(5400, 600, 'transformer capacity', color=layout.loc['Electrical_grid', color], fontsize=18)
+    # ax.text(3700, 380, 'grid export', color="grey", fontsize=18)
+    # ax.text(3700, -520, 'grid import', color="grey", fontsize=18)
+    # ax.set_xlim([-2, 8800])
+    # ax.set_ylim([-1000, 2000])
     plt.rc('xtick', labelsize=16)
[…]
