### 2024-05-13 fe1ac84
cleaning + refactoring for compliance with PEP8
