+++ reho/model/sub_problem.py
@@ -1,16 +1,16 @@
-import os
-from amplpy import AMPL, Environment
 import itertools as itertools
-import reho.model.preprocessing.buildings_profiles as BP
-import reho.model.preprocessing.weather as WD
-import reho.model.preprocessing.skydome as SKD
-import reho.model.preprocessing.emissions_parser as emissions
-import reho.model.preprocessing.EV_profile_generator as EV_gen
-from reho.model.preprocessing.QBuildings import *
 import logging
 
+from amplpy import AMPL, Environment
 
-class compact_optimization:
+import reho.model.preprocessing.EV_profile_generator as EV_gen
+import reho.model.preprocessing.buildings_profiles as buildings_profiles
+import reho.model.preprocessing.emissions_parser as emissions
+import reho.model.preprocessing.weather as weather
+from reho.model.preprocessing.QBuildings import *
+
+
+class SubProblem:
     """
     Collects all the data input and sends it an AMPL model, solves the optimization.
 
@@ -19,9 +19,9 @@ class compact_optimization:
     district : district
         Instance of the class district, contains relevant structure in the district such as Units or grids.
     buildings_data : dict
-        Dictionary containing relevant Building data.
+        Building-specific data.
     local_data : dict
-        Dictionary containing relevant location-specific data.
+        Location-specific data.
     parameters : dict, optional
         Dictionary containing 'new' parameters for the AMPL model. If incomplete, uses data from buildings_data.
     set_indexed : dict, optional
@@ -35,62 +35,39 @@ class compact_optimization:
     solver : str, optional
         Chosen solver for AMPL (gurobi, cplex, HiGHS, cbc...).
     qbuildings_data : dict, optional
-        Dictionary containing input data for the buildings.
+        Input data for the buildings.
 
     See also
     --------
     reho.model.reho.reho
-    reho.model.district_decomposition.district_decomposition
+    reho.model.master_problem.MasterProblem
 
     """
+
     def __init__(self, district, buildings_data, local_data, parameters, set_indexed, cluster, scenario, method, solver, qbuildings_data=None):
 
-        self.buildings_data_compact = buildings_data
+        self.buildings_data_sp = buildings_data
         if method['use_facades']:
-            self.facades_compact = qbuildings_data['facades_data']
-            self.shadows_compact = qbuildings_data['shadows_data']
+            self.facades_sp = qbuildings_data['facades_data']
+            self.shadows_sp = qbuildings_data['shadows_data']
         if method['use_pv_orientation']:
-           self.roofs_compact = qbuildings_data['roofs_data']
-        self.infrastructure_compact = district
+            self.roofs_sp = qbuildings_data['roofs_data']
+        self.infrastructure_sp = district
         self.local_data = local_data
-        self.parameters_compact = parameters
-        self.set_indexed_compact = set_indexed
-        self.cluster_compact = cluster
+        self.parameters_sp = parameters
+        self.set_indexed_sp = set_indexed
+        self.cluster_sp = cluster
         if 'exclude_units' not in scenario:
             scenario['exclude_units'] = []
         if 'enforce_units' not in scenario:
             scenario['enforce_units'] = []
-        self.scenario_compact = scenario
-        self.method_compact = method
+        self.scenario_sp = scenario
+        self.method_sp = method
         self.solver = solver
         self.parameters_to_ampl = dict()
 
[…]
