+++ scripts/examples/1b_Building-scale_Pareto.py
@@ -9,7 +9,7 @@ if __name__ == '__main__':
     reader.establish_connection('Suisse')
     qbuildings_data = reader.read_db(3658, nb_buildings=1)
 
-    # Select weather data
+    # Select clustering options for weather data
     cluster = {'Location': 'Geneva', 'Attributes': ['T', 'I', 'W'], 'Periods': 10, 'PeriodDuration': 24}
 
     # Set scenario
@@ -28,7 +28,7 @@ if __name__ == '__main__':
     method = {'building-scale': True}
 
     # Run optimization
-    reho = reho(qbuildings_data=qbuildings_data, units=units, grids=grids, cluster=cluster, scenario=scenario, method=method, solver="gurobi")
+    reho = REHO(qbuildings_data=qbuildings_data, units=units, grids=grids, cluster=cluster, scenario=scenario, method=method, solver="gurobi")
     reho.generate_pareto_curve()  # multi-objective optimization
 
     # Save results
