+++ reho/plotting/rainbow_plots.py
@@ -9,35 +9,36 @@ __doc__ = """
 Reconstructs data from a Pareto optimization to produce a rainbow plot.
 """
 
-def linerize_equal_steps (x, y):
+
+def linerize_equal_steps(x, y):
     x.name = 'x'
     y.name = 'y'
     step = (x.max() - x.min()) / len(x)
-    df = pd.concat([x, y], names =['x','y'], axis=1)
+    df = pd.concat([x, y], names=['x', 'y'], axis=1)
 
     df_new = pd.DataFrame(columns=['y'])
-    for s in  np.arange(1,  len(x)+1 ):
-        x_new = s*step
+    for s in np.arange(1, len(x) + 1):
+        x_new = s * step
         ub = df[df.x >= x_new].x.min()
         lb = df[df.x < x_new].x.max()
-        if math.isnan(ub) == True: # happens in case x_new == to last x in array
-            y_new = df[df.x == lb].iloc[0].y # just return last point
+        if math.isnan(ub):  # happens in case x_new == to last x in array
+            y_new = df[df.x == lb].iloc[0].y  # just return last point
         else:
             upper_bound = df[df.x == ub].iloc[0]
             lower_bound = df[df.x == lb].iloc[0]
 
             m = ((upper_bound.y - lower_bound.y) / (upper_bound.x - lower_bound.x))
-            b = lower_bound.y- m * lower_bound.x
+            b = lower_bound.y - m * lower_bound.x
 
-            y_new = m*x_new + b
+            y_new = m * x_new + b
         df_new.at[x_new, 'y'] = y_new
     return df_new
 
-def split_orientation(az):
 
-    if (az < 45) or  (az > 315):
+def split_orientation(az):
+    if (az < 45) or (az > 315):
         orientation = 'North'
-    elif  (az >= 45) and (az < 135):
+    elif (az >= 45) and (az < 135):
         orientation = 'East'
     elif (az >= 135) and (az < 225):
         orientation = 'South'
@@ -48,16 +49,16 @@ def split_orientation(az):
 
     return orientation
 
-def get_roof_max(df_nbr, hs_A):
 
+def get_roof_max(df_nbr, hs_A):
     df_n = df_nbr.groupby(['Pareto_ID', 'Surface', 'Azimuth', 'Tilt']).sum()
     roofs = df_n[df_n.index.get_level_values(level='Tilt') != 90]
 
-    #get total m2 PV
+    # get total m2 PV
     total_roofs = roofs.groupby('Pareto_ID').sum()
-    total_roofs['PV_m2'] = total_roofs.PVA_module_nbr*1.6
+    total_roofs['PV_m2'] = total_roofs.PVA_module_nbr * 1.6
     # get occupation based on azimuth orientation
-    roofs = roofs.reset_index(level = ['Azimuth'])
+    roofs = roofs.reset_index(level=['Azimuth'])
     roofs['Orientation'] = roofs.apply(lambda x: split_orientation(x['Azimuth']), axis=1)
     roofs = roofs.set_index('Orientation', append='True')
     roofs = roofs.groupby(['Pareto_ID', 'Orientation']).sum()
@@ -65,13 +66,13 @@ def get_roof_max(df_nbr, hs_A):
     roofs['occupancy'] = roofs.PVA_module_nbr / max
 
     roofs_max = total_roofs.PV_m2.max()
-    roofs_max_m2 = roofs_max /hs_A
+    roofs_max_m2 = roofs_max / hs_A
 
     facades = df_n[df_n.index.get_level_values(level='Tilt') == 90]
     if not facades.empty:
         # get total m2 facades
         total_facades = facades.groupby('Pareto').sum()
[…]
