+++ reho/model/postprocessing/sensitivity_analysis.py
@@ -1,12 +1,11 @@
-from reho.model.reho import *
-from SALib.sample import sobol as sobol_sample
+import matplotlib.pyplot as plt
+from SALib.analyze import morris as morris_analyze
 from SALib.analyze import sobol as sobol_analyze
 from SALib.sample import morris as morris_sample
-from SALib.analyze import morris as morris_analyze
-import pandas as pd
-import pickle
-import matplotlib.pyplot as plt
+from SALib.sample import sobol as sobol_sample
 from qmcpy import Sobol
+
+from reho.model.reho import *
 
 
 class SensitivityAnalysis:
@@ -16,26 +15,25 @@ class SensitivityAnalysis:
 
     Parameters
     ----------
-    district_name
-    reho_model : reho object
+    reho : reho object
         Model of the district, obtained via the reho class.
     SA_type : str
         Type of SA (Morris or Sobol).
     sampling_parameters : int
-        Number of trajectory for the sampling of the solution space.
+        Number of trajectories for the sampling of the solution space.
     upscaling_factor : int
         To represent the effective ERA of the typical districts.
 
     Notes
     -------
     The framework is designed to be performed using TOTEX minimization but can easily be modified,
-    just change the objective function of the reho_model and the KPI saved into OBJ in the function run_SA()
+    just change the objective function of the reho and the KPI saved into OBJ in the function run_SA()
     """
 
-    def __init__(self, district_name, reho_model, SA_type, sampling_parameters=0, upscaling_factor=1):
+    def __init__(self, reho, SA_type, sampling_parameters=0, upscaling_factor=1):
 
-        self.district_name = district_name
-        self.reho_model = reho_model
+        self.district_name = reho.scenario["name"]
+        self.reho = reho
         self.SA_type = SA_type
         self.upscaling_factor = upscaling_factor
         self.sampling_parameters = sampling_parameters
@@ -55,16 +53,16 @@ class SensitivityAnalysis:
         f.close()
 
     def get_lists(self):
-        unit_list = self.reho_model.infrastructure.UnitTypes.tolist()
+        unit_list = self.reho.infrastructure.UnitTypes.tolist()
         try:
-            for excluded_unit in self.reho_model.scenario['exclude_units']:
+            for excluded_unit in self.reho.scenario['exclude_units']:
                 unit_list.remove(excluded_unit)
         except ValueError:
             pass
         KPI_list = ['OPEX', 'CAPEX', 'TOTEX', 'GWP']
         return unit_list, KPI_list
 
-    def build_SA(self, unit_parameter=['Cost_inv1', 'Cost_inv2'], SA_parameter={}):
+    def build_SA(self, unit_parameter=['Cost_inv1', 'Cost_inv2'], SA_parameters={}):
         """
         Description:
         - Generate the list of parameters for the SA, their values and type of variation range
@@ -74,7 +72,7 @@ class SensitivityAnalysis:
         Parameters
         ----------
         unit_parameter [list]: Units parameters wanted in the SA, by default all parameters are included
-        SA_parameter [dict]: Parameters with their bounds set by the script
+        SA_parameters [dict]: Parameters with their bounds set by the script
 
         Returns
         ---------
@@ -88,10 +86,12 @@ class SensitivityAnalysis:
 
         # 1) Generate the list of parameters
 
-        default_units_values = self.reho_model.infrastructure.Units_Parameters  # Extract default unit values of the district
-        units = np.unique([unit.split('_Building')[0] for unit in self.reho_model.infrastructure.Units_Parameters.index.to_list()]).tolist()
[…]
