### 2024-05-14 6ac56c7
cleaning reho.py (nesting Pareto related functions)
