+++ reho/model/master_problem.py
@@ -274,7 +274,7 @@ class MasterProblem:
             print('INITIATE HOUSE: ' + h)
 
         # find district structure and parameter for one single building
-        buildings_data_SP, parameters_SP = self.__split_parameter_sets_per_building(h)
+        buildings_data_SP, parameters_SP = self.split_parameter_sets_per_building(h)
 
         # epsilon constraints on districts may lead to infeasibilities on building level -> apply them in MP only
         if epsilon_init is not None and self.method['building-scale']:
@@ -383,8 +383,8 @@ class MasterProblem:
         # Load Master Problem (MP) Formulation
         ampl_MP.cd(path_to_ampl_model)
         ampl_MP.read('master_problem.mod')
-        if self.method["actors_cost"]:
-            ampl_MP.read('master_problem_actors.mod')
+        if self.method["actors_problem"]:
+            ampl_MP.read('actors_problem.mod')
 
         if len(self.infrastructure.UnitsOfDistrict) > 0:
             ampl_MP.cd(path_to_district_units)
@@ -499,7 +499,7 @@ class MasterProblem:
 
         MP_set_indexed['FeasibleSolutions'] = df_Performance.index.unique('FeasibleSolution').to_numpy()  # index to array as set
 
-        if self.method['actors_cost']:
+        if self.method['actors_problem']:
             # MP_parameters['Costs_tot_actors_min'] = df_Performance[["Costs_op", "Costs_inv", "Costs_rep"]].sum(axis=1).groupby("house").min()
             MP_set_indexed['ActorObjective'] = self.set_indexed["ActorObjective"]
 
@@ -645,7 +645,7 @@ class MasterProblem:
 
         self.feasible_solutions += 1  # after each 'round' of SP execution-> increase
 
-    def SP_execution(self, scenario, Scn_ID, Pareto_ID, House):
+    def SP_execution(self, scenario, Scn_ID, Pareto_ID, h):
         """
         Insert dual variables in ampl model, apply scenario, adapt model depending on the methods and get results.
 
@@ -657,7 +657,7 @@ class MasterProblem:
             scenario ID
         Pareto_ID : int
             pareto ID
-        House : string
+        h : string
             house ID
 
         Returns
@@ -671,13 +671,13 @@ class MasterProblem:
         ------
         ValueError: If the SP optimization did not converge
         """
-        self.logger.info('iterate HOUSE: ' + House + 'iteration: ' + str(self.iter))
+        self.logger.info('iterate HOUSE: ' + h + 'iteration: ' + str(self.iter))
 
         # Give dual variables to Subproblem
-        pi = self.get_dual_values_SPs(Scn_ID, Pareto_ID, self.iter - 1, House, 'pi').reorder_levels(['Layer', 'Period', 'Time'])
-        pi_GWP = self.get_dual_values_SPs(Scn_ID, Pareto_ID, self.iter - 1, House, 'pi_GWP').reorder_levels(['Layer', 'Period', 'Time'])
-        pi_lca = self.get_dual_values_SPs(Scn_ID, Pareto_ID, self.iter - 1, House, 'pi_lca')
-        pi_h = pd.concat([pi], keys=[House], names=['House']).reorder_levels(['House', 'Layer', 'Period', 'Time'])
+        pi = self.get_dual_values_SPs(Scn_ID, Pareto_ID, self.iter - 1, h, 'pi').reorder_levels(['Layer', 'Period', 'Time'])
+        pi_GWP = self.get_dual_values_SPs(Scn_ID, Pareto_ID, self.iter - 1, h, 'pi_GWP').reorder_levels(['Layer', 'Period', 'Time'])
+        pi_lca = self.get_dual_values_SPs(Scn_ID, Pareto_ID, self.iter - 1, h, 'pi_lca')
[…]
