+++ reho/model/reho.py
@@ -11,18 +11,16 @@ from reho.paths import *
 
 class REHO(MasterProblem):
     """
-    Wrapper class used to perform the single or multi-objective optimization.
+    Wrapper class used to perform the single or multi-objective optimization (Pareto curve).
 
-    Parameters are inherited from `MasterProblem`
+    Parameters are inherited from `MasterProblem`.
 
     See also
     --------
     reho.model.master_problem.MasterProblem
-
     """
 
-    def __init__(self, qbuildings_data, units, grids, parameters=None, set_indexed=None,
-                 cluster=None, method=None, scenario=None, solver="highs", DW_params=None):
+    def __init__(self, qbuildings_data, units, grids, parameters=None, set_indexed=None, cluster=None, method=None, scenario=None, solver="highs", DW_params=None):
 
         super().__init__(qbuildings_data, units, grids, parameters, set_indexed, cluster, method, solver, DW_params)
         self.initialize_optimization_tracking_attributes()
@@ -55,415 +53,12 @@ class REHO(MasterProblem):
         self.solver_attributes = pd.DataFrame()
         self.epsilon_constraints = {}
 
-    def add_constraints_from_self_scenario(self):
-        scenario = {'EMOO': {'EMOO_grid': self.scenario['EMOO']['EMOO_grid']},
-                    'specific': self.scenario['specific'],
-                    'exclude_units': self.scenario['exclude_units'],
-                    'enforce_units': self.scenario['enforce_units'],
-                    }
-        return scenario
-
-    def return_epsilon_init(self, C_max, C_min, pareto_max, pareto, objective):
-        if self.method['building-scale']:
-            if objective == self.scenario["Objective"][0]:
-                epsilon_init = (C_max - C_min) / (pareto_max + 1) * (pareto - 1) + C_min
-            elif objective == self.scenario["Objective"][1]:
-                epsilon_init = (C_max - C_min) / (pareto_max + 1) * (pareto + 1) + C_min
-            else:
-                epsilon_init = None
-        else:
-            epsilon_init = None
-        return epsilon_init
-
-    def __annualized_investment(self, ampl, surfaces, Scn_ID, Pareto_ID):
-
-        if self.method['building-scale'] or self.method['district-scale']:
-            df_inv = self.results[Scn_ID][Pareto_ID]["df_Performance"]
-            district = (df_inv.Costs_inv[-1] + df_inv.Costs_rep[-1]) / self.ERA
-            buildings = df_inv.Costs_inv[:-1].div(surfaces.ERA) + df_inv.Costs_rep[:-1].div(surfaces.ERA)
-        else:
-            tau = ampl.getParameter('tau').getValues().toList()  # annuality factor
-            df_h = write_results.get_ampl_data(ampl, 'Costs_House_inv', multi_index=False)
-            df1_h = write_results.get_ampl_data(ampl, 'Costs_House_rep', multi_index=False)
-            df = write_results.get_ampl_data(ampl, 'Costs_inv', multi_index=False)
-            df1 = write_results.get_ampl_data(ampl, 'Costs_rep', multi_index=False)
-            # annualized investment costs with replacements (important for BAT and NG_Cogeneration)
-            district = (df.sum()[0] + df1.sum()[0]) * tau[0] / surfaces.sum()[0]  # for compact formulation
-            buildings = (df_h.Costs_House_inv.div(surfaces.ERA) + df1_h.Costs_House_rep.div(surfaces.ERA)) * tau[0]  # for decomposition formulation
-        return district, buildings
-
-    def __opex_per_house(self, ampl, surfaces, Scn_ID, Pareto_ID):
-        if self.method['building-scale'] or self.method['district-scale']:
[…]
