### 2024-05-15 023893a
10 buildings per district iteration
