+++ scripts/solene/10buildings_scenario.py
@@ -1,0 +1,158 @@
+import pandas as pd
+
+from reho.model.reho import *
+from reho.model.preprocessing.EV_profile_generator import *
+import datetime
+
+
+if __name__ == '__main__':
+    date = datetime.datetime.now().strftime("%d_%H%M")
+
+    districts = [3658,3112,277]
+
+    # Initialization of scenarios - Generic parameters
+    ## Set building parameters
+    reader = QBuildingsReader()
+    reader.establish_connection('Suisse')
+
+    ## Select weather data
+    cluster = {'Location': 'Geneva', 'Attributes': ['I', 'T', 'W'], 'Periods': 10, 'PeriodDuration': 24}
+
+    ## Set scenario
+    scenario = dict()
+    scenario['Objective'] = 'TOTEX'
+    scenario['EMOO'] = {}
+    scenario['specific'] = []
+    scenario['name'] = 'totex'
+    scenario['exclude_units'] = ['NG_Cogeneration']
+    scenario['enforce_units'] = ['EV_district']
+
+    ## Initialize available units and grids
+    grids = infrastructure.initialize_grids({'Electricity': {},
+                                             'NaturalGas': {},
+                                             'FossilFuel': {},
+                                             'Mobility': {},
+                                             })
+    units = infrastructure.initialize_units(scenario, grids, district_data=True)
+
+    ## Set method options
+    method = {'building-scale': True}
+
+    # Initialization of scenarios - District parameters
+    for transformer in districts:
+        ## Set building parameters
+        qbuildings_data = reader.read_db(transformer=transformer, nb_buildings=10)
+        functionnal_area = np.sum([qbuildings_data["buildings_data"][bui]["ERA"] for bui in qbuildings_data["buildings_data"]])
+        population = functionnal_area / 46.5 # average m2/pers in Switzerland
+        ## District parameters
+        ext_districts = [d for d in districts if d != transformer]
+        parameters = {'Population': population,
+                      "Districts" : ext_districts }
+                      # "share_district_activity": rho_param(ext_districts,1) } # other districts
+
+        vars()['reho_' + str(transformer)] = reho(qbuildings_data=qbuildings_data, units=units, grids=grids, cluster=cluster, scenario=scenario,
+                    method=method, parameters=parameters, solver="gurobiasl")
+
+    # if you want to add non-zero params  ["outside_charging_price","charging_externalload"] for the init scenario
+    PARAM_INIT = False
+    with open('data/mobility/parameters_standaloneiter2.pickle', 'rb') as handle:
+        parameters_init = pickle.load(handle)
+
+    # Run optimization - iterations
+    df_pi = pd.DataFrame()
+    df_externalcharging = pd.DataFrame()
+    variables = dict()
+    parameters = dict()
+    deltas = list()
+    df_delta = pd.DataFrame()
+    
+    
+    # Standalone initializing run
+    i = 0
+    for transformer in districts:
+        print(f"iteration {i} : district {transformer}")
+
+        # Some customed parameters (if I want to start with another type of INIT RUN)
+        if PARAM_INIT:
+            ext_districts = [d for d in districts if d != transformer]
+            vars()['reho_' + str(transformer)].parameters["Districts"] = ext_districts
+            vars()['reho_' + str(transformer)].parameters["share_district_activity"] = rho_param(ext_districts,1)
+            for param in parameters_init[transformer].keys():
+                vars()['reho_' + str(transformer)].parameters[param] = parameters_init[transformer][param]
+
+        # Run
+        vars()['reho_' + str(transformer)].single_optimization(Pareto_ID=i)
+
[…]
