+++ reho/model/preprocessing/EV_profile_generator.py
@@ -394,3 +394,113 @@ def scenario_profiles_temp(cluster):
         "Daily_Profile": bikedailyprofile
     }
     return output
+
+
+
+def rho_param(ext_districts,S,activities = ["work","leisure","travel"]):
+    """
+    This function is used in the iterative scenario to iteratively calculate multiple districts with EVs being able to charge at the different districts.
+    This function is used to calculate the parameter rho from the share of activities S(a) in each districts.
+    To be updated when I have the values S.
+    """
+    rho = pd.DataFrame(index=ext_districts,columns=activities).fillna(1/len(ext_districts))
+    rho = rho.stack().to_frame(name= "share_district_activity").reorder_levels([1,0])
+    rho.loc['travel'] = 0 # additionnal precaution
+    return rho
+
+
+def compute_iterative_parameters(variables,parameters,only_pi = False):
+    """"
+    This function is used in the iterative scenario to iteratively calculate multiple districts with EVs being able to charge at the different districts.
+    """
+    if only_pi:
+        df_prices = pd.DataFrame()
+        for d in variables.keys():
+            price = variables[d]['pi'].to_frame().copy()
+            price['district'] = d
+            price = price.set_index('district',append = True).reorder_levels([2,0,1])
+            df_prices = pd.concat([df_prices,price])
+
+        for d in variables.keys():
+            parameters[d] = {   "outside_charging_price"    : df_prices[df_prices.index.get_level_values(level="district") != d].rename(columns = {'pi' : 'outside_charging_price'}),
+                                "externalload_sellingprice" : variables[d]['pi'].rename("externalload_sellingprice")
+                                }
+
+    else:
+        df_load = pd.DataFrame()
+        df_prices = pd.DataFrame()
+        for d in variables.keys():
+            df_load = pd.concat([df_load,variables[d]['externaldemand']])
+            price = variables[d]['pi'].to_frame().copy()
+            price['district'] = d
+            price = price.set_index('district',append = True).reorder_levels([2,0,1])
+            df_prices = pd.concat([df_prices,price])
+
+        df_load = df_load.groupby(["district" ,"Period", "Time"]).agg('sum').stack()
+        df_load = df_load.unstack(level='district').reorder_levels([2,0,1])
+        
+        for d in variables.keys():
+                parameters[d] = {   "charging_externalload"     : df_load[[str(d)]].rename(columns={str(d) :"charging_externalload"}),
+                                    "outside_charging_price"    : df_prices[df_prices.index.get_level_values(level="district") != d].rename(columns = {'pi' : 'outside_charging_price'}),
+                                    "externalload_sellingprice" : variables[d]['pi'].to_frame(name = "externalload_sellingprice")}
+
+
+
+def check_convergence(deltas,df_delta,variables,iteration,criteria = ('total')):
+    """"
+    This function is used in the iterative scenario to iteratively calculate multiple districts with EVs being able to charge at the different districts.
+    Parameters
+    ----------
+    criteria : tuple, optional
+        Choose on which indexes to match the load and demand. Default is time matching (p,t). Write 'by_activity' for activity matching (a,p,t)
+
+    """
+    termination_threshold = 0.1 # 10% TODO : mettre ces tuning parametres somewhere else
+    termination_iter = 3
+    
+    # Compute Delta
+    df_demand = pd.DataFrame()
+    df_load = pd.DataFrame()
+    for k in variables.keys():
+        df = variables[k]['externaldemand'].groupby(["Period","Time"]).agg("sum")
+        if 'by_activity' in criteria:
+            df.columns.name = "Activity"
[…]
