### 2024-05-16 2528ccb
README + missing DOI in paper.bib
