+++ reho/plotting/plotting.py
@@ -11,6 +11,7 @@ from reho.model.sub_problem import *
 from reho.plotting.utils import *
 from reho.plotting import sankey
 
+
 __doc__ = """
 Contains ready-to-use representations for results generated by REHO.
 """
@@ -25,7 +26,7 @@ def plot_performance(results, plot='costs', indexed_on='Scn_ID', label='EN_long'
                      scc=0.177,
                      filename=None, export_format='html', scaling_factor=1, return_df=False):
     """
-    Plot performance based on REHO results.
+    Plots performance based on REHO results.
 
     Parameters
     ----------
@@ -60,7 +61,7 @@ def plot_performance(results, plot='costs', indexed_on='Scn_ID', label='EN_long'
     return_df: bool
         A dataframe can be returned for further post-processing or reporting purposes.
 
-   Returns
+    Returns
     ----------
     plotly.graph_objs.Figure
         The generated plotly figure.
@@ -296,7 +297,7 @@ def plot_expenses(results, plot='costs', indexed_on='Scn_ID', label='EN_long', p
                   scc=0.177,
                   filename=None, export_format='html', scaling_factor=1, return_df=False):
     """
-    Plot expenses based on REHO results.
+    Plots expenses based on REHO results.
 
     Parameters
     ----------
@@ -331,7 +332,7 @@ def plot_expenses(results, plot='costs', indexed_on='Scn_ID', label='EN_long', p
     return_df: bool
         A dataframe can be returned for further post-processing or reporting purposes.
 
-   Returns
+    Returns
     ----------
     plotly.graph_objs.Figure
         The generated plotly figure.
@@ -563,7 +564,7 @@ def plot_expenses(results, plot='costs', indexed_on='Scn_ID', label='EN_long', p
 
 def plot_sankey(df_Results, label='EN_long', color='ColorPastel', filename=None, export_format='html', scaling_factor=1, return_df=False):
     """
-    Plot a Sankey plot based on the results DataFrame.
+    Plots a Sankey plot based on the results DataFrame.
 
     Parameters
     ----------
@@ -582,7 +583,7 @@ def plot_sankey(df_Results, label='EN_long', color='ColorPastel', filename=None,
     return_df: bool
         A dataframe can be returned for further post-processing or reporting purposes.
 
-   Returns
+    Returns
     ----------
     plotly.graph_objs.Figure
         The generated plotly figure.
@@ -632,7 +633,7 @@ def plot_sankey(df_Results, label='EN_long', color='ColorPastel', filename=None,
 def plot_profiles(df_Results, units_to_plot, style='plotly', label='EN_long', color='ColorPastel', resolution='weekly', plot_curtailment=False,
                   filename=None, export_format='html', return_df=False):
     """
-    Plot an hourly profile for an entire year of operation.
+    Plots an hourly profile for an entire year of operation.
 
     Parameters
     ----------
@@ -657,7 +658,7 @@ def plot_profiles(df_Results, units_to_plot, style='plotly', label='EN_long', co
     return_df: bool
         A dataframe can be returned for further post-processing or reporting purposes.
 
-   Returns
+    Returns
     ----------
     plotly.graph_objs.Figure
         The generated plotly figure.
@@ -899,7 +900,7 @@ def plot_resources(results, label='EN_long', color='ColorPastel', filename=None,
 
 def plot_sunburst_eud(results, label='EN_long', filename=None, export_format='html', scaling_factor=1, return_df=False):
     """
-    Plot a Sunburst for End Use Demand (EUD) based on REHO results, grouped by buildings' class.
+    Plots a Sunburst for End Use Demand (EUD) based on REHO results, grouped by buildings' class.
 
     Parameters
     ----------
[…]
