+++ reho/plotting/utils.py
@@ -5,9 +5,11 @@ import pandas as pd
 
 from reho.paths import *
 
+
 __doc__ = """
-Contains ready-to-use representations for results generated by REHO.
+Utilities for plotting functions.
 """
+
 
 # Definition of colors and labels
 cm = dict({'ardoise': '#413D3A', 'perle': '#CAC7C7', 'rouge': '#FF0000', 'groseille': '#B51F1F',

[…]
