### 2024-05-16 IPESE/REHO#20 @nmstreethran
[JOSS Review] psycopg2 installation
Hi, I'm reviewing your JOSS submission at https://github.com/openjournals/joss-reviews/issues/6734.

Your recommended installation method is using the requirements file. This currently has both `psycopg2` and `psycopg2-binary`, which is redundant. As stated in the [`psycopg` documentation](https://www.psycopg.org/docs/install.html):

> The psycopg2-binary package is meant for beginners to start playing with Python and PostgreSQL without the need to meet the build requirements.
>
> If you are the maintainer of a published package depending on psycopg2 you shouldn’t use psycopg2-binary as a module dependency. **For production use you are advised to use the source distribution.**

I suggest removing the binary package from your requirements.

I noticed that in your `setup.py`, you use the binary package if the source installation fails (also it seems like this `install_dependencies` function isn't being used anywhere, so the pip installation doesn't install `psycopg2`):

```py
def install_dependencies():
    try:
        subprocess.run([sys.executable, '-m', 'pip', 'install', 'psycopg2'])
    except subprocess.CalledProcessError:
        subprocess.run([sys.executable, '-m', 'pip', 'install', 'psycopg2-binary'])
```

I think you should address the root cause of the `psycopg2` installation error, instead of using the binary as a workaround. The issue is that some prerequisites are missing (i.e. the PostgreSQL library and Python development tools).

These are my steps to successfully install your package on my Ubuntu-based system:

```sh
sudo apt install python3-dev libpq-dev
python -m venv .venv
source .venv/bin/activate
python -m pip install -r requirements.txt
```

See <https://stackoverflow.com/a/5450183>.
