### 2024-05-16 IPESE/REHO#21 @nmstreethran
[JOSS review] Improvements to "Getting started" in docs
https://github.com/openjournals/joss-reviews/issues/6734

In your [documentation](https://reho.readthedocs.io/en/main/sections/5_Getting_started.html), you start off with two alternatives for installing REHO: PyPI and from source. However, the 'Requirements' section only talks about installing REHO from source. I suggest slightly reorganising your steps as follows:

1. Prerequisites - Python 3 and AMPL installations
1. Clone the repository and set up the AMPL licence path and key
1. Installation:
   - from PyPI
   - from source (and checking if it's installed properly)
1. Running REHO

Furthermore, you should recommend users to install the package in a `venv` (you have stated this in the README but not in the docs). Also, the `example.env` file in the docs has only one variable (`AMPL_PATH`), but the one included in the REHO repo has two (`AMPL_PATH` and `API_VESE_KEY`).

Regarding the relative path issue which you have documented in https://github.com/IPESE/REHO/issues/13, I think this is not limited to VS Code. If I run your scripts in a standalone terminal, I still encounter the issue. Also, setting `"python.terminal.executeInFileDir": true` didn't work for me as it produced `FileNotFoundError: The relative path that was given is not a valid file.`. Do you have a solution for this? I think terminals usually expect the paths to be defined relative to the working directory. I was able to run your scripts by first doing `cd scripts/template/` and then `python run.py` and `python plot.py`.

I was able to run `run.py` and `plot.py` when REHO is installed from source, but not when it's installed from PyPI. It looks like the PyPI version is out of date and is missing dependencies, i.e. `coloredlogs` and `pvlib`.

Additionally, when I run `run.py`, the following warning occurs:

```
/mnt/Backup/Downloads/REHO/reho/model/preprocessing/QBuildings.py:436: UserWarning: Geometry column does not contain geometry.
  new_buildings_data[REHO_index] = df_buildings[key]
```
