### 2024-05-16 IPESE/REHO#22 @nmstreethran
[JOSS review] Tests
https://github.com/openjournals/joss-reviews/issues/6734

JOSS requires:

> Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

Do you have any tests implemented? I can't seem to find them.
