### 2024-05-17 0893291
removing psycopg2-binary and improving Getting started
