+++ README.md
@@ -22,7 +22,7 @@ Cédric Terrier <cedric.terrier@epfl.ch>
 ## Licence
 Copyright (C) <2021-2024> <Ecole Polytechnique Fédérale de Lausanne (EPFL), Switzerland>
 
-Licensed under the Apache License, Version0 (the "License");
+Licensed under the Apache License, (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
 
