+++ documentation/sections/5_Getting_started.rst
@@ -1,9 +1,6 @@
 Getting started
 ++++++++++++++++
 
-Installation
-============
-
 .. grid:: 1 2 2 2
     :gutter: 4
 
@@ -35,25 +32,17 @@ Installation
 
             git clone https://github.com/IPESE/REHO.git
 
-.. note::
-    The REHO repository can be forked or cloned depending on the intended use. As soon as the repository is cloned, don't forget to check out your own branch from the ``main`` latest version!
+Prerequisites
+=============
 
-Requirements
+Python 3
 ------------------
-
-Python
-~~~~~~~~~~~~~~~~~~~~~~~~
 
 You will need `Python3 <https://www.python.org/downloads/>`_, just pick the latest version.
 As IDE we recommend to use `PyCharm <https://www.jetbrains.com/pycharm/>`_.
 
-.. warning:: VScode
-
-    If you are a VScode user, you may have path issues, either when importing the REHO module or when providing a relative path.
-    Please refer to the instructions in `REHO/Issues/Relative Path in VScode <https://github.com/IPESE/REHO/issues/13>`_.
-
 AMPL
-~~~~~~~~~~~~~~~~~~~~~~~~
+------------------
 
 As REHO is based on AMPL, it requires a licence of AMPL and at least one LP solver.
 
@@ -63,20 +52,8 @@ As REHO is based on AMPL, it requires a licence of AMPL and at least one LP solv
 
 Plenty of text editors exist which feature AMPL. We recommend using `Sublime Text <https://www.sublimetext.com/>`_, which provides the `AMPL Highlighting package <https://github.com/JackDunnNZ/sublime-ampl>`_.
 
-Setup environment
-~~~~~~~~~~~~~~~~~~~~~~~~
-
-Requirements
-^^^^^^^^^^^^^^^^^^^^^^^
-
-The Python packages required by REHO can be installed with:
-
-.. code-block:: bash
-
-   pip install -r requirements.txt
-
 AMPL license path
-^^^^^^^^^^^^^^^^^^^^^^^
+~~~~~~~~~~~~~~~~~~
 
 You need to include a ``.env`` file at the project root folder. This one should contain the path to your AMPL license (cf. file ``example.env``):
 
@@ -84,22 +61,34 @@ You need to include a ``.env`` file at the project root folder. This one should 
 
    AMPL_PATH = "path_to_your_license"
 
-.. warning:: Relative path error
+.. warning:: Running code from VScode or Terminal
 
-    Getting a *"AMPL_PATH is not defined"* error is frequent for VScode users. Please refer to `REHO/Issues/Relative Path in VScode <https://github.com/IPESE/REHO/issues/13>`_ to overcome this problem.
+    Running scripts from VScode or from a standalone terminal may lead to path issues, either when importing the REHO module or when providing the path to ``.env``.
 
-Git tracking
-^^^^^^^^^^^^^^^^^^^^^^^
+    Please refer to the instructions in `REHO/Issues/Relative Path in VScode <https://github.com/IPESE/REHO/issues/13>`_.
 
-The subfolders ``scripts/examples/`` and ``scripts/template/`` are both git-tracked.
 
-For your future work and own case-studies with REHO, you can create any additional subfolders in ``scripts/``. These will not be git-tracked.
+Installation
+============
 
+From PyPI
+---------
+
+Command
+~~~~~~~
+
+Select the project root containing a Python environment and install the latest stable release of REHO with:
+
+.. code-block:: bash
+
+    pip install --extra-index-url https://pypi.ampl.com REHO
 
 Checking proper installation
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
[…]
