+++ requirements.txt
@@ -9,7 +9,6 @@ scikit-learn>=1.2.2,<2.0.0
 scikit-learn-extra>=0.3.0
 sqlalchemy>=1.4.42,<2.0.0
 psycopg2>=2.9.4,<3.0.0
-psycopg2-binary>=2.9.9,<3.0.0
 geopandas>=0.12.2,<1.0.0
 matplotlib>=3.6.1,<4.0.0
 plotly>=5.10,<6.0.0
