### 2024-05-17 IPESE/REHO#18 @DorsanL
> 1. To check proper installation, I still have to clone the repo

I clarified the « Checking proper installation » paragraph for an installation from PyPI or from source.
Still, I recognize there is no automated test directly within the package. I guess we can discuss this in https://github.com/IPESE/REHO/issues/22 ?

> 2. /scripts/template/run.py fails to recognize the .env file (copied the example.env file and added ampl.lic path as well as my ampl key) placed in the root directory

I updated the documentation and the paragraph [AMPL license path](https://reho.readthedocs.io/en/main/sections/5_Getting_started.html#ampl-license-path), which redirects to [REHO/Issues/Relative Path in VScode](https://github.com/IPESE/REHO/issues/13). I hope this will clarify things.

> 2b. and fails to find another file Sion-hour.dat

This file was indeed not available on the repository, my apologies for this. Fortunately, the use of local weather files is no longer necessary, since now REHO uses the [pvlib](https://pvlib-python.readthedocs.io/en/stable/) library to connect to the PVGIS dabatase and to extract the weather data based on the building's coordinates.
https://github.com/IPESE/REHO/commit/3ecf7cef15e0b50bf977cdc2f7fb9c34fe155b3d

Local weather files (``xx-hour.dat``) have therefore all been removed.
https://github.com/IPESE/REHO/commit/47dec01c998555034f9e24452743eefb8ad6398c
