### 2024-05-17 IPESE/REHO#20 @DorsanL
Thank you for pointing this out and for your feedback for solving it.

``psycopg2-binary`` has been removed from the requirements, as well as the ``install_dependencies()`` function.
https://github.com/IPESE/REHO/commit/0893291bce6685a2088a3e9152bcc27430176c77

The documentation has been improved (cf [Getting started > Installation > From source > Requirements](https://reho.readthedocs.io/en/main/sections/5_Getting_started.html#requirements)), and a warning box added.
