### 2024-05-17 IPESE/REHO#21 @DorsanL
Thank you, your recommendations have been carefully considered!

> I suggest slightly reorganising your steps as follows

A new structure has been set for [Getting started]((https://reho.readthedocs.io/en/main/sections/5_Getting_started.html), based on your suggestions.

> you should recommend users to install the package in a venv

Yes, this has been precised.

> API_VESE_KEY in .env

This ``API_VESE_KEY`` relates to a specific example (``scripts/examples/3i_Electricity_prices.py``) where we connect to an API to extract electricity prices from a public database.
This environment variable is not crucial. If it's not defined, the function ``get_vese_key()`` from ``reho/model/preprocessing/electricity_prices.py`` will retrieve the key from a URL (I just need to conduct some tests to verify it is accessible outside EPFL intranet).

> I think terminals usually expect the paths to be defined relative to the working directory.

I will discuss this with a colleague and try to enhance the documentation regarding such issue. Nice that you still managed to run the scripts!

> I was able to run run.py and plot.py when REHO is installed from source, but not when it's installed from PyPI. It looks like the PyPI version is out of date and is missing dependencies, i.e. coloredlogs and pvlib.

This is correct, the package had not been updated for quite some time. A new [REHO 1.1.0](https://pypi.org/project/REHO/) version is now available, I hope everything will work fine.
