### 2024-05-17 #6734 @DorsanL
Dear all,

Thank you for your interest in our REHO tool and for taking the time and effort to review it.

I have been attentive to the points raised here, as well as to the various issues that have already been opened by hgandhi2411 and nmstreethran.

Here are a few specific updates:

- @mbarzegary the issue of missing DOIs of some references should be fine.
- @hgandhi2411 I answered issues https://github.com/IPESE/REHO/issues/17 and https://github.com/IPESE/REHO/issues/18.
- @nmstreethran I answered issues https://github.com/IPESE/REHO/issues/20, https://github.com/IPESE/REHO/issues/21, and https://github.com/IPESE/REHO/issues/22.

Regarding these issues (and the next that will come), is it preferred that reviewers close them themselves or shall I take care of it?

More generally, I devoted time to clean and refactor many parts of the code, to be as compliant as possible with PEP8 standards. Files, classes and functions have sometimes been renamed and their DocStrings enriched, to improve the overall clarity of the tool.

Two other points to mention regarding the examples:

- REHO can load its input data from .csv files or by connecting to a GIS database. The complete [QBuildings database](https://ipese-web.epfl.ch/lepour/qbuildings/index.html) (which characterizes the entire Swiss building stock) is in the process of being made publicly available (up to now, it requires to be on the EPFL intranet). Meanwhile, we already deployed a public access set of the database for the Geneva region. All examples provided in ``scripts/examples/`` connect to this QBuildings-Geneva database.

- Following willu47’s remark, I verified that all scripts would be running fine with HiGHS. In ``scripts/examples/``, the solver is explicitly defined as Gurobi to reduce calculation time, but one may simply remove the ``REHO(solver="gurobi")`` argument and it will by default substitute the open-source solver HiGHS.

All these recent changes have been pushed to the ``main``, and a version 1.1.0 is available on PyPI.
