### 2024-05-17 #6734 @DorsanL
I’ve just noticed an important discrepancy between the Docstrings and the [Package structure](https://reho.readthedocs.io/en/documentation/sections/4_Package_structure.html) section of the documentation.

Compiling the documentation locally produces the complete and exhaustive summary of all functions and classes. But for a reason yet to be investigated, the latest compilation of the readthedocs has rendered a summary where many important functions are missing (from the principal files such as subproblem_py, master_problem.py, reho.py, plotting.py…)

This shouldn't prevent review in any way (since Docstrings are visible directly in the code), but I'll try to sort it out as soon as possible for your convenience.
