+++ scripts/solene/mobility_scenario.py
@@ -2,10 +2,11 @@ from reho.model.reho import *
 import datetime
 
 if __name__ == '__main__':
+    date = datetime.datetime.now().strftime("%d_%H%M")
     # Set building parameters
     reader = QBuildingsReader()
     reader.establish_connection('Suisse')
-    qbuildings_data = reader.read_db(transformer=3658, nb_buildings=2) #3112
+    qbuildings_data = reader.read_db(transformer=3658, nb_buildings=10) #3112
 
     # Select weather data
     cluster = {'Location': 'Geneva', 'Attributes': ['I', 'T', 'W'], 'Periods': 10, 'PeriodDuration': 24}
@@ -27,7 +28,7 @@ if __name__ == '__main__':
                                              })
     units = infrastructure.initialize_units(scenario, grids, district_data=True)
 
-    parameters = {'Population': 9}
+    parameters = {'Population': 20}
 
     # Set method options
     method = {'building-scale': True}
@@ -36,7 +37,7 @@ if __name__ == '__main__':
 
     # Set specific parameters
     # reho.parameters['TransformerCapacity'] = np.array([1e6, 1e6, 0, 1e6])  # TODO : robustesse of mobility Network
-    reho.parameters['Mode_Speed'] = {'Bike2_district': 20}
+    # reho.parameters['Mode_Speed'] = {'Bike2_district': 20}
     # reho.parameters['DailyDist'] = 9
 
     # reho.parameters.update(EV_gen.scenario_profiles_temp(cluster))
@@ -48,7 +49,6 @@ if __name__ == '__main__':
     reho.single_optimization()
 
     # Save results
-    date = datetime.datetime.now().strftime("%d_%H%M")
     reho.save_results(format=['xlsx', 'pickle'], filename='Mob')
     reho.save_mobility_results(filename = f"3f_mobility{date}")
 
