### 2024-05-22 IPESE/REHO#20 @nmstreethran
Thanks for addressing this!

I have the following recommendations:

1.  Since you suggest Windows users to install `psycopg2-binary`, I recommend modifying your `requirements.txt` with the following:

    ```
    psycopg2>=2.9.4,<3.0.0 ; platform_system != "Windows"
    psycopg2-binary>=2.9.9,<3.0.0 ; platform_system == "Windows"
    ```

    This ensures that the correct package is installed depending on the user's platform. You can then update the documentation to state that the prerequisites for `psycopg2` are only required for Linux and Mac users.

1.  Your `setup.py` is missing `psycopg2`. You can similarly modify `install_requires` to include the following:

    ```
    'psycopg2>=2.9.4,<3.0.0 ; platform_system != "Windows"',
    'psycopg2-binary>=2.9.9,<3.0.0 ; platform_system == "Windows"'
    ```

Feel free to close my issues once you address them.
