### 2024-05-22 IPESE/REHO#22 @nmstreethran
I agree that unit tests are needed. The tests don't need to be automated; manual steps to run them are fine. You can have a look at any recent JOSS papers for suggestions, e.g. [pyheatdemand](https://github.com/AlexanderJuestel/pyheatdemand).
