+++ reho/plotting/plotting.py
@@ -11,7 +11,6 @@ from reho.model.sub_problem import *
 from reho.plotting.utils import *
 from reho.plotting import sankey
 
-
 __doc__ = """
 Contains ready-to-use representations for results generated by REHO.
 """
@@ -589,7 +588,6 @@ def plot_sankey(df_Results, label='EN_long', color='ColorPastel', filename=None,
         The generated plotly figure.
     pd.DataFrame
         (Optional) A dataframe for further post-processing or reporting purposes.
-    
     """
     source, target, value, label_, color_ = sankey.df_sankey(df_Results, label=label, color=color, precision=2,
                                                              units='MWh', display_label_value=True,
@@ -664,7 +662,6 @@ def plot_profiles(df_Results, units_to_plot, style='plotly', label='EN_long', co
         The generated plotly figure.
     pd.DataFrame
         (Optional) A dataframe for further post-processing or reporting purposes.
-
     """
     if resolution == 'monthly':
         items_average = 730
@@ -853,7 +850,7 @@ def plot_profiles(df_Results, units_to_plot, style='plotly', label='EN_long', co
 
 def plot_resources(results, label='EN_long', color='ColorPastel', filename=None, export_format='html'):
     """
-    Under construction.
+    TO DO
     """
     
     scenarios = list(results.keys())
@@ -923,7 +920,6 @@ def plot_sunburst_eud(results, label='EN_long', filename=None, export_format='ht
         The generated plotly figure.
     pd.DataFrame
         (Optional) A dataframe for further post-processing or reporting purposes.
-
     """
 
     def add_class(row):
@@ -1040,7 +1036,6 @@ def plot_unit_monthly(results, unit_to_plot, label='EN_short', filename=None, ex
         The generated plotly figure.
     pd.DataFrame
         (Optional) A dataframe for further post-processing or reporting purposes.
-
     Examples
     --------
     >>> reho_results = pd.read_pickle('results/progressive_scenario.pickle')
@@ -1720,6 +1715,9 @@ def plot_composite_curve(results, cluster, periods=["Yearly"], filename=None, ex
 
 
 def plot_EVs(results, era, label='EN_long', color='ColorPastel'):
+    """
+    TO DO
+    """
     fig, ax = plt.subplots(1, figsize=(5.5, 4.2))
     ax2 = ax.twinx()
     data = {}
