### 2024-05-22 03e5a73
ActorsProblem class + 6a_Actors_problem.py (need to debug)
