+++ reho/model/actors_problem.py
@@ -1,0 +1,134 @@
+from scipy.stats import qmc
+
+from reho.model.reho import *
+
+__doc__ = """
+File for constructing and solving the optimization for the actor-based problem formulation.
+"""
+
+
+class ActorsProblem(REHO):
+    """
+    Performs an actor-based optimization.
+
+    Parameters are inherited from the REHO class.
+
+    See also
+    --------
+    reho.model.reho.REHO
+    
+    Notes
+    -------
+    This class is still under construction.
+    """
+
+    def __init__(self, qbuildings_data, units, grids, parameters=None, set_indexed=None, cluster=None, method=None, scenario=None, solver="highs", DW_params=None):
+
+        super().__init__(qbuildings_data, units, grids, parameters, set_indexed, cluster, method, scenario, solver, DW_params)
+    
+    def execute_actors_problem(self, n_sample=10, bounds=None, actor="Owners"):
+        self.method["actors_problem"] = True
+        self.method["include_all_solutions"] = True
+        self.method['building-scale'] = True
+        self.scenario["Objective"] = "TOTEX_bui"
+        self.set_indexed["ActorObjective"] = np.array([actor])
+        if bounds is not None:
+            sampler = qmc.LatinHypercube(d=2)
+            sample = sampler.random(n=n_sample)
+            l_bound = [bounds[key][0] for key in bounds]
+            u_bound = [bounds[key][1] for key in bounds]
+            samples = pd.DataFrame(qmc.scale(sample, l_bound, u_bound), columns=['utility_portfolio', 'owner_portfolio'])
+            self.samples = samples.round(4)
+        else:
+            self.samples = pd.DataFrame([[None, None]] * n_sample, columns=['utility_portfolio', 'owner_portfolio'])
+    
+        Scn_ID = self.scenario['name']
+        self.pool = mp.Pool(mp.cpu_count())
+        results = {ids: self.pool.apply_async(self.run_actors_optimization, args=(self.samples, ids)) for ids in self.samples.index}
+        while len(results[list(self.samples.index)[-1]].get()) != 2:
+            time.sleep(1)
+        for ids in self.samples.index:
+            df_Results, df_Results_MP = results[ids].get()
+            self.add_df_Results(None, Scn_ID, ids, self.scenario)
+            self.get_KPIs(Scn_ID, ids)
+            self.results_MP[Scn_ID][ids] = df_Results_MP
+    
+        self.samples["objective"] = None
+        for i in self.results_MP[self.scenario["name"]]:
+            if self.results_MP[self.scenario["name"]][i] is not None:
+                self.samples.loc[i, "objective"] = self.results_MP[self.scenario["name"]][i][0]["df_District"]["Objective"]["Network"]
+    
+        self.pool.close()
+    
+        gc.collect()  # free memory
+    
+    def run_actors_optimization(self, samples, ids):
+        if any([samples[col][0] for col in samples]):  # if samples contain values
+            param = samples.iloc[ids]
+            self.parameters = {'utility_portfolio_min': param['utility_portfolio'], 'owner_portfolio_min': param['owner_portfolio']}
+        scenario, SP_scenario, SP_scenario_init = self.select_SP_obj_decomposition(self.scenario)
+        try:
+            scn = self.scenario["name"]
+            self.MP_iteration(scenario, Scn_ID=scn, binary=True, Pareto_ID=0)
+            self.add_df_Results(None, scn, 0, self.scenario)
+            return self.results[scn][0], self.results_MP[scn][0]
+        except:
+            return None, None
+    
+    def read_configurations(self, path=None):
+        if path is None:
[…]

[…]
