### 2024-05-22 IPESE/REHO#20 @DorsanL
> I recommend modifying your requirements.txt with the following

Done, and documentation updated accordingly (https://reho.readthedocs.io/en/main/sections/5_Getting_started.html#requirements)

> Your setup.py is missing psycopg2

Indeed, it used to be installed through the former `install_dependencies()` function. It's now part of install_requires() as suggested.

I can close the issue, please let me know if some issue persists :)
