### 2024-05-27 IPESE/REHO#21 @nmstreethran
I have the following additional suggestions for your documentation.

1.  Consider adding a Jupyter notebook for the scripts in the `template` directory and including that in your Sphinx documentation. I managed to run those scripts in a notebook which you can view here: https://nbviewer.org/gist/nmstreethran/2475549ab51b02c70eb48ad438781ea1/template.ipynb. I think this could be a nice way to display all the expected results and the interactive plots within the documentation.

1.  Consider including titles and additional context (e.g. explanation of acronyms) in the interactive plots. Currently, only the time series plot in the template scripts has a title.

1.  There are some boxes / annotations (for example, [here](https://reho.readthedocs.io/en/main/sections/3_Model.html#annual-operating-expenses) and [here ("TO DO")](https://reho.readthedocs.io/en/main/sections/4_Package_structure.html#reho.plotting.plotting.plot_pareto_by_objectives)) which should probably be commented out.

1.  Consider adding a `CONTRIBUTING.md` file in your GitHub repository and copying the contents of [contributing guidelines](https://reho.readthedocs.io/en/main/sections/7_Contribute.html).

I have created a separate issue for the warnings that appeared when running the scripts at #24.
