+++ scripts/solene/results_analysis.py
@@ -8,17 +8,22 @@ import datetime
 
 if __name__ == '__main__':
 
-    run_label = "10buil_14_1640"
-    districts = [277,3658,3112]
+    # SET-UP =====================================================================================================
+    # Results files 
+    # run_label = "10buil_14_1640"
+    # districts = [277,3658,3112]
+    run_label = "1district"
+    districts = ["noconstraints","maxshare","maxshare10km","relaxed"]
+    districts = ['28_1106']
     pickle_files = [f'results/{run_label}_{d}.pickle' for d in districts] # filename format example : results/10buil_14_1640_277.pickle
     
-    if len(pickle_files) == 1:
-        with open(pickle_files[0], 'rb') as handle:
-            results = pickle.load(handle)
-    else:
-        for file,label in zip(pickle_files,districts):
-            with open(file, 'rb') as handle:
-                vars()[f"results{label}"] = pickle.load(handle)
+    # Specifications for the graphs
+    Pareto_IDs = [0] # which iteration(s) to generate for graphs that only plot 1 iter (especially for the pkm graphs)
+
+    # PLOTTING =====================================================================================================
+    for file,label in zip(pickle_files,districts):
+        with open(file, 'rb') as handle:
+            vars()[f"results{label}"] = pickle.load(handle)
 
     # from plotting library
     rehos_dict = dict()
@@ -29,9 +34,45 @@ if __name__ == '__main__':
     fig.write_html('plots\plot.html')
     fig.show()
 
-    df_pkm, fig = plot_pkm(results3112.results['totex'][4])
-    fig.write_html('plots\plot2.html')
-    fig.show()
+    ## pkm profiles per typical day p.
+    for p in Pareto_IDs:
+        for label in districts:
+            df_pkm, fig = plot_pkm(vars()[f"results{label}"].results['totex'][p],run_label=f"{run_label} - district {label} - iter {p}")
+            fig.write_html(f'plots\pkm{label}{p}.html')
+            fig.show()
+
+            df_share = pd.DataFrame()
+            df_share['PT']= df_pkm.loc[:,df_pkm.columns.get_level_values(level=0).str.startswith('PT')].sum(axis=1)
+            df_share['Bike']= df_pkm.loc[:,df_pkm.columns.get_level_values(level=1).str.startswith('Bike')].sum(axis=1)
+            df_share['Total'] = df_pkm.loc[:,df_pkm.columns.get_level_values(level=0) == "Domestic_energy"]
+
+            df_share = df_share.groupby(['Layer',"Period"]).agg('sum')
+            df_share = df_share.div(df_share['Total'],axis = 0)
+
+            if "T" and "I" and "W" in vars()[f"results{label}"].cluster['Attributes']:
+                attributes = "T_I_W"
+                File_ID = "timestamp_{location}_{periods}_{periodduration}_{attributes}".format(
+                    location = vars()[f"results{label}"].cluster['Location'],
+                    periods = vars()[f"results{label}"].cluster['Periods'],
+                    periodduration = vars()[f"results{label}"].cluster['PeriodDuration'],
+                    attributes = attributes
+                )
+            else:
+                print("Achtung : check how the timestamp file is label, and especially Attribute TIW? ")
+                File_ID = "timestamp_{location}_{periods}_{periodduration}".format(
+                    location = vars()[f"results{label}"].cluster['Location'],
+                    periods = vars()[f"results{label}"].cluster['Periods'],
+                    periodduration = vars()[f"results{label}"].cluster['PeriodDuration']
+                )
+                print(f"File_ID tested : {File_ID}")
+
[…]
