### 2024-05-28 30db715
debugging example
