### 2024-05-31 175d078
added electric bikes + adapted some parameters
