### 2024-06-03 775ba42
Lucerne scenario v1
