+++ scripts/solene/iterative_scenario_lucerne.py
@@ -6,10 +6,31 @@ import datetime
 
 
 if __name__ == '__main__':
+    date = datetime.datetime.now().strftime("%d_%H%M")
 
-    districts = [3658,3112,277]
+    # Parameters ==================================================================================================================
+    districts = [ 7724,8538,13569,13219,13228]
+    district_parameters = {
+        7724  : {"PopHouse" : 4.52, "rho" : pd.Series([95,3,2],index=['household','industry','service']) } ,
+        8538  : {"PopHouse" : 10.5, "rho" : pd.Series([89,2,6],index=['household','industry','service']) } ,
+        13569 : {"PopHouse" : 7.71, "rho" : pd.Series([96,2,1],index=['household','industry','service']) },
+        13219 : {"PopHouse" : 9.59, "rho" : pd.Series([61,17,17],index=['household','industry','service']) },
+        13228 : {"PopHouse" : 11.29, "rho" : pd.Series([52,21,5],index=['household','industry','service']) }
 
-    # Initialization of scenarios - Generic parameters
+    }
+    
+    # to change easily the size of population depending of the nb of buildings
+    n_buildings = 2
+    for k in district_parameters.keys():
+        district_parameters[k]['Population'] = n_buildings * district_parameters[k]['PopHouse'] 
+
+    # if you want to add non-zero params  ["outside_charging_price","charging_externalload"] for the init scenario
+    PARAM_INIT = False
+    with open('data/mobility/parameters_iteration.pickle', 'rb') as handle:
+        parameters_init = pickle.load(handle)
+
+
+    # Initialization of scenarios - Generic parameters ==========================================================================================
     ## Set building parameters
     reader = QBuildingsReader()
     reader.establish_connection('Suisse')
@@ -40,24 +61,25 @@ if __name__ == '__main__':
     # Initialization of scenarios - District parameters
     for transformer in districts:
         ## Set building parameters
-        qbuildings_data = reader.read_db(transformer=transformer, nb_buildings=2)
+        qbuildings_data = reader.read_db(transformer=transformer, nb_buildings=n_buildings)
         
         ## District parameters
         ext_districts = [d for d in districts if d != transformer]
-        parameters = {'Population': 9,
-                      "Districts" : ext_districts,
-                      "share_district_activity": rho_param(ext_districts,1) } # other districts 
+        df_rho = pd.DataFrame()
+        for k in district_parameters.keys():
+            df_rho[k] = district_parameters[k]['rho']
+        df_rho = df_rho.T.rename(columns = {'industry' : 'work', 'service': 'leisure'})
+        parameters = {'Population': district_parameters[transformer]['Population'],
+                      "Districts" : ext_districts}
+                    #   "share_district_activity": rho_param(ext_districts,df_rho) } # other districts 
 
         vars()['reho_' + str(transformer)] = reho(qbuildings_data=qbuildings_data, units=units, grids=grids, cluster=cluster, scenario=scenario,
                     method=method, parameters=parameters, solver="gurobiasl")
 
-    # if you want to add non-zero params  ["outside_charging_price","charging_externalload"] for the init scenario
-    PARAM_INIT = False
-    with open('data/mobility/parameters_iteration.pickle', 'rb') as handle:
-        parameters_init = pickle.load(handle)
+    
 
 
-    # Run optimization
+    # Run optimization =====================================================================================================================
     df_pi = pd.DataFrame()
     df_externalcharging = pd.DataFrame()
     variables = dict()
@@ -65,8 +87,49 @@ if __name__ == '__main__':
     deltas = list()
     df_delta = pd.DataFrame()
 
+    # Standalone initializing run
[…]
