+++ scripts/solene/results_analysis.py
@@ -12,10 +12,14 @@ if __name__ == '__main__':
     # Results files 
     # run_label = "10buil_14_1640"
     # districts = [277,3658,3112]
-    run_label = "EBIKE"
-    run_label = "EVactivity"
-    districts = ["noconstraints","maxshare","maxshare10km","relaxed"]
-    districts = ["calibrage","calibrated"]
+    # run_label = "EBIKE"
+    # run_label = "EVactivity"
+    # districts = ["noconstraints","maxshare","maxshare10km","relaxed"]
+    # districts = ["calibrage","calibrated"]
+
+    run_label = "lucerne_31_1600"
+    districts = [ 7724,8538,13569,13219,13228]
+
     pickle_files = [f'results/{run_label}_{d}.pickle' for d in districts] # filename format example : results/10buil_14_1640_277.pickle
     
     # Specifications for the graphs
