### 2024-06-03 0604243
script lucerne - commit bf debugging negative OPEX
