+++ scripts/solene/results_analysis.py
@@ -17,7 +17,7 @@ if __name__ == '__main__':
     # districts = ["noconstraints","maxshare","maxshare10km","relaxed"]
     # districts = ["calibrage","calibrated"]
 
-    run_label = "lucerne_31_1600"
+    run_label = "lucerne_03_0857"
     districts = [ 7724,8538,13569,13219,13228]
 
     pickle_files = [f'results/{run_label}_{d}.pickle' for d in districts] # filename format example : results/10buil_14_1640_277.pickle
