### 2024-06-03 6ec978c
BUG OPEX negative (POP = 0)
