### 2024-06-07 62aa25c
script lucerne : remis les coûts de vente charging_ext_load (+ fonctions d'affichage)
