+++ scripts/solene/results_analysis.py
@@ -5,6 +5,88 @@ from reho.plotting.plotting import *
 
 import datetime
 
+def plot_cost_EV(reho_dict,run_label = "",CS_unit = ['EVcharging_district'],Scn_ID = 'totex'):
+    """
+    reho_dict : dict
+        dictionnary of reho objects. The keys must be the names of the districts (transformer number)
+    results : reho.results[Scn_ID] 
+        ex: reho.results['totex']
+    run_label : str
+        used for the title of the graph
+    CS_unit : list
+        temporary, to get all the types of charging station, ideally this function should fetch in district_units.csv all the UnitsofType[charging station]
+
+    """
+    try: # if reho
+        Pareto_ID_number = len(rehos_dict[list(rehos_dict.keys())[0]].results[Scn_ID].keys())
+        REHO = True
+    except: # if reho.results => this can't get the correct results bc we're missing the dual variables. 
+        Pareto_ID_number = len(rehos_dict[list(rehos_dict.keys())[0]][Scn_ID].keys())
+        REHO = False
+        assert("Please provide a reho object and not reho.results for this function")
+
+
+    # DATA FETCHING    
+    # pi
+    df_pi = pd.DataFrame()
+    for Pareto_ID in range(Pareto_ID_number):
+        if REHO:
+            for name, reho in rehos_dict.items():
+                pi = reho.results_MP[Scn_ID][Pareto_ID][0]["df_Dual_t"]["pi"].xs("Electricity")
+                df_pi[name] = pi
+    new_columns = pd.MultiIndex.from_product([['pi'], df_pi.columns], names=['activity', 'load'])
+    df_pi.columns = new_columns
+
+    # gathering electric load data
+    df_costs = pd.DataFrame()
+    for name, reho in rehos_dict.items():
+        for Pareto_ID in range(Pareto_ID_number):
+            df = reho.results[Scn_ID][Pareto_ID]["df_Unit_t"].copy()
+            df = df.loc[:,df.columns.str.contains('EV_E_charged_outside')].xs('Electricity').dropna()
+            df['demand'] = name
+            df['Pareto_ID'] = Pareto_ID
+            df.set_index("demand",append=True,inplace=True)
+            df.set_index("Pareto_ID",append=True,inplace=True)
+            new_columns = [x.split('[')[1].split(']')[0].split(',') for x in df.columns ]
+            new_columns = [(x,int(y)) for (x,y) in new_columns]
+            df.columns = pd.MultiIndex.from_tuples(new_columns,names=('activity','load'))
+            df = df.merge(df_pi,left_index = True,right_index = True)
+            df = df.reorder_levels(['load','activity'],axis=1)
+            for l in df.columns.get_level_values('load').unique():
+                results = df.loc[:,l].mul(df.loc[:,l]['pi'],axis = 0)
+                results.columns = pd.MultiIndex.from_product([[l],results.columns])
+                df.loc[:,l] = results
+            df = df.loc[:, df.columns.get_level_values('activity') != 'pi']
+            df_costs = pd.concat([df_costs,df])   
+
+
+    df_home = pd.DataFrame()
+    for name, reho in rehos_dict.items():
+        for Pareto_ID in range(Pareto_ID_number):
+
+            df_costNetwork = reho.results[Scn_ID][Pareto_ID]['df_Grid_t'].xs(('Electricity','Network'))[['Cost_supply']]
+            df_CSdemand = reho.results[Scn_ID][Pareto_ID]['df_Unit_t'].xs('Electricity')
+            df_CSdemand = df_CSdemand[df_CSdemand.index.get_level_values('Unit').isin(CS_unit)][['Units_demand']]
[…]
