### 2024-06-10 70c2aea
update methods storing results to compute KPIs + generalize plot_profiles
