+++ reho/plotting/plotting.py
@@ -146,7 +146,7 @@ def remove_building_from_index(df):
 
 
 def plot_performance(results, plot='costs', indexed_on='Scn_ID', label='EN_long', add_annotation=True, per_m2=False, additional_costs={}, additional_gwp={}, scc=0.177,
-                     filename=None, export_format='html', scaling_factor=1, return_df=False):
+                     filename=None, export_format='html', scaling_factor=1, return_df=False, unit_costs=' CHF'):
     """
     Plot performance based on REHO results.
 
@@ -183,6 +183,8 @@ def plot_performance(results, plot='costs', indexed_on='Scn_ID', label='EN_long'
         Scales linearly the REHO results for the plot.
     return_df: bool
         A dataframe can be returned for further post-processing or reporting purposes.
+    unit_costs: string
+        To specify the cost unit (' CHF', ' kCHF', ' MCHF', ' GCHF')
 
     Returns
     ----------
@@ -198,12 +200,14 @@ def plot_performance(results, plot='costs', indexed_on='Scn_ID', label='EN_long'
 
     change_data = pd.DataFrame()
     change_data.index = ['x_axis_1', 'x_axis_2', 'y_axis', 'keyword', 'total', 'unites', 'scc_legend']
-    decimal = 0
+    decimal = 1
     lang = re.split('_', label)[0]
+    unit_prefix = {' CHF': 1, ' kCHF': 1e3, ' MCHF': 1e6, ' GCHF': 1e9}
+    prefix = unit_prefix[unit_costs]
 
     if plot == 'costs':
-        change_data['FR'] = ['CAPEX', 'OPEX', 'Coûts [CHF/an]', 'Coûts', 'TOTEX', ' CHF', '']
-        change_data['EN'] = ['CAPEX', 'OPEX', 'Costs [CHF/y]', 'Costs', 'TOTEX', ' CHF', '']
+        change_data['FR'] = ['CAPEX', 'OPEX', 'Coûts [CHF/an]', 'Coûts', 'TOTEX', unit_costs, '']
+        change_data['EN'] = ['CAPEX', 'OPEX', 'Costs [CHF/y]', 'Costs', 'TOTEX', unit_costs, '']
         df_costs = df_Economics.xs('costs', level='Perf_type')
         if per_m2:
             df_costs = df_costs / era.sum()
@@ -280,9 +284,9 @@ def plot_performance(results, plot='costs', indexed_on='Scn_ID', label='EN_long'
     xtick = [x + 1 / 6 for x in x1]
 
     text_capacities = ["<b>" + change_data.loc['x_axis_1', lang] + "</b><br>" + str(custom_round(cp, decimal))
-                   for cp in combined_capacities]
+                   for cp in combined_capacities / prefix]
     text_resources = ["<b>" + change_data.loc['x_axis_2', lang] + "</b><br>" + str(custom_round(op, decimal))
-               for op in combined_resources]
+               for op in combined_resources / prefix]
     pos_resources = data_resources[indexes][data_resources[indexes] > 0].sum(axis=0).astype(int).reset_index(drop=True) + \
                data_scc_resources[indexes][data_scc_resources[indexes] > 0].sum(axis=0).astype(int).reset_index(drop=True)
 
@@ -305,7 +309,7 @@ def plot_performance(results, plot='costs', indexed_on='Scn_ID', label='EN_long'
                                )
             fig.add_annotation(x=xtick[i], y=max(combined_capacities[i], pos_resources[i],
                                                  combined_capacities[i] + combined_resources[i]) + text_placeholder,
-                               text="<b>Total</b><br>" + str(custom_round(combined_capacities[i] + combined_resources[i], decimal)) + change_data.loc['unites', lang],
[…]
