+++ reho/model/postprocessing/write_results.py
@@ -363,8 +363,10 @@ def get_df_Results_from_SP(ampl, scenario, method, buildings_data, filter=True):
         df_Results["df_Buildings_t"] = set_df_buildings_t(ampl)
         df_Results["df_Unit_t"] = df_Unit_t
     else:
-        for i in ['Cost_demand', 'Cost_supply', 'GWP_demand', 'GWP_supply', "Uncontrollable_load"]:
-            del df_Results["df_Grid_t"][i]
+        df_Results["df_Buildings_t"] = set_df_buildings_t(ampl)[["House_Q_heating", "T_in", "House_Q_DHW"]]
+        bat = df_Unit_t[df_Unit_t.index.get_level_values("Unit").str.contains("Battery")][['Units_demand', 'Units_supply']]
+        pv = df_Unit_t[df_Unit_t.index.get_level_values("Unit").str.contains("PV")][['Units_supply', 'Units_curtailment']]
+        df_Results["df_Unit_t"] = pd.concat([pv, bat])
 
     if method["save_streams"]:
         df_Results["df_Streams_t"] = set_df_streams_t(ampl)
