+++ reho/plotting/plotting.py
@@ -1544,7 +1544,7 @@ def plot_load_duration_curve(results, ids, save_fig=False, label='EN_long', colo
     ax.set_xlabel('time [hours]', fontsize=19)
     ax.legend(loc="lower left", fontsize=17)
     ax.grid()
-    ax.set_title('Load duration curve of each district',fontsize=25)
+    ax.set_title('Load duration curve of each district',fontsize=20)
     # ax.hlines(570, -20, 12000, linewidth=1.0, linestyle="--", color=layout.loc['Electrical_grid', color])
     # ax.hlines(-570, -20, 12000, linewidth=1.0, linestyle="--", color=layout.loc['Electrical_grid', color])
     # ax.text(5400, 600, 'transformer capacity', color=layout.loc['Electrical_grid', color], fontsize=18)
@@ -1780,7 +1780,7 @@ def unit_monthly_plot(results, to_plot, label='EN_short', save_path="", filename
 
     return fig
 
-def plot_pathway(results,scn_id=None, EMOO_list=[], y_span=[],Battery_yearly=[],EV_yearly=[],y_span_yearly=[],label='EN_short', save_path="", filename=None, export_format='html', objective=None, constraint=None):
+def plot_pathway(results,scn_id=None, EMOO_list=[], y_span=[],Battery_yearly=[],EV_yearly=[],y_span_yearly=[],label='EN_short', save_path="", filename=None, export_format='html', objective=None, constraint=None, district=None, scn_number=None,era=None):
     """
         Generate a line plot showing the investment pathway for the respective Objective function and EMOO selected
 
@@ -1824,7 +1824,12 @@ def plot_pathway(results,scn_id=None, EMOO_list=[], y_span=[],Battery_yearly=[],
     fig, ax1=plt.subplots(figsize=(12,7))
 
     # Set title
-    ax1.set_title("Investment Pathway", fontsize=35)
+    title_str="Investment Pathway"
+    if district is not None:
+        title_str=title_str+" in district "+str(district)
+    if scn_number is not None:
+        title_str=title_str+" for scenario " + str(scn_number)
+    ax1.set_title(title_str, fontsize=25)
 
     # Set x-axis parameters
     ax1.tick_params(axis='x', labelsize=15)
@@ -1833,14 +1838,16 @@ def plot_pathway(results,scn_id=None, EMOO_list=[], y_span=[],Battery_yearly=[],
     obj=[]
     obj2=[]
     add_line=True
-    era = results[scn_id][0]['df_Buildings']['ERA'].sum()
+    if era is None:
+        era = results[scn_id][0]['df_Buildings']['ERA'].sum()
+
     if objective=='CAPEX':
         obj_label = 'Cost [CHF/m$^2$/y]'
         for i in range(len(EMOO_list)):
             obj=obj+[results[scn_id][i]['df_KPIs']['capex_m2']['Network']]
             obj2 = obj2 + [results[scn_id][i]['df_KPIs']['opex_m2']['Network']]
         obj_old = objective
-        objective = 'Cumulative CAPEX'
+        objective = 'CAPEX' #'Cumulative CAPEX'
         lns4 = ax1.plot(y_span, obj2, label='OPEX', color='C0', linewidth=2, linestyle='--')
     elif objective=='TOTEX':
         obj_label = 'Cost [CHF/m$^2$/y]'
@@ -1848,10 +1855,10 @@ def plot_pathway(results,scn_id=None, EMOO_list=[], y_span=[],Battery_yearly=[],
             obj = obj + [results[scn_id][i]['df_KPIs']['capex_m2']['Network']]
             obj2= obj2 + [results[scn_id][i]['df_KPIs']['opex_m2']['Network']]
         obj_old = objective
-        objective='Cumulative CAPEX'
[…]
