### 2024-06-14 b803724
two minor bugs debugged
