+++ reho/model/preprocessing/weather.py
@@ -7,6 +7,7 @@ import datetime as dt
 from reho.paths import *
 from reho.model.preprocessing.clustering import Clustering
 import pvlib
+import geopandas as gpd
 from pyproj import Transformer
 
 __doc__ = """
@@ -50,7 +51,8 @@ def get_cluster_file_ID(cluster):
     else:
         E = ''
 
-    File_ID = cluster['Location'] + '_' + str(cluster['Periods']) + '_' + str(cluster['PeriodDuration']) + T + I + W + E
+    File_ID = cluster['Location'] + '_' + str(cluster['Periods']) + '_' + str(cluster['PeriodDuration']) + \
+              T + I + W + E
 
     return File_ID
 
@@ -76,6 +78,8 @@ def generate_weather_data(cluster, qbuildings_data):
     if 'E' in cluster['Attributes']:
         attributes.append('Emissions')
 
+    df3 = pd.read_csv(os.path.join(path_to_weather, 'Elec_CO2_2023.txt'), index_col=[0], header=None)
+    df['Emissions'] = df3
     df = df[attributes]
     cl = Clustering(data=df, nb_clusters=[cluster['Periods']], option={"year-to-day": True, "extreme": []}, pd=cluster['PeriodDuration'])
     cl.run_clustering()
@@ -97,7 +101,7 @@ def get_weather_data(qbuildings_data):
     weather_data['Month'] = weather_data.index.month
     weather_data['Day'] = weather_data.index.day
     weather_data['Hour'] = weather_data.index.hour + 1
-    weather_data['id'] = (weather_data.reset_index().index + 1).to_list()
+    weather_data['id'] = (weather_data.reset_index().index+1).to_list()
     weekday = pd.read_csv(os.path.join(path_to_weather, 'weekday.txt'), index_col=[0], header=None)
     weather_data['Weekday'] = weekday[1].tolist()
 
@@ -109,7 +113,7 @@ def get_weather_data(qbuildings_data):
 def read_custom_weather(path_to_weather_file):
     """
     From the current directory, looks for a custom weather file.
-    This file should be a .csv with the same structure as the examples provided in ``reho/scripts/examples/data/profiles/``.
+    This file should be a .csv with the same structure as the template provided in ``reho/scripts/template/data/profiles/``.
     """
 
     df = file_reader(path_handler(path_to_weather_file))
@@ -208,7 +212,11 @@ def generate_output_data(cl, attributes, location):
 
     return print(f'The data have been computed and saved in {path_to_clustering}.')
 
-
+def get_metric(cluster):
+    if 'E' in cluster['Attributes']:
+        return 'GWP100a'
+    else:
+        return 'method 1'
 def write_dat_files(attributes, location, values_cluster, index_inter):
     """
     Writes the clustering results computed from ``generate_output_data`` as .dat files.
@@ -271,9 +279,49 @@ def write_dat_files(attributes, location, values_cluster, index_inter):
     # -------------------------------------------------------------------------------------
     # T
     # -------------------------------------------------------------------------------------
+
     df_T = values_cluster['Text']
     filename = os.path.join(path_to_clustering, 'T_' + File_ID + '.dat')
     df_T.to_csv(filename, index=False, header=False)
+
[…]
