+++ reho/plotting/plotting.py
@@ -1,6 +1,6 @@
 import calendar
 import locale
-import re
+import os
 
 import matplotlib.patches as mpatches
 import plotly.graph_objects as go
@@ -11,6 +11,7 @@ from plotly.subplots import make_subplots
 
 from reho.model.sub_problem import *
 from reho.plotting import sankey
+import plotly.graph_objects as go
 
 __doc__ = """
 Contains ready-to-use representations for results generated by REHO.
@@ -1735,6 +1736,90 @@ def plot_profiles(df, units_to_plot, style='plotly', label='EN_long', color='Col
             return fig, pd.DataFrame()
         else:
             return fig
+
+def yearly_demand_plot(file_path):
+
+    df_index = pd.read_excel(file_path, sheet_name='df_Index')
+    df_Buildings_t = pd.read_excel(file_path, sheet_name='df_Buildings_t')
+
+    df_Buildings = pd.read_excel(file_path, sheet_name='df_Buildings')
+    df_External = pd.read_excel(file_path, sheet_name='df_External')
+
+    # Fill missing values and filter DataFrame
+    df_Buildings_t['Hub'] = df_Buildings_t['Hub'].fillna(method='ffill')
+    df_Buildings_t['Period'] = df_Buildings_t['Period'].fillna(method='ffill')
+    df_External['Period'] = df_External['Period'].fillna(method='ffill')
+
+    df_output = df_Buildings_t[df_Buildings_t['Period'] < 11]
+    df_output_2 = df_External[df_External['Period'] < 11]
+
+    # Group by columns and sum
+    df_output = df_output.groupby(['Hub', 'Period', 'Time']).sum()
+    df_output_2 = df_External.groupby(['Period', 'Time']).sum()
+
+    # Build profiles
+    KeyWord = 'House_Q_heating'
+    KeyWord2 = 'House_Q_DHW'
+    Hub_List = set([k[0] for k in df_output[KeyWord].keys()])
+
+    HourList = range(8760)
+    HourList2 = list(range(167, 8760))
+
+    Heat_DemandProfilesByHub = {}
+    Heat_DemandProfilesByHub_plot = {}
+    DHW_DemandProfilesByHub = {}
+    for hub in Hub_List:
+        Heat_DemandProfile = []
+        DHW_DemandProfile = []
+        Heat_DemandProfilesByHub[hub] = Heat_DemandProfile
+        Heat_DemandProfilesByHub_plot[hub] = Heat_DemandProfile
+        DHW_DemandProfilesByHub[hub] = DHW_DemandProfile
+        for hour in HourList:
+            typical_day = df_index['PeriodOfYear'][hour]
+            current_hour = (hour) % 24 + 1
+            current_key = (hub, typical_day, current_hour)
+            Heat_DemandProfile.append(df_output[KeyWord][current_key])
+            DHW_DemandProfile.append(df_output[KeyWord2][current_key])
+
+        Heat_DemandProfilesByHub_plot[hub] = np.convolve(Heat_DemandProfilesByHub_plot[hub], np.ones(168),
+                                                         'valid') / 168
+
+    # turn profiles into a dataframe
+    df_Heat_DemandProfilesByHub = pd.DataFrame(Heat_DemandProfilesByHub, index=HourList)  # in kWh
+    df_DHW_DemandProfilesByHub = pd.DataFrame(DHW_DemandProfilesByHub, index=HourList)  # in kWh
+
+    fig1 = go.Figure()
+    for k in Hub_List:
+        fig1.add_trace(go.Scatter(x=HourList2, y=Heat_DemandProfilesByHub_plot[k], mode='lines', name=k))
+
+    fig1.update_layout(
+        title='Heat Demand Profiles by Building',
[…]
