+++ reho/model/preprocessing/clustering.py
@@ -19,7 +19,7 @@ class ClusterClass:
 
     """
 
-    def __init__(self, data, nb_clusters=None, option=None, pd=None, outliers=None):
+    def __init__(self, data, nb_clusters=None,  option=None, pd=None, outliers=None):
         super().__init__()
         if nb_clusters is None:
             self.nb_clusters = [8]
@@ -32,26 +32,27 @@ class ClusterClass:
 
         self.data_org = data  # input data- df: attribute vectors as columns
         self.data_nor = None
-        self.attr_org = []  # matrix or attributes, arrays:  len(attr_org) = total number of periods per year, len(attr_org[0])= number of attributes x period duration
+        self.attr_org = []  # matrix or attributes, arrays:  len(attr_org) = total number of periods per year , len(attr_org[0])= number of attributes x period duration
         self.attr_nor = []
         self.attr_clu = None
         self.modula = None
-        self.mod_org = []  # modulo data, f.e. last day of year (24hrs) for  pd = 7*24
-        self.mod_nor = []
+        self.mod_org =[]  # modulo data, f.e. last day of year (24hrs) for  pd = 7*24
+        self.mod_nor =[]
         if option is None:
             self.option = {"year-to-day": True, "extreme": [(1, "min"), (2, "max")]}
         else:
-            self.option = option
+            self.option= option
 
         self.results = {"idx": None}  # pd.DataFrame()
 
         self.kpis_clu = None
         self.nbr_opt = None
 
-        if pd is None:  # period duration, f.e. 1 period = 1 day => pd =  24 (default)
+        if pd is None:              # period duration, f.e. 1 period = 1 day => pd =  24 (default)
             self.pd = 24
         else:
             self.pd = pd
+
 
     def run_clustering(self):
         self.__do_normalization()
@@ -59,21 +60,21 @@ class ClusterClass:
         self.__compute_kpis()
         self.__select_optimal()
 
+
     def __do_normalization(self):
         # - Load data set
         self.data_nor = pd.DataFrame([], index=self.data_org.index, columns=self.data_org.columns)
         # - Create dissimilarity matrix
         for id, col in enumerate(self.data_org):
-            self.data_nor.loc[:, col] = ((self.data_org.loc[:, col] - self.data_org.loc[:, col].min()) / (
-                        self.data_org.loc[:, col].max() - self.data_org.loc[:, col].min()))  # normalize
+            self.data_nor.loc[:, col] = ((self.data_org.loc[:, col] - self.data_org.loc[:, col].min()) / (self.data_org.loc[:, col].max() - self.data_org.loc[:, col].min())) #normalize
             # - OPTION : re-arrange dataframe
             if self.option["year-to-day"]:
-                nbr_p = int(self.data_nor.shape[0] // self.pd)  # ! integer division / number of periods default: 365
+                nbr_p = int(self.data_nor.shape[0] // self.pd) #! integer division / number of periods default: 365
                 self.modulo = int(self.data_nor.shape[0] % self.pd)
-                stop = self.data_nor.shape[0] - self.modulo
-                self.attr_org.append(np.reshape(self.data_org.loc[: (stop - 1), col].tolist(), (nbr_p, self.pd)))
-                self.attr_nor.append(np.reshape(self.data_nor.loc[: (stop - 1), col].tolist(), (nbr_p, self.pd)))
-                self.mod_org.append(self.data_org.loc[stop:, col].tolist())
+                stop = self.data_nor.shape[0] -self.modulo
+                self.attr_org.append(np.reshape(self.data_org.loc[: (stop -1), col].tolist(), (nbr_p, self.pd)))
[…]
