### 2024-06-19 a5baebc
cleaning ampl files
