### 2024-06-20 6ab0f93
version of the model bf LU meeting
