### 2024-06-24 59ffcc7
updated constraints, max and min shares
