### 2024-06-24 #6734 @nmstreethran
Hi @mbarzegary, there are two main issues from my side that are still to be addressed:

- https://github.com/IPESE/REHO/issues/22 - tests are missing and need to be implemented
- https://github.com/IPESE/REHO/issues/24 - these errors and warnings should be fixed

I'm happy to complete my checklist once these are resolved.
