### 2024-06-25 5ea9a79
modified constraint on demand pkm
