### 2024-06-26 7950aa0
Suurstoffi script + changes so that the model can run without mobility + maxminshare = [0,1]
