+++ scripts/solene/Suurstoffi_script.py
@@ -1,0 +1,110 @@
+from reho.model.reho import *
+from reho.plotting.plotting import *
+
+
+if __name__ == '__main__':
+    reader = QBuildingsReader()
+    qbuildings_data = reader.read_csv(buildings_filename='data/clustering/buildings_suurstoffi.csv', nb_buildings=100)
+
+    cluster = {'Location': 'Lucern', 'Attributes': ['T', 'I', 'W'], 'Periods': 10, 'PeriodDuration': 24}
+    method = {'building-scale': True}
+    parameters = {}
+
+    scenario = dict()
+    scenario['Objective'] = 'TOTEX'
+    scenario['EMOO'] = {}
+    scenario['specific'] = []
+
+    # NO MOBILITY ======================================================================================================
+
+    # SCENARIO 1 Gas boiler
+    scenario['name'] = 'Gas'
+    scenario['exclude_units'] = ['HeatPump', 'PV', "Battery_district", 'EVcharging_district' , 'NG_Cogeneration']
+    scenario['enforce_units'] = ['NG_Boiler']
+
+    grids = infrastructure.initialize_grids({'Electricity': {},
+                                             'NaturalGas': {},
+                                             })
+    units = infrastructure.initialize_units(scenario, grids,district_data=True)
+
+    reho = reho(qbuildings_data=qbuildings_data, units=units, grids=grids, parameters=parameters, cluster=cluster,
+                scenario=scenario, method=method, solver="gurobiasl")
+
+    reho.single_optimization()
+
+    # SCENARIO 2 HP + PV
+    scenario['name'] = 'HP'
+    scenario['exclude_units'] = ['Battery_district', 'EVcharging_district', 'NG_Cogeneration']
+    scenario['enforce_units'] = []
+    scenario['specific'] = ["enforce_PV_max"]
+    units = infrastructure.initialize_units(scenario, grids)
+
+    reho.scenario = scenario
+    reho.units = units
+    reho.infrastructure = infrastructure.infrastructure(qbuildings_data, units, grids)
+    reho.single_optimization()
+
+    # WITH MOBILITY ====================================================================================================
+    reho.parameters['Population'] = 1500
+
+    grids = infrastructure.initialize_grids({'Electricity': {},
+                                             'NaturalGas': {},
+                                             'FossilFuel': {},
+                                             'Mobility': {},
+                                             })
+
+    # SCENARIO 3 Full ICE
+    scenario['name'] = 'ICE'
+    scenario['specific'] = []
+    scenario['exclude_units'] = ['NG_Cogeneration']
+    scenario['enforce_units'] = ['EV_district']
+    units = infrastructure.initialize_units(scenario, grids, district_data=True)
+
+    reho.parameters['min_share_ICE'] = 0.9
+    reho.parameters['n_ICEperhab'] = 10
+    reho.parameters['max_share_EBikes'] = 0 # to prevent bug.
+
+    reho.scenario = scenario
+    reho.units = units
+    reho.grids = grids
+    reho.infrastructure = infrastructure.infrastructure(qbuildings_data, units, grids)
+    reho.single_optimization()
+
+    # SCENARIO 4 OFS SHARES
+    scenario['name'] = 'OFS_modalshares'
+
+    shares = {  "share_cars" : 0.66,
+                "share_PT"  : 0.24,
+                "share_MD" : 0.1, # mobilité douce : "soft mobility" ? (from FSO : include biking, walking, electric biking)
+                "share_ICE" : 0.66,
+                "share_PT_train" : 0.2,
+            }
[…]

[…]
