### 2024-06-28 IPESE/REHO#18 @hgandhi2411
@DorsanL Thanks for removing redundant files.

I reinstalled reho using the pip install instructions. I still had the reho repo cloned. Tried to run run.py but still wasnt able to. A few suggestions based on the continued inability to run `/scripts/template/run.py`

1.  When running `python /scripts/template/run.py` from REHO root dir, I get this error:
```
Traceback (most recent call last):
  File "/Users/hetagandhi/Documents/REHO/scripts/template/run.py", line 8, in <module>
    qbuildings_data = reader.read_csv(buildings_filename='data/buildings.csv', nb_buildings=1)
  File "/opt/anaconda3/envs/reho-review/lib/python3.10/site-packages/reho/model/preprocessing/QBuildings.py", line 137, in read_csv
    self.data['buildings'] = file_reader(buildings_filename)
  File "/opt/anaconda3/envs/reho-review/lib/python3.10/site-packages/reho/paths.py", line 74, in file_reader
    file = Path(path_handler(file))
  File "/opt/anaconda3/envs/reho-review/lib/python3.10/site-packages/reho/paths.py", line 69, in path_handler
    raise FileNotFoundError('The relative path that was given is not a valid file.')
FileNotFoundError: The relative path that was given is not a valid file.
```
Please use relative paths in your paths.py file - these should be relative to the file being run rather than relative to the working directory.

2. When I navigate to /scripts/template/ and run the file run.py file, the path issue is solved. However, there is a NameError:
```
(reho-review) hetagandhi@Abhisheks-Air REHO % cd scripts/template
(reho-review) hetagandhi@Abhisheks-Air template % python run.py
/opt/anaconda3/envs/reho-review/lib/python3.10/site-packages/reho/model/preprocessing/QBuildings.py:426: UserWarning: Geometry column does not contain geometry.
  new_buildings_data[REHO_index] = df_buildings[key]
Traceback (most recent call last):
  File "/Users/hetagandhi/Documents/REHO/scripts/template/run.py", line 30, in <module>
    reho = REHO(qbuildings_data=qbuildings_data, units=units, grids=grids, cluster=cluster, scenario=scenario, method=method, solver="highs")
NameError: name 'REHO' is not defined
```
I believe this is because you haven't configured imports correctly in your `__init__.py` file. My recommendation is to configure your init files correctly.

3. I tried starting a python shell to test reho install. Here's the output - confirming my suspicion of imports not configured properly in your __init__.py files

4. Please also set up some basic tests to test import issues, and fundamental functions so that review of this package can be accelerated. As stated in #22.
