### 2024-06-28 #6734 @hgandhi2411
@mbarzegary  @DorsanL has addressed some of my concerns, however, there are still some issues with imports and how the package is structured. I would like this issue addressed  https://github.com/IPESE/REHO/issues/18 along with some basic unit tests put in place as described by @nmstreethran
