### 2024-06-28 daf0293
working on plots
