+++ scripts/solene/Suurstoffi_plots.py
@@ -22,18 +22,16 @@ if __name__ == '__main__':
     report_folder = "C:/masterthesis/images/"
 
     fig = plot_performance(results, plot='costs', indexed_on='Scn_ID', label='EN_long' ) # , additional_costs=additional_costs)
-    fig.write_html(f"{report_folder}C1_performance_Suurstoffi_baseline.html")
-    fig.write_image(f"{report_folder}C1_performance_Suurstoffi_baseline.png")
+    fig.write_html(f"{report_folder}C1_performance_{run_label}.html")
+    fig.write_html(f"plots/C1_performance_{run_label}.html")
+    fig.write_image(f"{report_folder}C1_performance_{run_label}.png")
 
-    fig2, df = plot_sankey(results["M3_EV10"][0], return_df=True)
-    fig2.write_html(f"{report_folder}C1_sankeyM3_EV10_Suurstoffi_baseline.html")
-    fig2.write_image(f"{report_folder}C1_sankeyM3_EV10_Suurstoffi_baseline.png")
+    for scn_ID in results.keys():
+        fig2, df2 = plot_sankey(results[scn_ID][0], return_df=True)
+        fig2.write_html(f"plots/C1_sankey{scn_ID}_{run_label}.html")
 
-    fig4 = plot_sankey(results["M1_ICE"][0])
-    fig4.write_html(f"plots/Sankey_ICE{run_label}.html")
-
-    fig3 = plot_profiles(results["Modalshares_EV10"][0], units_to_plot=["EV_district","ICE_district"])
-    fig3.write_html(f"plots/Profile_EV10{run_label}.html")
+    fig3 = plot_profiles(results["M3_EV10"][0], units_to_plot=["EV_district","ICE_district"])
+    fig3.write_html(f"plots/Profile_M3_EV10_{run_label}.html")
 
     fig5 = plot_profiles(results["ICE"][0], units_to_plot=["EV_district","ICE_district"])
-    fig5.write_html(f"plots/Profile_ICE{run_label}.html")
+    fig5.write_html(f"plots/Profile_ICE_{run_label}.html")
