### 2024-07-03 40ee955
Suurstoffi script (for EVpenetration_lowprice)
