### 2024-07-11 #6734 @DorsanL
Dear @mbarzegary, thank you for your attention and for this reminder!

These past few weeks have been quite busy with other projects on my end, and I regret not having been able to dedicate time to these pointed issues. These latter seem quite clear to me, and I plan to resolve them over the course of the next week.

I thank the reviewers for their careful work, and I will get back to all of you within 7 days with (hopefully!) all the points addressed. 🙏
