### 2024-07-12 31d5adc
cleaning unused variables
