+++ documentation/sections/3_Model.rst
@@ -93,8 +93,6 @@ List of symbols
             +-------------------+------------------------------------------+--------------------------------------+
             | :math:`\alpha`    | azimuth angle                            | :math:`^{\circ}`                     |
             +-------------------+------------------------------------------+--------------------------------------+
-            | :math:`b`         | baremodule                               | :math:`-`                            |
-            +-------------------+------------------------------------------+--------------------------------------+
             | :math:`\beta`     | limiting angle                           | :math:`^{\circ}`                     |
             +-------------------+------------------------------------------+--------------------------------------+
             | :math:`c`         | energy tariff                            | :math:`\text{currency}/kWh`          |
