+++ documentation/sections/2_Releases.rst
@@ -98,7 +98,7 @@ The thesis also focuses in particular on the role of integrating photovoltaic pa
         The results highlight that a low-emission electrical grid mix has a high impact on sustainable design of renewable energy hubs at the building scale and led to less independent system configurations. Optimally integrating of solar systems had a significant impact on their interaction with the electrical grid: rotating the panels 20° westwards reduced the grid exchange peak by 50% while increasing cost by only 8.3%. Moreover, the studied district could achieve carbon neutrality based on PV energy alone, whereas self-sufficiency is more ambitious that confirmed the importance of storage systems: even with 100% round-trip efficiency of storage systems, the required ratio of area covered in PV modules to the ERA was 0.44 and 16% of available facades were needed to be covered with PV modules . However, energy demand reduction through thermal renovation would allow self-sufficiency with half of the PV and storage capacity.  Overall, this work demonstrates that moving from a decentralized to coordinated and centralized design strategy allows a higher electrification rate and an increased integration of renewable energy in the district for the same total expenses. The centralized investment strategy differed most from the decentralized strategy for PV panels; using the centralized strategy, a wide range of PV installation on less--optimal surfaces became economically interesting. The most economically convenient solution to overcome transformer limitations were district storage for peak shaving and photovoltaic curtailment. The cost increase were around 600 CHF per kWyr annual capacity shortage, regardless of the considered district energy system.
 
 
-Acamedic publications
+Academic publications
 ------------------------
 
 .. dropdown:: :bdg-primary:`Journal` Identification of typical district configurations: A two-step global sensitivity analysis framework, Chuat 2024 :cite:`chuatIdentificationTypicalDistrict2024`
