### 2024-07-22 c14cf74
cleaning plotting functions
