+++ reho/plotting/plotting.py
@@ -1,24 +1,16 @@
 import locale
 
-import matplotlib.patches as mpatches
 import plotly.graph_objects as go
 from matplotlib import pyplot as plt
-from matplotlib.legend_handler import HandlerTuple
-from matplotlib.lines import Line2D
 from plotly.subplots import make_subplots
 
 from reho.model.sub_problem import *
-from reho.plotting.utils import *
 from reho.plotting import sankey
+from reho.plotting.utils import *
 
 __doc__ = """
 Contains ready-to-use representations for results generated by REHO.
 """
-
-
-################################################################################################################################################################
-# Plotly
-################################################################################################################################################################
 
 
 def plot_performance(results, plot='costs', indexed_on='Scn_ID', label='EN_long', add_annotation=True, per_m2=False, additional_costs=None, additional_gwp=None,
@@ -73,9 +65,10 @@ def plot_performance(results, plot='costs', indexed_on='Scn_ID', label='EN_long'
         additional_gwp = {}
     if additional_costs is None:
         additional_costs = {}
+
     sc = list(results.keys())[0]
     id = list(results[sc].keys())[0]
-    era = results[sc][id]['df_Buildings'].ERA
+    era = results[sc][id]['df_Buildings'].ERA.sum()
 
     df_Economics = dict_to_df(results, 'df_Economics')
 
@@ -91,7 +84,7 @@ def plot_performance(results, plot='costs', indexed_on='Scn_ID', label='EN_long'
         change_data['EN'] = ['CAPEX', 'OPEX', 'Costs [CHF/y]', 'Costs', 'TOTEX', unit_costs, '']
         df_costs = df_Economics.xs('costs', level='Perf_type')
         if per_m2:
-            df_costs = df_costs / era.sum()
+            df_costs = df_costs / era
             change_data.loc['y_axis']['FR'] = "Coûts [CHF/m2/an]"
             change_data.loc['y_axis']['EN'] = "Costs [CHF/m2/y]"
             decimal = 1
@@ -109,7 +102,7 @@ def plot_performance(results, plot='costs', indexed_on='Scn_ID', label='EN_long'
         change_data['EN'] = ['Capacities', 'Resources', 'Emissions [kgCO2/y]', 'Emissions', 'Total', ' kgCO2', '']
         df_impact = df_Economics.xs('impact', level='Perf_type')
         if per_m2:
-            df_impact = df_impact / era.sum()
+            df_impact = df_impact / era
             change_data.loc['y_axis']['FR'] = "Émissions [kgCO2/m2/an]"
             change_data.loc['y_axis']['EN'] = "Emissions [kgCO2/m2/y]"
             decimal = 1
@@ -129,8 +122,8 @@ def plot_performance(results, plot='costs', indexed_on='Scn_ID', label='EN_long'
         df_costs = df_Economics.xs('costs', level='Perf_type')
         df_impact = df_Economics.xs('impact', level='Perf_type')
         if per_m2:
-            df_costs = df_costs / era.sum()
-            df_impact = df_impact / era.sum()
+            df_costs = df_costs / era
+            df_impact = df_impact / era
             change_data.loc['y_axis']['FR'] = "Coûts [CHF/m2/an]"
             change_data.loc['y_axis']['EN'] = "Costs [CHF/m2/y]"
             decimal = 1
@@ -343,7 +336,7 @@ def plot_expenses(results, plot='costs', indexed_on='Scn_ID', label='EN_long', p
 
     sc = list(results.keys())[0]
     id = list(results[sc].keys())[0]
-    era = results[sc][id]['df_Buildings'].ERA
[…]
