+++ reho/plotting/rainbow_plots.py
@@ -1,344 +1,0 @@
-import pandas as pd
-import matplotlib.pyplot as plt
-import numpy as np
-from scipy.interpolate import interp1d
-import math
-import scipy as sp
-
-__doc__ = """
-Reconstructs data from a Pareto optimization to produce a rainbow plot.
-"""
-
-
-def linerize_equal_steps(x, y):
-    x.name = 'x'
-    y.name = 'y'
-    step = (x.max() - x.min()) / len(x)
-    df = pd.concat([x, y], names=['x', 'y'], axis=1)
-
-    df_new = pd.DataFrame(columns=['y'])
-    for s in np.arange(1, len(x) + 1):
-        x_new = s * step
-        ub = df[df.x >= x_new].x.min()
-        lb = df[df.x < x_new].x.max()
-        if math.isnan(ub):  # happens in case x_new == to last x in array
-            y_new = df[df.x == lb].iloc[0].y  # just return last point
-        else:
-            upper_bound = df[df.x == ub].iloc[0]
-            lower_bound = df[df.x == lb].iloc[0]
-
-            m = ((upper_bound.y - lower_bound.y) / (upper_bound.x - lower_bound.x))
-            b = lower_bound.y - m * lower_bound.x
-
-            y_new = m * x_new + b
-        df_new.at[x_new, 'y'] = y_new
-    return df_new
-
-
-def split_orientation(az):
-    if (az < 45) or (az > 315):
-        orientation = 'North'
-    elif (az >= 45) and (az < 135):
-        orientation = 'East'
-    elif (az >= 135) and (az < 225):
-        orientation = 'South'
-    elif (az >= 225) and (az <= 315):
-        orientation = 'West'
-    else:
-        orientation = 'Other'
-
-    return orientation
-
-
-def get_roof_max(df_nbr, hs_A):
-    df_n = df_nbr.groupby(['Pareto_ID', 'Surface', 'Azimuth', 'Tilt']).sum()
-    roofs = df_n[df_n.index.get_level_values(level='Tilt') != 90]
-
-    # get total m2 PV
-    total_roofs = roofs.groupby('Pareto_ID').sum()
-    total_roofs['PV_m2'] = total_roofs.PVA_module_nbr * 1.6
-    # get occupation based on azimuth orientation
-    roofs = roofs.reset_index(level=['Azimuth'])
-    roofs['Orientation'] = roofs.apply(lambda x: split_orientation(x['Azimuth']), axis=1)
-    roofs = roofs.set_index('Orientation', append='True')
-    roofs = roofs.groupby(['Pareto_ID', 'Orientation']).sum()
-    max = roofs.PVA_module_nbr.groupby(level="Orientation").max()
-    roofs['occupancy'] = roofs.PVA_module_nbr / max
-
-    roofs_max = total_roofs.PV_m2.max()
-    roofs_max_m2 = roofs_max / hs_A
-
-    facades = df_n[df_n.index.get_level_values(level='Tilt') == 90]
-    if not facades.empty:
-        # get total m2 facades
-        total_facades = facades.groupby('Pareto').sum()
-        total_facades['PV_m2'] = total_facades.PVA_module_nbr * 1.6
-        # get occupation based on azimuth orientation
-        facades = facades.reset_index(level=['Azimuth'])
-        facades['Orientation'] = facades.apply(lambda x: split_orientation(x['Azimuth']), axis=1)
-        facades = facades.set_index('Orientation', append='True')
-        facades = facades.groupby(['Pareto', 'Orientation']).sum()
-        max = facades.PVA_module_nbr.max(level=1)
-        facades['occupancy'] = facades.PVA_module_nbr / max
[…]
