### 2024-07-22 4a63b03
fixing warnings, adding entry points for test
