+++ reho/model/__init__.py
@@ -1,3 +1,5 @@
+from . import reho
+
 __doc__ = """
 Directory for model-related code.
 """
