+++ documentation/conf.py
@@ -95,3 +95,4 @@ sys.modules['sklearn_extra'] = MagicMock()
 sys.modules['sklearn_extra.cluster'] = MagicMock()
 sys.modules['sqlalchemy'] = MagicMock()
 sys.modules['sqlalchemy.dialects'] = MagicMock()
+sys.modules['sqlalchemy.exc'] = MagicMock()
