### 2024-07-23 IPESE/REHO#24 @DorsanL
All the mentioned issues have been addressed here:
https://github.com/IPESE/REHO/commit/4a63b03aec4f85ae097582e4bd1ab2c3b5d03e0c

- no more warnings regarding geometry, division by 0, or deprecated methods
- I placed the data/ folder directly within scripts/examples/ to avoid downloading it separately
- regarding the reho/data/ folder, it is indeed part of the package as it contains essential data for the model to run (as well as default data, which the user can replace by supplying specific files for weather, units and grids specifications)
- the Windows-specific command `cmd /c "ampl_lic restart"` has been placed in an appropriate condition (this command is used when the model is run with distributed computing, for a very large number of optimizations)
