### 2024-07-23 IPESE/REHO#22 @DorsanL
Entry points for test have been added.
https://github.com/IPESE/REHO/commit/4a63b03aec4f85ae097582e4bd1ab2c3b5d03e0c

The "Checking proper installation" section of the documentation has been updated accordingly.
https://reho.readthedocs.io/en/main/sections/5_Getting_started.html#checking-proper-installation
