### 2024-07-23 IPESE/REHO#21 @DorsanL
Thank you for these suggestions!

1. Right, that's an interesting idea. Just to make sure I understand: are you suggesting integrating the notebook into the repository (and not into the package directly), and as well displaying its contents directly in the documentation? (I'll have to see if that's possible for a "for a direct integration", or if it has to go through an external link like yours)
2. I understand. My choice to work with html files allows me to generally adjust all this in post-processing, but I can add optional title and context.
3. This has been addressed.
4. I added a "Contribute" paragraph  in the README.md, which redirects to the [Contribute](https://reho.readthedocs.io/en/main/sections/7_Contribute.html section) section of the documentation. (I thought it would be redundant and less efficient for maintenance to copy the content? but if you would prefer to have a separate CONTRIBUTING.md file, I can add it)
