+++ requirements.txt
@@ -23,3 +23,4 @@ pvlib==0.10.*
 pyproj>=3.6.1
 urllib3>=2.1.0
 shapely>=2.0.2
+pytest>=8.3.0
