### 2024-07-23 IPESE/REHO#18 @DorsanL
Thank you for the feedback.

Currently, most paths in the code are defined with `os.path` rather than `Path`.
Unfortunately, we had noticed that `os.get_cwd()` command leads to different results when running the script from Pycharm (path of the file itself) or from VScode or terminal (path of the working directory). Cf [corresponding issue here](https://github.com/IPESE/REHO/issues/13).

However, the `Path(__file__).parent` function seems to always give the same result (whether running from Pycharm / VScode / terminal), and appears to solve the problem of relative paths when loading files such as `working_directory/data/filename.csv`.

Do you recommend using Path objects everywhere (code would need a bit of adaptation)? More specifically, what would you recommend instead of defining the relative paths based on `os.path.dirname(__file__)` in paths.py ?

1. `scripts/test/run.py` should now work.
2. and 3. I added `from . import reho` in `model/__init__.py`, does it solve your problem?
4. I added some entry points for testing the package without having to download the full (or parts of the) repository. I also included the `pytest` package to verify proper imports (https://github.com/IPESE/REHO/commit/84feda1aa251d7a3879843b6280b399b06cb996b). Any further suggestion is welcome :)
