### 2024-07-23 #6734 @DorsanL
@mbarzegary Here are some updates 👌

I've been able to work on the various issues mentioned and propose fixes. Some issues are still open, and feedback from reviewers @willu47 @hgandhi2411 @nmstreethran is welcome 👋

As the version is not yet stable, I have not deployed a new version of the package in PyPI. But the latest developments have been pushed to the `main`, and a local build allows you to check that the package is functional with:

```
python setup.py sdist bdist_wheel
pip install .
```

Entry points have also been included, and pytest package allows to verify proper installation.

```
reho-run-test
reho-plot-test
reho-examples-test
```

(In particular, 2 of the 19 examples -3e and 4b- do not work as intended. I am well aware of this and intend to correct it as soon as possible.)
