### 2024-07-25 6cf8ffe
debug calculation emissions with dynamic profile
