+++ reho/plotting/sankey.py
@@ -4,7 +4,7 @@ import numpy as np
 from reho.paths import *
 
 __doc__ = """
-Builds a dataframe for the visualization of annual flows from REHO results in the form of a Sankey diagram.
+Builds the dataframe for the visualization of annual flows from REHO results in the form of a Sankey diagram.
 """
 
 # Colors and labels for units and layers
@@ -17,17 +17,13 @@ def update_label(source_name, target_name, df_label):
 
     Parameters
     ----------
-    source_name: str
-        Source to update
-    target_name: str
-        Target to update
-    df_label: pd.DataFrame
-        Labels
+    source_name (string) : first name to update
+    target_name (string) : second name to update
+    df_label (df) : pd.DataFrame with labels
 
     Returns
     -------
-    pd.DataFrame
-        df_label updated with the source and target values
+    df_label (df) updated
     """
     if not (source_name in df_label.index):  # create label 'source' if not existing yet
         df_label.loc[source_name, 'pos'] = len(df_label)
@@ -43,19 +39,14 @@ def add_label_value(df_label, df_stv, precision, units):
 
     Parameters
     __________
-    df_label: pd.DataFrame
-        Labels
-    df_stv: pd.DataFrame
-        Source, target and value
-    precision: int
-        Precision of the displayed numbers (default = 2)
-    units: str
-        Unit of the values (default MWh)
+    df_label (df) : pd.DataFrame of labels
+    df_stv (df) : pd.DataFrame of source, target and value
+    precision (int): precision of the displayed numbers (default = 2)
+    units (string): unit of the values (default MWh)
 
     Returns
     _______
-    pd.DataFrame
-        df_label pdated with the label values
+    df label updated with the label values
     """
     df_source_value = pd.DataFrame()
     df_source_value.index = df_label.pos
@@ -73,35 +64,23 @@ def add_label_value(df_label, df_stv, precision, units):
 
 def add_flow(source, dest, layer, hub, dem_sup, df_annuals, df_label, df_stv, check_dest_2=False, dest_2=None, adjustment=0, fact=1):
     """
-    Adds an energy flow for the sankey diagram according cell(s) of df_annuals if cell not null
+    Adds an energy flow for the sankey diagramm for 'sankey_plot' according (a) cell(s) of df_annuals if cell not null
 
     Parameters
     ----------
 
-    source: str
-        name of the source
-    dest: str
-        name of the destination
-    layer: str
-        name of the layer of the considered cell(s)
-    hub: str
-        name of the hub of the considered cell(s)
-    dem_sup: str
-        'Supply_MWh' or 'Demand_MWh', column to take (! no control)
-    df_annuals: pd.DataFrame
-
-    df_label: pd.DataFrame
-
-    df_stv: pd.DataFrame
-
-    check_dest_2: bool
-        if True dest_2 substitute dest (default False)
-    dest_2: str
-        second possible destination (default None)
-    adjustment: float
-        offset added to the cell value (default 0)
-    fact: float
-        factor multiplied to the cell value (default 1)
+    source (string) : name of the source
+    dest (string) : name of the destination
+    layer (string) : name of the layer of the considered cell(s)
+    hub (string) : name of the hub of the considered cell(s)
+    dem_sup (string) : 'Supply_MWh' or 'Demand_MWh', column to take (! no control)
+    df_annuals (df) : df_annuals dataframe
[…]
