### 2024-08-02 fd2ae5b
adapt generation parameters to speed up preprocessing
