### 2024-08-19 IPESE/REHO#24 @nmstreethran
There's an error when running `4b_Sensitivity_analysis.py`:

```
python 4b_Sensitivity_analysis.py
Traceback (most recent call last):
  File "/media/nithiya/SSD2/REHO-git/scripts/examples/4b_Sensitivity_analysis.py", line 9, in <module>
    reader = QBuildingsReader()
             ^^^^^^^^^^^^^^^^
NameError: name 'QBuildingsReader' is not defined
```
