### 2024-08-19 IPESE/REHO#20 @nmstreethran
Re: my previous comment. Can you please check whether you are able to install `ampl_module_base` and `ampl_module_highs` using the requirements file?
