### 2024-08-19 IPESE/REHO#22 @nmstreethran
I still can't seem to find any tests to verify the functionality and methods in the package, e.g. assertion and exception testing. The tests you currently have only seem to check whether the imports work and if the scripts manage to execute. (Please correct me if I'm wrong.)

The pyheatdemand package I linked in my earlier comment has these tests for example: https://github.com/AlexanderJuestel/pyheatdemand/blob/main/test/test_processing.py.
