### 2024-08-29 a85c096
debug DHN example and clean data process DHN
