### 2024-08-29 3860931
all examples working with entry points
