+++ reho/model/postprocessing/__init__.py
@@ -1,4 +1,4 @@
-from . import sensitivity_analysis
+# from . import sensitivity_analysis (this line was causing issues with import in example 4b)
 
 __doc__ = """
 Directory where the output of the optimization from the AMPL model is extracted and processed to give a ``reho.results`` dictionary.
