+++ reho/tests/__init__.py
@@ -1,0 +1,3 @@
+__doc__ = """
+Contains test functions.
+"""
