### 2024-08-30 fae5097
add titles in plotting
