+++ scripts/examples/0_Compact_formulation.py
@@ -40,6 +40,6 @@ if __name__ == '__main__':
     reho.save_results(format=['xlsx', 'pickle'], filename='0')
 
     # Plot results
-    plotting.plot_performance(reho.results, plot='costs', indexed_on='Scn_ID', label='EN_long').show()
-    plotting.plot_performance(reho.results, plot='gwp', indexed_on='Scn_ID', label='EN_long').show()
-    plotting.plot_sankey(reho.results['totex'][0], label='EN_long', color='ColorPastel').show()
+    plotting.plot_performance(reho.results, plot='costs', indexed_on='Scn_ID', label='EN_long', title="Economical performance").show()
+    plotting.plot_performance(reho.results, plot='gwp', indexed_on='Scn_ID', label='EN_long', title="Environmental performance").show()
+    plotting.plot_sankey(reho.results['totex'][0], label='EN_long', color='ColorPastel', title="Sankey diagram").show()
