+++ scripts/examples/3e_DHN.py
@@ -42,5 +42,5 @@ if __name__ == '__main__':
     reho.save_results(format=['xlsx', 'pickle'], filename='3e')
 
     # Plot results
-    plotting.plot_performance(reho.results, plot='costs', indexed_on='Scn_ID', label='EN_long').show()
-    plotting.plot_sankey(reho.results['totex'][0], label='EN_long', color='ColorPastel').show()
+    plotting.plot_performance(reho.results, plot='costs', indexed_on='Scn_ID', label='EN_long', title="Economical performance").show()
+    plotting.plot_sankey(reho.results['totex'][0], label='EN_long', color='ColorPastel', title="Sankey diagram").show()
