+++ reho/plotting/plotting.py
@@ -14,7 +14,7 @@ Contains ready-to-use representations for results generated by REHO.
 
 
 def plot_performance(results, plot='costs', indexed_on='Scn_ID', label='EN_long', add_annotation=True, per_m2=False, additional_costs=None, additional_gwp=None,
-                     scc=0.177, filename=None, export_format='html', scaling_factor=1, return_df=False, unit_costs=' CHF'):
+                     scc=0.177, title=None, filename=None, export_format='html', scaling_factor=1, return_df=False):
     """
     Plots performance based on REHO results.
 
@@ -42,6 +42,8 @@ def plot_performance(results, plot='costs', indexed_on='Scn_ID', label='EN_long'
         Additional gwp to include (choose between 'isolation', 'mobility', and 'ict') and scaling values.
     scc: float
         Carbon externalities, expressed in *CHF/kgCO2*. Default value is the *Social Cost of Carbon*, from `Rennert, 2022 <https://www.nature.com/articles/s41586-022-05224-9>`_.
+    title: str
+        Title for the plot.
     filename: str
         Name of the file to be saved.
     export_format: str
@@ -50,8 +52,6 @@ def plot_performance(results, plot='costs', indexed_on='Scn_ID', label='EN_long'
         Scales linearly the REHO results for the plot.
     return_df: bool
         A dataframe can be returned for further post-processing or reporting purposes.
-    unit_costs: string
-        To specify the cost unit (' CHF', ' kCHF', ' MCHF', ' GCHF')
 
     Returns
     ----------
@@ -74,14 +74,12 @@ def plot_performance(results, plot='costs', indexed_on='Scn_ID', label='EN_long'
 
     change_data = pd.DataFrame()
     change_data.index = ['x_axis_1', 'x_axis_2', 'y_axis', 'keyword', 'total', 'unites', 'scc_legend']
-    decimal = 1
+    decimal = 0
     lang = re.split('_', label)[0]
-    unit_prefix = {' CHF': 1, ' kCHF': 1e3, ' MCHF': 1e6, ' GCHF': 1e9}
-    prefix = unit_prefix[unit_costs]
 
     if plot == 'costs':
-        change_data['FR'] = ['CAPEX', 'OPEX', 'Coûts [CHF/an]', 'Coûts', 'TOTEX', unit_costs, '']
-        change_data['EN'] = ['CAPEX', 'OPEX', 'Costs [CHF/y]', 'Costs', 'TOTEX', unit_costs, '']
+        change_data['FR'] = ['CAPEX', 'OPEX', 'Coûts [CHF/an]', 'Coûts', 'TOTEX', ' CHF', '']
+        change_data['EN'] = ['CAPEX', 'OPEX', 'Costs [CHF/y]', 'Costs', 'TOTEX', ' CHF', '']
         df_costs = df_Economics.xs('costs', level='Perf_type')
         if per_m2:
             df_costs = df_costs / era
@@ -154,9 +152,9 @@ def plot_performance(results, plot='costs', indexed_on='Scn_ID', label='EN_long'
     xtick = [x + 1 / 6 for x in x1]
 
     text_capacities = ["<b>" + change_data.loc['x_axis_1', lang] + "</b><br>" + str(custom_round(cp, decimal))
-                       for cp in combined_capacities / prefix]
+                       for cp in combined_capacities]
     text_resources = ["<b>" + change_data.loc['x_axis_2', lang] + "</b><br>" + str(custom_round(op, decimal))
-                      for op in combined_resources / prefix]
+                      for op in combined_resources]
[…]
