### 2024-08-30 b7d869a
include jupyter notebook in documentation
