### 2024-08-30 20bafe8
include demo jupyter notebook
