+++ documentation/requirements.txt
@@ -3,3 +3,4 @@ sphinxcontrib-bibtex
 sphinx-design
 pydata_sphinx_theme
 sphinx-copybutton
+nbsphinx
