+++ documentation/sections/Notebook.rst
@@ -1,0 +1,8 @@
+Notebook demo
++++++++++++++++++
+
+This jupyter notebook aims to illustrate a simple workflow with REHO.
+
+.. toctree::
+
+   demo
