### 2024-08-30 IPESE/REHO#21 @DorsanL
> There are ways to integrate Jupyter notebooks with Sphinx. See https://docs.readthedocs.io/en/stable/guides/jupyter.html. You can consider these in the future to enhance the documentation. This is not a requirement for the review.

I included a demo Jupyter Notebook in the documentation, now visible here: https://reho.readthedocs.io/en/main/sections/demo.html (https://github.com/IPESE/REHO/commit/20bafe85e5d47e342d9945a2a2a44599dd59fb45)

 > There are a number of plots generated when running the examples and a title (at least) would provide some context to users to better interpret the results.

 All the functions in plotting.py now have a `title` argument, and examples have been enhanced accordingly (https://github.com/IPESE/REHO/commit/fae50970a0484d7782b4d71e0e9852b23143fed0)
