### 2024-08-30 #6734 @DorsanL
Dear @mbarzegary,
I'm also back from my holiday period and fully available to take the necessary steps forward!

I've just answered and closed the remaining issues raised by the reviewers, made a push on the main (https://github.com/IPESE/REHO/commits/main/) and updated the documentation accordingly (https://reho.readthedocs.io/en/main/).
