### 2024-09-02 #6734 @mbarzegary
@nmstreethran and @hgandhi2411, can you please have a look at the fixes applied by @DorsanL and see if they resolve the issues raised by you. If so, I appreciate it if you finalize your reviews.
