### 2024-09-02 #6734 @nmstreethran
Hi @mbarzegary and @DorsanL. I made some comments on the closed issues some time ago that I'd like to bring to @DorsanL's attention. Ideally these should be resolved before I finalise my review.

1. https://github.com/IPESE/REHO/issues/20#issuecomment-2132479275
    The `ampl_module_base` and `ampl_module_highs` packages fail to install when using `requirements.txt`. I suggest the authors to test out the installation using a clean environment to ensure it proceeds and, if possible, set up continuous integration to test the installation on both Windows and Unix-like platforms.

1. https://github.com/IPESE/REHO/issues/24#issuecomment-2297146803
    One of the examples failed to run (`4b_Sensitivity_analysis.py`). @DorsanL also pointed out in https://github.com/openjournals/joss-reviews/issues/6734#issuecomment-2245293322 that there were issues with two of the examples. Have these been fixed?

1. https://github.com/IPESE/REHO/issues/22#issuecomment-2297210111
    Some tests have been implemented but these seem to only test the module imports and script execution. I can't find any functionality tests, so I think the authors still need to work on this issue. JOSS specifically requires "automated tests or manual steps described so that the functionality of the software can be verified".
