+++ setup.py
@@ -14,7 +14,7 @@ def read_file(file_path):
 setup(
 
     name='REHO',
-    version='1.1.0',
+    version='1.1.1',
     entry_points={
         'console_scripts': [
             'reho-test-run = reho.tests.test_scripts:test_run',
@@ -24,33 +24,38 @@ setup(
     },
     packages=find_packages(),
     include_package_data=True,
-    install_requires=['amplpy>=0.12.0,<0.13.0',
-                      'ampl_module_base',
-                      'ampl_module_highs',
-                      'pandas>=1.5.3,<2.0.0',
-                      'openpyxl>=3.1.2,<4.0.0',
-                      'numpy>=1.23.4,<2.0.0',
-                      'scipy>=1.9.2,<2.0.0',
-                      'scikit-learn>=1.2.2,<2.0.0',
-                      'scikit-learn-extra>=0.3.0',
-                      'sqlalchemy>=1.4.42,<2.0.0',
-                      'psycopg2>=2.9.4,<3.0.0 ; platform_system != "Windows"',
-                      'psycopg2-binary>=2.9.9,<3.0.0 ; platform_system == "Windows"',
-                      'geopandas>=0.12.2,<1.0.0',
-                      'matplotlib>=3.6.1,<4.0.0',
-                      'plotly>=5.10,<6.0.0',
-                      'kaleido>=0.2.1,<1.0.0',
-                      'python-dotenv>=1.0',
-                      'requests>=2.0.0,<3.0.0',
-                      'coloredlogs>=15.0.1',
-                      'SALib>=1.4.8',
-                      'qmcpy>=1.4.5',
-                      'pvlib==0.10.*',
-                      'pytest>=8.3.0',
-                      ],
+    install_requires=[
+        "amplpy",
+        "ampl_module_base",
+        "ampl_module_highs",
+        "coloredlogs",
+        "geopandas<1.0.0",
+        "kaleido",
+        "matplotlib",
+        "numpy",
+        "openpyxl",
+        "pandas<2.0.0",
+        "plotly",
+        "psycopg2<3.0.0 ; platform_system != 'Windows'",
+        "psycopg2-binary<3.0.0 ; platform_system == 'Windows'",
+        "pvlib",
+        "pyproj",
+        "python-dotenv",
+        "pytest",
+        "qmcpy",
+        "requests",
+        "SALib",
+        "scikit-learn==1.3.2",
+        "scikit-learn-extra",
+        "scipy",
+        "shapely",
+        "sqlalchemy<2.0.0",
+        "urllib3",
+    ],
+
     package_data={
-          '': ['*.csv', '*.xlsx', '*.dat', '*.txt' '*.mod', '*.ini'],
-      },
+        '': ['*.csv', '*.xlsx', '*.dat', '*.txt' '*.mod', '*.ini'],
+    },
     author='Dorsan Lepour',
     author_email='dorsan.lepour@epfl.ch',
     maintainer="IT team of IPESE Laboratory from EPFL",
