+++ documentation/conf.py
@@ -68,28 +68,33 @@ toc_object_entries_show_parents = 'all'
 
 # ------------ Autodoc ------------------------------------
 autodoc_member_order = 'bysource'
-autodoc_mock_imports = ['amplpy',
-                        'pandas',
-                        'openpyxl',
-                        'numpy',
-                        'scipy',
-                        'scikit-learn',
-                        'scikit-learn-extra',
-                        'sqlalchemy',
-                        'psycopg2',
-                        'geopandas',
-                        'matplotlib',
-                        'plotly',
-                        'kaleido',
-                        'dotenv',
-                        'requests',
-                        'coloredlogs',
-                        'SALib',
-                        'qmcpy',
-                        'pvlib',
-                        'pyproj',
-                        'shapely',
-                        'pytest']
+autodoc_mock_imports = [
+    "amplpy",
+    "ampl_module_base",
+    "ampl_module_highs",
+    "coloredlogs",
+    "geopandas",
+    "kaleido",
+    "matplotlib",
+    "numpy",
+    "openpyxl",
+    "pandas<2.0.0",
+    "plotly",
+    "psycopg2",
+    "pvlib",
+    "pyproj",
+    "python-dotenv",
+    "pytest",
+    "qmcpy",
+    "requests",
+    "SALib",
+    "scikit-learn",
+    "scikit-learn-extra",
+    "scipy",
+    "shapely",
+    "sqlalchemy",
+    "urllib3",
+]
 sys.modules['scikit-learn'] = MagicMock()
 sys.modules['sklearn'] = MagicMock()
 sys.modules['sklearn.metrics'] = MagicMock()
