+++ requirements.txt
@@ -1,26 +1,27 @@
-amplpy>=0.12.0,<0.13.0
---extra-index-url https://pypi.ampl.com ampl_module_base
---extra-index-url https://pypi.ampl.com ampl_module_highs
-pandas>=1.5.3,<2.0.0
-openpyxl>=3.1.2,<4.0.0
-numpy>=1.23.4,<2.0.0
-scipy>=1.9.2,<2.0.0
-scikit-learn>=1.2.2,<2.0.0
-scikit-learn-extra>=0.3.0
-sqlalchemy>=1.4.42,<2.0.0
-psycopg2>=2.9.4,<3.0.0 ; platform_system != "Windows"
-psycopg2-binary>=2.9.9,<3.0.0 ; platform_system == "Windows"
-geopandas>=0.12.2,<1.0.0
-matplotlib>=3.6.1,<4.0.0
-plotly>=5.10,<6.0.0
-kaleido>=0.2.1,<1.0.0
-python-dotenv>=1.0
-requests>=2.0.0,<3.0.0
-coloredlogs>=15.0.1
-SALib>=1.4.8
-qmcpy>=1.4.5
-pvlib==0.10.*
-pyproj>=3.6.1
-urllib3>=2.1.0
-shapely>=2.0.2
-pytest>=8.3.0
+amplpy
+--extra-index-url https://pypi.ampl.com
+ampl_module_base
+ampl_module_highs
+coloredlogs
+geopandas<1.0.0
+kaleido
+matplotlib
+numpy
+openpyxl
+pandas<2.0.0
+plotly
+psycopg2<3.0.0 ; platform_system != "Windows"
+psycopg2-binary<3.0.0 ; platform_system == "Windows"
+pvlib
+pyproj
+python-dotenv
+pytest
+qmcpy
+requests
+SALib
+scikit-learn==1.3.2
+scikit-learn-extra
+scipy
+shapely
+sqlalchemy<2.0.0
+urllib3
