### 2024-09-09 IPESE/REHO#20 @DorsanL
I updated the requirements.txt according to your suggestion, and verified that the package and all examples are properly working in a fully new virtual environment. 👌
(https://github.com/IPESE/REHO/commit/70d7b11dabd67d50d87c9025ea896435e377f53c)
