### 2024-09-09 #6734 @DorsanL
> Regarding these issues (and the next that will come), is it preferred that reviewers close them themselves or shall I take care of it?

Dear @mbarzegary, I'm sorry for the confusion! As you can see above, I asked the reviewers whether I should close the issues myself, and the answer was positive. But I agree that it's preferable they are closed after validation by the reviewers.
(I therefore just reopened the 3 issues mentioned above)

@nmstreethran: I answered https://github.com/IPESE/REHO/issues/20#issuecomment-2132479275 and https://github.com/IPESE/REHO/issues/24#issuecomment-2297146803, and I'm working on functionality tests that I hope to commit tomorrow 🤞
