### 2024-09-10 622cffb
erratum : bug with constraints activation
