+++ reho/model/district_decomposition.py
@@ -94,7 +94,7 @@ class district_decomposition:
                                                 "DailyDist","Mode_Speed","outside_charging_price","charging_externalload","Districts","share_district_activity","externalload_sellingprice",
                                                 "max_share_cars" ,  "min_share_cars" ,  "max_share_PT" , "min_share_PT" , "max_share_MD" , "min_share_MD" , "max_share_ICE" ,  "min_share_ICE" ,
                                                  "max_share_EV" , "min_share_EV" , "max_share_PT_train" ,    "min_share_PT_train" ,    "max_share_EBikes" , "n_ICEperhab"], # attention District is a SET
-                         "list_constraints_MP": []
+                         "list_constraints_MP": ["EV_supplyprofile1","EV_supplyprofile2"]
                          }
 
         self.df_fix_Units = pd.DataFrame()
