### 2024-09-10 fda9b0e
some functionality tests with pytest
