+++ reho/model/preprocessing/weather.py
@@ -9,6 +9,8 @@ from reho.model.preprocessing.clustering import Clustering
 import pvlib
 import geopandas as gpd
 from pyproj import Transformer
+import csv
+
 
 __doc__ = """
 Generates the meteorological data (temperature and solar irradiance).
@@ -134,6 +136,19 @@ def read_custom_weather(path_to_weather_file):
     df2 = pd.read_csv(os.path.join(path_to_weather, 'weekday.txt'), index_col=[0], header=None)
     df['Weekday'] = df2
 
+    df3 = pd.read_csv(os.path.join(path_to_weather, 'Elec_CO2_2023.txt'), index_col=[0], header=None)
+    df['Emissions'] = df3
+
+    df4 = pd.read_csv(os.path.join(path_to_weather, 'yearly_data_centre_profile_repeated.csv'), index_col= False, header=None)
+    df4 = df4.iloc[1:, :]
+
+
+    df4 = df4.reset_index()
+    df4 = df4.iloc[:,2]
+    print(df4)
+    df4 = df4.astype(float)
+    df['DataLoad'] = df4
+
     return df
 
 
@@ -222,13 +237,39 @@ def generate_output_data(cl, attributes, location):
     # Call for the write_dat_files function
     write_dat_files(attributes, location, data_cls, data_idy)
 
-    return print(f'The data have been computed and saved in {path_to_clustering}.')
+    print(f'The data have been computed and saved in {path_to_clustering}.')
+    return data_cls
+def data_centre_profiles(value):
+    df_D = value['DataLoad']
+    return df_D.to_numpy()
 
 def get_metric(cluster):
     if 'E' in cluster['Attributes']:
         return 'GWP100a'
     else:
         return 'method 1'
+
+def data_centre_profile(size): # size to be mentioned in kW
+    df_load_profile = pd.read_csv('/Users/ravi/Desktop/PhD/My_Reho_Qgis_files/Reho_Sai_Fork/scripts/template/data/profiles/data_centre_hourly_week.csv')
+    df_load_profile= df_load_profile['Load_Profile'].div(50).mul(size) #profile data created by observing trend from this study: https://arxiv.org/abs/1804.00703
+    df_load_annual = pd.concat([df_load_profile]*70).to_frame().reset_index()
+    df_load_annual = df_load_annual['Load_Profile'].to_frame()
+  # Replace with your actual data
+
+# Repeat data for the entire year (8760 hours)
+
+
+# Prepare data in the format for CSV
+    csv_data = []
+    for hour in range(8760):
+        value = df_load_annual['Load_Profile'][hour]  # Repeat data cyclically
+        csv_data.append([hour + 1, value])  # Hour starts from 1
+
+    csv_filename = '/Users/ravi/Desktop/PhD/My_Reho_Qgis_files/Reho_Sai_Fork/reho/data/weather/yearly_data_centre_profile_repeated.csv'
+    with open(csv_filename, mode='w', newline='') as file:
+        writer = csv.writer(file)
+        writer.writerow(['Hour', 'Load_Profile'])  # Header row
+        writer.writerows(csv_data)
 def write_dat_files(attributes, location, values_cluster, index_inter):
     """
     Writes the clustering results computed from ``generate_output_data`` as .dat files.
[…]
