+++ scripts/examples/1b_Building-scale_Pareto.py
@@ -10,7 +10,7 @@ if __name__ == '__main__':
     qbuildings_data = reader.read_db(transformer=234, nb_buildings=1)
 
     # Select clustering options for weather data
-    cluster = {'Location': 'Geneva', 'Attributes': ['T', 'I', 'W'], 'Periods': 10, 'PeriodDuration': 24}
+    cluster = {'Location': 'Geneva', 'Attributes': ['T', 'I'], 'Periods': 10, 'PeriodDuration': 24}
 
     # Set scenario
     scenario = dict()
