### 2024-09-10 IPESE/REHO#22 @DorsanL
I started implementing some functionality tests with pytest (https://github.com/IPESE/REHO/commit/fda9b0ef9b6ff29f01211cc0b6a6e6cb66a8eca9). This is only a draft, I will extend them in the coming days.
