### 2024-09-11 a1c9168
cleaned old EV stuff
