### 2024-09-13 ee797a1
Clean code
