### 2024-09-17 254a8bc
enhancing functionality tests
