+++ requirements.txt
@@ -22,6 +22,7 @@ SALib
 scikit-learn==1.3.2
 scikit-learn-extra
 scipy
+setuptools
 shapely
 sqlalchemy<2.0.0
 urllib3

[…]
