+++ scripts/examples/3d_EVs.py
@@ -33,7 +33,7 @@ if __name__ == '__main__':
     # Set specific parameters
 
     # Run optimization
-    reho = reho(qbuildings_data=qbuildings_data, units=units, grids=grids, cluster=cluster, scenario=scenario,
+    reho = REHO(qbuildings_data=qbuildings_data, units=units, grids=grids, cluster=cluster, scenario=scenario,
                 method=method, parameters=parameters, solver="gurobi")
     reho.single_optimization()
 

[…]
