### 2024-09-17 IPESE/REHO#22 @DorsanL
A folder `reho/test/` is now part of the package, containing the following files and functionality tests:

- _test_import.py_ verifies the import of the different **REHO** modules
- _test_clustering.py_ verifies the behavior of **Clustering** class
- _test_infrastructure.py_ verifies the behavior of **Infrastructure** class
- _test_run.py_ ensures that a classic optimisation works as intended (and thereby that the **SubProblem**, **MasterProblem** and **REHO** classes work properly)
- _test_plot.py_ (calls _test_run_ to generate results) and verifies various plotting functions
- _test_examples.py_ runs each of the examples available in `scripts/examples/` (and downloads them locally beforehand if they have not been cloned)

Developments have been pushed in https://github.com/IPESE/REHO/commit/254a8bc1476dc9250208d94b6446b4953017c783
