### 2024-09-18 3250004
code polishing from Cedric's comments
