+++ reho/model/district_decomposition.py
@@ -91,10 +91,11 @@ class district_decomposition:
         self.lists_MP = {"list_parameters_MP": ['utility_portfolio_min', 'owner_portfolio_min', 'EMOO_totex_renter', 'TransformerCapacity',
                                                 'EV_y', 'EV_plugged_out', 'n_vehicles', 'EV_capacity', 'monthly_grid_connection_cost',
                                                 "area_district", "velocity", "density", "delta_enthalpy", "cinv1_dhn", "cinv2_dhn","Population","transport_Units",
-                                                "DailyDist","Mode_Speed","outside_charging_price","charging_externalload","Districts","share_district_activity","externalload_sellingprice",
+                                                "DailyDist","Mode_Speed","outside_charging_price","charging_externalload","share_district_activity","externalload_sellingprice",
                                                 "max_share_cars" ,  "min_share_cars" ,  "max_share_PT" , "min_share_PT" , "max_share_MD" , "min_share_MD" , "max_share_ICE" ,  "min_share_ICE" ,
-                                                 "max_share_EV" , "min_share_EV" , "max_share_PT_train" ,    "min_share_PT_train" ,    "max_share_EBikes" , "n_ICEperhab"], # attention District is a SET
-                         "list_constraints_MP": ["EV_supplyprofile1","EV_supplyprofile2"]
+                                                 "max_share_EV" , "min_share_EV" , "max_share_PT_train" ,    "min_share_PT_train" ,    "max_share_EBikes" , "n_ICEperhab"],
+                         "list_constraints_MP": ["EV_supplyprofile1","EV_supplyprofile2"],
+                         "list_set_indexed_MP" : ["Districts"]
                          }
 
         self.df_fix_Units = pd.DataFrame()
@@ -255,7 +256,7 @@ class district_decomposition:
         print('INITIATE HOUSE: ', h)
 
         # find district structure and parameter for one single building
-        buildings_data_SP, parameters_SP, infrastructure_SP = self.__split_parameter_sets_per_building(h)
+        buildings_data_SP, parameters_SP, infrastructure_SP, set_indexed_SP = self.__split_parameter_sets_per_building(h)
 
         # epsilon constraints on districts may lead to infeasibilities on building level -> apply them in MP only
         if epsilon_init is not None and self.method['building-scale']:
@@ -273,9 +274,9 @@ class district_decomposition:
             parameters_SP['beta_duals'] = beta_list
 
         if self.method['use_facades'] or self.method['use_pv_orientation']:
-            REHO = compact_optimization(infrastructure_SP, buildings_data_SP, parameters_SP, self.set_indexed, self.cluster, scenario, self.method, self.solver, self.qbuildings_data)
+            REHO = compact_optimization(infrastructure_SP, buildings_data_SP, parameters_SP, set_indexed_SP, self.cluster, scenario, self.method, self.solver, self.qbuildings_data)
         else:
-            REHO = compact_optimization(infrastructure_SP, buildings_data_SP, parameters_SP, self.set_indexed, self.cluster, scenario, self.method, self.solver)
+            REHO = compact_optimization(infrastructure_SP, buildings_data_SP, parameters_SP, set_indexed_SP, self.cluster, scenario, self.method, self.solver)
         ampl = REHO.build_model_without_solving()
 
         if self.method['fix_units']:
@@ -443,8 +444,6 @@ class district_decomposition:
 
 
         for key in self.lists_MP['list_parameters_MP']:
-            if key == "Districts": # bc it's a SET
-                continue
             if key in self.parameters.keys():
                 MP_parameters[key] = self.parameters[key]
 
@@ -456,17 +455,7 @@ class district_decomposition:
             if len(self.infrastructure.UnitsOfDistrict) != 0:
[…]
