+++ reho/model/preprocessing/EV_profile_generator.py
@@ -89,7 +89,7 @@ def generate_EV_plugged_out_profiles_district(cluster):
     return EV_plugged_out, EV_plugging_in
 
 
-def generate_mobility_parameters(cluster, population, dailydist,mode_speed_custom,transportunits):
+def generate_mobility_parameters(cluster, parameters,transportunits):
     """
     Based on EV_profile_generator_structure
 
@@ -99,27 +99,28 @@ def generate_mobility_parameters(cluster, population, dailydist,mode_speed_custo
     ----------
     cluster : 
         to get periods characterisations (p,t) => usually the value self.cluster_compact or 
-    mode_speed_custom : None or a dictionnay
-        either None or a dictionnay given by the user in the scenario initialisation
-    dailydist : float
-        duplicata of the ampl param Dailydist, declared later in the program (TODO : Could be coded better bc the duplicatas are not linked)
-    population : float
-        the input reho.parameters['Population'] in the initialisation of the scenario. duplicata of the ampl param Population, declared later in the program (TODO : Could be coded better bc the duplicatas are not linked)
+    parameters :
+        From the parameters will be extracted values related to the mobility namely DailyDist, Mode_Speed and Population. Population and DailyDist are float, Mode_Speed is a dictionnary given by the user in the scenario initialisation. It can contains customed values for only some modes while the other remain default. 
     transportunits : list
-        a list of all infrastructure units providing Mobility + "Public_transport" => which is the Network supply['Mobility']
+        a list of all infrastructure units providing Mobility + "Public_transport" => which is the Network_supply['Mobility']
 
     Returns:
     -------
     param_output : a dict of dataframes containing the profiles for each param. 
+
+    ..caution:
+    The default values in this function are a hardcoded copy of parameters DailyDist and Population in mobility.mod.
     """
     param_output = dict()
     
-    if dailydist is None:
-        dailydist = 36.8 # km per day
-    if mode_speed_custom is None:
+    if not "DailyDist" in parameters:
+        parameters['DailyDist'] = 36.8 # km per day
+    if not "Population" in parameters:
+        parameters['Population'] = 10 # km per day  
+    if "Mode_Speed" in parameters:
+        mode_speed_custom = parameters['Mode_Speed']
+    else:
         mode_speed_custom = pd.DataFrame()
-    if population is None:
-        population = 10 # km per day
 
     # Periods
     # TODO IMPLEMENTATION of flexible period duration
@@ -151,8 +152,8 @@ def generate_mobility_parameters(cluster, population, dailydist,mode_speed_custo
     # Domestic demand ================================================================================================
     # The labels look like this : demwdy => normalized mobility demand of a weekday
     columns = ["dem" + x for x in days_mapping.values()]
-    profiles_input[columns] *= population
-    profiles_input[columns] *= dailydist
+    profiles_input[columns] *= parameters['Population']
+    profiles_input[columns] *= parameters['DailyDist']
 
     mobility_demand = pd.DataFrame(columns=['l', 'p', 't', 'Domestic_energy'])
 
@@ -193,7 +194,7 @@ def generate_mobility_parameters(cluster, population, dailydist,mode_speed_custo
[…]
