+++ requirements.txt
@@ -13,14 +13,13 @@ plotly
 psycopg2<3.0.0 ; platform_system != "Windows"
 psycopg2-binary<3.0.0 ; platform_system == "Windows"
 pvlib
+pyclustering
 pyproj
 python-dotenv
 pytest
 qmcpy
 requests
 SALib
-scikit-learn==1.3.2
-scikit-learn-extra
 scipy
 setuptools
 shapely
