### 2024-09-18 b520885
cleaning clustering code and plot
