+++ reho/model/preprocessing/weather.py
@@ -77,7 +77,7 @@ def generate_weather_data(cluster, qbuildings_data):
         attributes.append('Emissions')
 
     df = df[attributes]
-    cl = Clustering(data=df, nb_clusters=[cluster['Periods']], option={"year-to-day": True, "extreme": []}, pd=cluster['PeriodDuration'])
+    cl = Clustering(data=df, nb_clusters=[cluster['Periods']], period_duration=cluster['PeriodDuration'], options={"year-to-day": True, "extreme": []})
     cl.run_clustering()
 
     generate_output_data(cl, attributes, cluster['Location'])
@@ -181,8 +181,8 @@ def generate_output_data(cl, attributes, location):
     # Get the max/min irradiance from the same day
     T_max = cl.data_org.iloc[[T_idx[1]]].copy()
     T_max['Irr'] = cl.data_org.loc[T_day[1] * 24: T_day[1] * 24 + 24, 'Irr'].max()
-    T_min.loc[:, ['time.dd', 'time.hh', 'dt']] = [T_day[0], 1, 1]
-    T_max.loc[:, ['time.dd', 'time.hh', 'dt']] = [T_day[1], 1, 1]
+    T_min[['time.dd', 'time.hh', 'dt']] = [T_day[0], 1, 1]
+    T_max[['time.dd', 'time.hh', 'dt']] = [T_day[1], 1, 1]
     data_cls = pd.concat([data_cls, T_min.rename({T_idx[0]: 240}), T_max.rename({T_idx[1]: 241})])
     # Add a 10% margin for the extreme over 20 years
     data_cls.loc[[240, 241], ['Text', 'Irr']] = data_cls.loc[[240, 241], ['Text', 'Irr']] * 1.1
@@ -200,7 +200,7 @@ def generate_output_data(cl, attributes, location):
     data_idp = pd.DataFrame(
         np.stack((np.arange(1, data_cls.shape[0] + 1, 1), np.arange(1, data_cls.shape[0] + 1, 1)), axis=1),
         columns=["IndexDy", "intra_t"])
-    data_idp["intra_end"] = [id + cl.pd if (id % cl.pd) == 0 else 0 for id in data_idp.index]
+    data_idp["intra_end"] = [id + cl.period_duration if (id % cl.period_duration) == 0 else 0 for id in data_idp.index]
 
     # Call for the write_dat_files function
     write_dat_files(attributes, location, data_cls, data_idy)
@@ -366,174 +366,146 @@ def write_dat_files(attributes, location, values_cluster, index_inter):
 
 
 def plot_cluster_KPI_separate(df, save_fig):
-    plt.rcParams.update({'font.size': 18})
-    print(df)
+    # Transpose the DataFrame
     df = df.transpose()
-    df.columns.names = ['KPI']
-    df.rename({'Irr': 'Global Irradiation', 'Text': 'Ambient Temperature'}, inplace=True)
-    df_irr = df.xs('Global Irradiation', level=1)
-    df_T = df.xs('Ambient Temperature', level=1)
 
+    # Ensure proper column renaming
+    df.columns.names = ['iteration', 'KPI']
+    df = df.rename(index={'Text': 'Ambient Temperature', 'Irr': 'Global Irradiation'})
+
+    # Select the relevant KPIs for plotting
+    df_irr = df.xs('Global Irradiation', axis=0)
+    df_T = df.xs('Ambient Temperature', axis=0)
+
+    # Plot RMSD and LDC for Global Irradiation and Ambient Temperature
     fig, ax = plt.subplots()
     fig.set_size_inches(4, 8)
[…]
