+++ reho/model/preprocessing/clustering.py
@@ -1,4 +1,3 @@
-import scipy.spatial
 from pyclustering.cluster.kmedoids import kmedoids
 from pyclustering.utils.metric import distance_metric, type_metric
 from pyclustering.utils import calculate_distance_matrix
@@ -12,49 +11,32 @@ Clustering algorithm for input data reduction.
 
 class Clustering:
     """
-    Executes a clustering for each number of clusters among a specified interval (nb_clusters), and selects the optimal one according to the MAPE criterion (Mean Average Percentage Error).
+    Executes a clustering for each number of clusters among a specified interval (nb_clusters),
+    and selects the optimal one according to the MAPE criterion (Mean Average Percentage Error).
 
-    Parameters
-    ----------
+    Parameters:
+    -----------
     data : pd.DataFrame
         Annual weather data
     nb_clusters : list
         Interval for the number of clusters possible.
-
     """
 
-    def __init__(self, data, nb_clusters=None, option=None, pd=None, outliers=None):
-        super().__init__()
-        if nb_clusters is None:
-            self.nb_clusters = [8]
-        else:
-            self.nb_clusters = nb_clusters
-
-        # org - original, nor - normalized, clu - clustered
-
+    def __init__(self, data, nb_clusters=None, period_duration=24, outliers=None, options=None):
+        """
+        Initializes the Clustering object with the given parameters.
+        """
+        self.nb_clusters = nb_clusters if nb_clusters is not None else [8]
         self.outlier_removal = outliers
-
-        self.data_org = data  # input data- df: attribute vectors as columns
+        self.data_org = data
+        self.period_duration = period_duration
         self.data_nor = None
-        self.attr_org = []  # matrix or attributes, arrays:  len(attr_org) = total number of periods per year, len(attr_org[0])= number of attributes x period duration
-        self.attr_nor = []
+        self.attr_org, self.attr_nor = [], []
+        self.mod_org, self.mod_nor = [], []
+        self.option = options or {"year-to-day": True, "extreme": [(1, "min"), (2, "max")]}
+        self.results = {"idx": None}
         self.attr_clu = None
-        self.mod_org = []  # modulo data, f.e. last day of year (24hrs) for  pd = 7*24
-        self.mod_nor = []
-        if option is None:
-            self.option = {"year-to-day": True, "extreme": [(1, "min"), (2, "max")]}
-        else:
-            self.option = option
-
-        self.results = {"idx": None}  # pd.DataFrame()
-
-        self.kpis_clu = None
-        self.nbr_opt = None
-
-        if pd is None:  # period duration, f.e. 1 period = 1 day => pd =  24 (default)
-            self.pd = 24
-        else:
-            self.pd = pd
+        self.kpis_clu, self.nbr_opt = None, None
 
     def run_clustering(self):
         self.__do_normalization()
@@ -63,173 +45,95 @@ class Clustering:
         self.__select_optimal()
 
     def __do_normalization(self):
-        # - Load data set
-        self.data_nor = pd.DataFrame([], index=self.data_org.index, columns=self.data_org.columns)
-        # - Create dissimilarity matrix
-        for id, col in enumerate(self.data_org):
-            self.data_nor.loc[:, col] = ((self.data_org.loc[:, col] - self.data_org.loc[:, col].min()) / (
-                        self.data_org.loc[:, col].max() - self.data_org.loc[:, col].min()))  # normalize
-            # - OPTION : re-arrange dataframe
+        """
+        Normalizes the data and reshapes it for clustering based on the defined period (e.g., daily).
+        """
[…]
