### 2024-09-19 e3e3307
nomenclature, clean EV model, examples 6ab
