+++ reho/model/district_decomposition.py
@@ -91,10 +91,10 @@ class district_decomposition:
         self.lists_MP = {"list_parameters_MP": ['utility_portfolio_min', 'owner_portfolio_min', 'EMOO_totex_renter', 'TransformerCapacity',
                                                 'EV_y', 'EV_plugged_out', 'n_vehicles', 'EV_capacity', 'monthly_grid_connection_cost',
                                                 "area_district", "velocity", "density", "delta_enthalpy", "cinv1_dhn", "cinv2_dhn","Population","transport_Units",
-                                                "DailyDist","Mode_Speed","outside_charging_price","charging_externalload","share_district_activity","externalload_sellingprice",
+                                                "DailyDist","Mode_Speed","Cost_demand_ext","EV_charger_supply_ext","share_activity","Cost_supply_ext",
                                                 "max_share_cars" ,  "min_share_cars" ,  "max_share_PT" , "min_share_PT" , "max_share_MD" , "min_share_MD" , "max_share_ICE" ,  "min_share_ICE" ,
                                                  "max_share_EV" , "min_share_EV" , "max_share_PT_train" ,    "min_share_PT_train" ,    "max_share_EBikes" , "n_ICEperhab"],
-                         "list_constraints_MP": ["EV_supplyprofile1","EV_supplyprofile2"],
+                         "list_constraints_MP": ["EV_chargingprofile1","EV_chargingprofile2"],
                          "list_set_indexed_MP" : ["Districts"]
                          }
 
@@ -370,11 +370,11 @@ class district_decomposition:
                 ampl_MP.read('mobility.mod')
             if "EV_district" in self.infrastructure.UnitsOfDistrict:
                 ampl_MP.read('evehicle.mod')
-                ampl_MP.getConstraint('EV_supplyprofile1').drop()
-                ampl_MP.getConstraint('EV_supplyprofile2').drop()
+                ampl_MP.getConstraint('EV_chargingprofile1').drop()
+                ampl_MP.getConstraint('EV_chargingprofile2').drop()
                 ampl_MP.getConstraint('ExternalEV_Costs_positive').drop()
             else:
-                scenario['exclude_units'] = scenario['exclude_units']  + ['EVCharging_district']
+                scenario['exclude_units'] = scenario['exclude_units']  + ['EV_charger_district']
             if "Bike_district" in self.infrastructure.UnitsOfDistrict:
                 ampl_MP.read('bike.mod')
             if "ElectricBike_district" in self.infrastructure.UnitsOfDistrict:
@@ -454,7 +454,6 @@ class district_decomposition:
         if 'EV_plugged_out' not in MP_parameters:
             if len(self.infrastructure.UnitsOfDistrict) != 0:
                 if 'EV_district' in self.infrastructure.UnitsOfDistrict:
-                    MP_parameters['EV_plugged_out'], MP_parameters['EV_plugging_in'] = EV_gen.generate_EV_plugged_out_profiles_district(self.cluster)
                     p = EV_gen.generate_mobility_parameters(self.cluster,self.parameters,
                                                             np.append(self.infrastructure.UnitsOfLayer["Mobility"],'Public_transport'))
                     MP_parameters.update(p)
