+++ reho/model/preprocessing/EV_profile_generator.py
@@ -4,91 +4,6 @@ import pandas as pd
 import numpy as np
 
 
-def generate_EV_plugged_out_profiles_district(cluster):
-    """
-    Computes hourly electric vehicle (EV) profiles for each typical day considering weekdays and weekends. Data are taken from
-    `UK Department for Transport 2013 <https://www.gov.uk/government/collections/energy-and-environment-statistics#publications>`_
-    and `SFSO 2015 <https://www.bfs.admin.ch/asset/fr/1840478>`_. The EV occupancy profiles are used to
-    optimize EV electricity demand profiles with the evehicle.mod ampl model.
-
-
-
-    Parameters
-    ----------
-    cluster : dict
-        Define location district, number of periods, and number of timesteps.
-
-    Returns
-    -------
-    EV_plugged_out : array
-        Hourly profile of the share of vehicles being plugged out of the district LV grid.
-    EV_plugging_in : array
-        Hourly profile of the share of vehicles connecting to the district LV grid.
-
-    Notes
-    -----
-    - EV_plugged_out, EV_plugging_in
-
-    """
-    # TODO IMPLEMENTATION of flexible period duration
-    File_ID = WD.get_cluster_file_ID(cluster)
-
-    if 'W' in File_ID.split('_'):
-        use_weekdays = True
-    else:
-        use_weekdays = False
-
-    if use_weekdays:
-        timestamp = np.loadtxt(os.path.join(path_to_clustering, 'timestamp_' + File_ID + '.dat'), usecols=(1, 2, 3),
-                               skiprows=1)
-        timestamp = pd.DataFrame(timestamp, columns=("Day", "Frequency", "Weekday"))
-    else:
-        df = pd.read_csv(os.path.join(path_to_clustering, 'timestamp_' + File_ID + '.dat'), delimiter='\t')
-        timestamp = df.fillna(1)  # only weekdays
-
-    # Federal Office of Statistic, Comportement de la population en matiere de transports, 2015
-    profiles_weekday = [0.01, 0.0, 0.0, 0.0, 0.0, 0.03, 0.1, 0.12, 0.11, 0.12, 0.12, 0.14, 0.13, 0.14, 0.13, 0.13, 0.18,
-                        0.2, 0.15, 0.1, 0.07, 0.06, 0.05, 0.03]
-    profiles_weekend = [0.01, 0.0, 0.0, 0.0, 0.0, 0.00, 0.0, 0.00, 0.03, 0.10, 0.12, 0.11, 0.12, 0.12, 0.14, 0.13, 0.14,
-                        0.13, 0.15, 0.1, 0.07, 0.06, 0.05, 0.03]
-
-    # UK department for transport Electric Chargepoint Analysis 2017: Domestics
-    profiles_weekday = [0.0, 0.0, 0.0, 0.0, 0.01, 0.02, 0.08, 0.2, 0.31, 0.35, 0.38, 0.39, 0.39, 0.38, 0.38, 0.36, 0.31,
-                        0.24, 0.18, 0.13, 0.09, 0.05, 0.03, 0.01]
-    profiles_weekend = [0.0, 0.0, 0.0, 0.0, 0.01, 0.02, 0.08, 0.2, 0.31, 0.35, 0.38, 0.39, 0.39, 0.38, 0.38, 0.36, 0.31,
-                        0.24, 0.18, 0.13, 0.09, 0.05, 0.03, 0.01]
-
-    plugging_in_weekday = [0.017, 0.006, 0.004, 0.003, 0.004, 0.005, 0.01, 0.019, 0.03, 0.034, 0.036, 0.042, 0.043,
[…]
