### 2024-09-19 18d56be
MaJ EBike model
