### 2024-09-19 02a33da
cleaning weather, add clustering test
