+++ reho/model/preprocessing/weather.py
@@ -14,51 +14,75 @@ Generates the meteorological data (temperature and solar irradiance).
 """
 
 
-def get_cluster_file_ID(cluster):
+def get_weather_data(qbuildings_data, export_filename=None):
     """
-    Gets the weather file ID that corresponds to the specifications provided in the reho initalization.
-
-    The file ID is built by concatenating Location_Periods_PeriodDuration_Attributes.
-    ``cluster = {'Location': 'Geneva', 'Attributes': ['T', 'I', 'W'], 'Periods': 10, 'PeriodDuration': 24}``
-    Will yield to:
-    ``File_ID = 'Geneva_10_24_T_I_W'``
-
-    Parameters
-    ----------
-    cluster : dict
-        Contains a 'Location' (str), some 'Attributes' (list, among 'T' (temperature), 'I' (irradiance), 'W' (weekday) and 'E' (emissions)), a number of periods 'Periods' (int) and a 'PeriodDuration' (int).
-
-    Returns
-    -------
-    str
-        A literal representation to identify the location and clustering attritutes.
+    Using the pvlib library, connects to the PVGIS dabatase to extract the weather data based on the building's coordinates.
     """
-    if 'T' in cluster['Attributes']:
-        T = '_T'
-    else:
-        T = ''
-    if 'I' in cluster['Attributes']:
-        I = '_I'
-    else:
-        I = ''
-    if 'W' in cluster['Attributes']:
-        W = '_W'
-    else:
-        W = ''
-    if 'E' in cluster['Attributes']:
-        E = '_E'
-    else:
-        E = ''
+    lat, long = Transformer.from_crs("EPSG:2056", "EPSG:4326").transform(qbuildings_data['buildings_data']['Building1']['x'],
+                                                                         qbuildings_data['buildings_data']['Building1']['y'])
 
-    File_ID = cluster['Location'] + '_' + str(cluster['Periods']) + '_' + str(cluster['PeriodDuration']) + T + I + W + E
+    pvgis_data = pvlib.iotools.get_pvgis_tmy(lat, long, outputformat='csv', startyear=2005, endyear=2016)
+    coordinates = pvgis_data[2]
+    weather_data = pvgis_data[0]
 
-    return File_ID
+    # Rename columns
+    weather_data = weather_data.rename(columns={'temp_air': 'Text', 'ghi': 'Irr'})
+
+    # Extract month, day, and hour
+    weather_data['Month'] = weather_data.index.month
+    weather_data['Day'] = weather_data.index.day
+    weather_data['Hour'] = weather_data.index.hour + 1
+
+    # Add a 'Weekday' column where 1 is a weekday and 0 is a weekend
+    weather_data['Weekday'] = weather_data.index.weekday < 5  # Monday-Friday are weekdays
+    weather_data['Weekday'] = weather_data['Weekday'].astype(int)  # Convert Boolean to int
+
+    # Add unique identifier column 'id'
+    weather_data['id'] = (weather_data.reset_index().index + 1).to_list()
+
+    print(f'The weather data have been loaded from the PVGIS database for a location with coordinates {coordinates}.')
+
+    if export_filename is not None:
+        weather_data = weather_data[['id', 'Month', 'Day', 'Hour', 'Irr', 'Text', 'Weekday']]
+        weather_data.to_csv(export_filename)
+
+    return weather_data
+
+
+def read_custom_weather(path_to_weather_file):
+    """
+    From the current directory, looks for a custom weather file.
+    This file should be a .csv with the same structure as the examples provided in ``reho/scripts/examples/data/profiles/``.
+    """
+
[…]
