+++ scripts/examples/3f_Custom_profiles.py
@@ -8,8 +8,9 @@ if __name__ == '__main__':
     reader.establish_connection('Geneva')
     qbuildings_data = reader.read_db(transformer=234, egid=['1017073/1017074', '1017109', '1017079', '1030377/1030380'])
 
-    # Select clustering options for weather data
-    cluster = {'Location': 'Geneva', 'Attributes': ['T', 'I', 'W'], 'Periods': 10, 'PeriodDuration': 24}
+    # Select clustering options for weather data, with custom profiles
+    custom_weather_profile = str(Path(__file__).parent / 'data' / 'profiles' / 'Geneva.csv')
+    cluster = {'custom_weather': custom_weather_profile, 'Location': 'Geneva', 'Attributes': ['T', 'I', 'W'], 'Periods': 10, 'PeriodDuration': 24}
 
     # Set scenario
     scenario = dict()
