+++ documentation/sections/4_Package_structure.rst
@@ -45,7 +45,6 @@ Directory for data-related files.
 - ``QBuildings/``
 - ``SIA/``
 - ``skydome/``
-- ``weather/``
 
 **model/**
 ==================
