### 2024-09-19 e8cf19b
fix for pyclustering, specify numpy version
(cherry picked from commit 187a42ea996d562eec77bfabc27ddc263ad7f9af)
