+++ requirements.txt
@@ -6,7 +6,7 @@ coloredlogs
 geopandas<1.0.0
 kaleido
 matplotlib
-numpy
+numpy<2.0.0
 openpyxl
 pandas<2.0.0
 plotly
