### 2024-09-19 187a42e
fix for pyclustering, specify numpy version
