### 2024-09-19 0fc04d3
fix for pyclustering, specify numpy version
