+++ reho/model/infrastructure.py
@@ -464,8 +464,6 @@ def initialize_units(scenario, grids=None, building_data=os.path.join(path_to_in
     """
 
     default_units_to_exclude = ['HeatPump_Anergy', 'HeatPump_Lake']
-    if "Mobility" not in grids.keys():
-        default_units_to_exclude = default_units_to_exclude + ['EVcharging_district']
     if "exclude_units" not in scenario:
         exclude_units = default_units_to_exclude
     else:
