### 2024-09-25 4beabb4
not ignoring template
