### 2024-09-26 20b0a9f
nomenclature debug + renamed EV_profile_generator.py to mobility_generator.py
