+++ reho/model/sub_problem.py
@@ -3,7 +3,7 @@ import logging
 
 from amplpy import AMPL, Environment
 
-import reho.model.preprocessing.EV_profile_generator as EV_gen
+import reho.model.preprocessing.mobility_generator as EV_gen
 import reho.model.preprocessing.buildings_profiles as buildings_profiles
 import reho.model.preprocessing.emissions_parser as emissions
 import reho.model.preprocessing.weather as weather
@@ -293,7 +293,7 @@ class SubProblem:
             if len(self.infrastructure_sp.UnitsOfDistrict) != 0:
                 if "EV_district" in self.infrastructure_sp.UnitsOfDistrict:
                     p = EV_gen.generate_mobility_parameters(self.cluster_sp,self.parameters_sp,
-                                                            np.append(self.infrastructure_sp.UnitsOfLayer["Mobility"],'Public_transport'))
+                                                            np.setdiff1d(np.append(self.infrastructure_sp.UnitsOfLayer["Mobility"],'Public_transport'),self.infrastructure_sp.UnitsOfType['EV_charger']))
                     self.parameters_to_ampl.update(p)
 
     def set_HP_parameters(self, ampl):
