+++ documentation/sections/4_Package_structure.rst
@@ -123,10 +123,10 @@ Core of the optimization model (model objectives, constraints, modelling equatio
 .. automodule:: reho.model.preprocessing.emissions_parser
     :members:
 
-`EV_profile_generator.py`
+`mobility_generator.py`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-.. automodule:: reho.model.preprocessing.EV_profile_generator
+.. automodule:: reho.model.preprocessing.mobility_generator
     :members:
 
 `local_data.py`
