+++ scripts/examples/6b_mobility_externaldistricts.py
@@ -1,4 +1,5 @@
 from reho.model.reho import *
+from reho.model.preprocessing.mobility_generator import *
 
 
 if __name__ == '__main__':
@@ -19,12 +20,16 @@ if __name__ == '__main__':
     scenario['exclude_units'] = ['Battery', 'NG_Cogeneration']
     scenario['enforce_units'] = ['EV_district']
 
+    ext_districts = [8538 ,13569]
+    df_rho = pd.DataFrame(columns=['work','travel','leisure'],index=ext_districts).fillna(0.5)
+
     parameters = {  "Population": 9,
-                    "max_share_cars": 0.7,
-                    "min_share_cars": 0.4,
+                    "max_share_EV": 0.7,
+                    "min_share_EV": 0.4,
+                    "share_activity": rho_param(ext_districts,df_rho)
                 }
     
-    set_indexed = {"Districts": [8538 ,13569]}
+    set_indexed = {"Districts": ext_districts }
 
     # Load parameters representing external district demands and supply
     with open(f'data/mobility/6b_extdistrict_parameters.pickle', 'rb') as handle:
@@ -45,7 +50,7 @@ if __name__ == '__main__':
               'external_district' : True
               }
     # Run optimization
-    reho = reho(qbuildings_data=qbuildings_data, units=units, grids=grids,parameters=parameters, cluster=cluster,set_indexed=set_indexed, scenario=scenario, method=method, solver="gurobiasl")
+    reho = REHO(qbuildings_data=qbuildings_data, units=units, grids=grids,parameters=parameters, cluster=cluster,set_indexed=set_indexed, scenario=scenario, method=method, solver="gurobiasl")
     reho.single_optimization()
 
     # Save results
