+++ reho/model/preprocessing/mobility_generator.py
@@ -300,6 +300,7 @@ def rho_param(ext_districts,rho,activities = ["work","leisure","travel"]):
     This function is used to calculate the parameter S from the share of activities S(a) in each districts.
     For each activity, if a distribution is provided by the parameter rho, then S = rho_d / sum over ext_d(rho)
     Otherwise, we assume equal distribution over the districts and S = 1/len(nb_ext_d)
+    
     Parameters
     ----------
     ext_districts : 
@@ -312,53 +313,85 @@ def rho_param(ext_districts,rho,activities = ["work","leisure","travel"]):
 
     """
     share = pd.DataFrame(index=ext_districts,columns=activities).fillna(1/len(ext_districts))
-    for act in rho.columns:
-        share[act] = rho[act] / rho[act][rho.index.isin(ext_districts)].sum()
+    for act in share.columns:
+        if act in rho.columns:
+            share[act] = rho[act] / rho[act][rho.index.isin(ext_districts)].sum()
     share = share.stack().to_frame(name= "share_activity").reorder_levels([1,0])
     share.loc['travel'] = 0 # additionnal precaution
     return share
 
 
-def compute_iterative_parameters(variables,parameters,district_parameters,only_pi = False):
+def compute_iterative_parameters(variables,district_parameters,only_prices = False):
     """"
     This function is used in the iterative scenario to iteratively calculate multiple districts with EVs being able to charge at the different districts.
-    The load is expressed using the corrective parameter f. 
+    The load is expressed using the corrective parameter f.
+
+    Parameters
+    ----------
+    variables : dict of dict
+        Each key of the dict refers to a district and should contain variables extracted from the previous optimisation, namely pi, df_Grid_t and df_Unit_t.
+    district_parameters : dict of dict
+        Each key of the dict refers to a district d. Used to extract the scale parameter f : district_parameters[d]['f']
+    only_prices : bool
+        if False, only returns the parameters Cost_demand_ext and Cost_supply_ext
+        if True, additionally returns the parameter EV_charger_supply_ext
+    Returns
+    -------
+    parameters : dict of dict
+        For each district d, returns a dict of the parameters to be inputted in the next optimisation. Parameters include Cost_demand_ext, Cost_supply_ext, EV_charger_supply_ext. 
 
     """
-    if only_pi:
-        df_prices = pd.DataFrame()
-        for d in variables.keys():
-            price = variables[d]['pi'].to_frame().copy()
-            price['district'] = d
-            price = price.set_index('district',append = True).reorder_levels([2,0,1])
-            df_prices = pd.concat([df_prices,price])
 
-        for d in variables.keys():
-            parameters[d] = {   "Cost_demand_ext"    : df_prices[df_prices.index.get_level_values(level="district") != d].rename(columns = {'pi' : 'Cost_demand_ext'}),
-                                "Cost_supply_ext" : variables[d]['pi'].rename("Cost_supply_ext")
-                                }
+    df_prices = pd.DataFrame()
+    for d in variables.keys():
+        price = variables[d]['pi'].to_frame().copy()
+        price['district'] = d
+        price = price.set_index('district',append = True).reorder_levels([2,0,1])
+        df_prices = pd.concat([df_prices,price])
+        df_prices = df_prices[df_prices.index.get_level_values("Period").isin(range(0,11))] # TODO : link this with Period and PeriodStandard
 
-    else:
+    parameters = dict()
+    for d in variables.keys():
[…]
