### 2024-10-02 IPESE/REHO#32 @willu47
[Bug]: JOSS REVIEW psycopg2 via pip install
### Is there an existing issue for this?

- [X] I have searched the existing issues

### What happened?

Installing v1.1.4 from the documented `pip install --extra-index-url https://pypi.ampl.com REHO` results in an error installing pyschopg2 due to missing postgres dependency.

I'm running OSX with an M3 chip.

### Version

1.1.4

### Relevant log output

```shell
pip install --extra-index-url https://pypi.ampl.com REHO
Looking in indexes: https://pypi.org/simple, https://pypi.ampl.com
Collecting REHO
  Downloading REHO-1.1.4-py3-none-any.whl.metadata (3.9 kB)
Collecting amplpy (from REHO)
  Downloading amplpy-0.14.0-cp312-cp312-macosx_10_9_x86_64.whl.metadata (4.0 kB)
Collecting ampl-module-base (from REHO)
  Downloading https://portal.ampl.com/dl/pypi/ampl-module-base/ampl_module_base-20240812-py3-none-macosx_10_9_x86_64.whl (5.9 MB)
Collecting ampl-module-highs (from REHO)
  Downloading https://portal.ampl.com/dl/pypi/ampl-module-highs/ampl_module_highs-20240724-py3-none-macosx_10_9_x86_64.whl (5.5 MB)
Collecting coloredlogs (from REHO)
  Downloading coloredlogs-15.0.1-py2.py3-none-any.whl.metadata (12 kB)
Collecting geopandas<1.0.0 (from REHO)
  Downloading geopandas-0.14.4-py3-none-any.whl.metadata (1.5 kB)
Collecting kaleido (from REHO)
  Downloading kaleido-0.2.1-py2.py3-none-macosx_10_11_x86_64.whl.metadata (15 kB)
Collecting matplotlib (from REHO)
  Using cached matplotlib-3.9.2-cp312-cp312-macosx_10_12_x86_64.whl.metadata (11 kB)
Collecting numpy<2.0.0 (from REHO)
  Using cached numpy-1.26.4-cp312-cp312-macosx_10_9_x86_64.whl.metadata (61 kB)
Collecting openpyxl (from REHO)
  Using cached openpyxl-3.1.5-py2.py3-none-any.whl.metadata (2.5 kB)
Collecting pandas<2.0.0 (from REHO)
  Downloading pandas-1.5.3.tar.gz (5.2 MB)
  Installing build dependencies ... done
  Preparing metadata (pyproject.toml) ... done
Collecting plotly (from REHO)
  Downloading plotly-5.24.1-py3-none-any.whl.metadata (7.3 kB)
Collecting pvlib (from REHO)
  Downloading pvlib-0.11.1-py3-none-any.whl.metadata (2.8 kB)
Collecting pyclustering (from REHO)
  Downloading pyclustering-0.10.1.2.tar.gz (2.6 MB)
  Preparing metadata (setup.py) ... done
Collecting pyproj (from REHO)
  Downloading pyproj-3.7.0-cp312-cp312-macosx_12_0_x86_64.whl.metadata (31 kB)
Collecting python-dotenv (from REHO)
  Downloading python_dotenv-1.0.1-py3-none-any.whl.metadata (23 kB)
Collecting pytest (from REHO)
  Downloading pytest-8.3.3-py3-none-any.whl.metadata (7.5 kB)
Collecting qmcpy (from REHO)
  Downloading qmcpy-1.5.tar.gz (7.5 MB)
  Preparing metadata (setup.py) ... done
Collecting requests (from REHO)
  Using cached requests-2.32.3-py3-none-any.whl.metadata (4.6 kB)
Collecting SALib (from REHO)
  Downloading salib-1.5.1-py3-none-any.whl.metadata (11 kB)
Collecting scipy (from REHO)
  Using cached scipy-1.14.1-cp312-cp312-macosx_14_0_x86_64.whl.metadata (60 kB)
Requirement already satisfied: setuptools in /usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages (from REHO) (75.1.0)
Collecting shapely (from REHO)
  Downloading shapely-2.0.6-cp312-cp312-macosx_10_9_x86_64.whl.metadata (7.0 kB)
Collecting sqlalchemy<2.0.0 (from REHO)
  Downloading SQLAlchemy-1.4.54-cp312-cp312-macosx_10_9_universal2.whl.metadata (10 kB)
Collecting urllib3 (from REHO)
  Downloading urllib3-2.2.3-py3-none-any.whl.metadata (6.5 kB)
Collecting psycopg2<3.0.0 (from REHO)
  Downloading psycopg2-2.9.9.tar.gz (384 kB)
  Preparing metadata (setup.py) ... error
  error: subprocess-exited-with-error

  × python setup.py egg_info did not run successfully.
  │ exit code: 1
  ╰─> [23 lines of output]
      running egg_info
      creating /private/var/folders/_7/bccbfn110b501smf76mty7xh0000gp/T/pip-pip-egg-info-lhjbyb35/psycopg2.egg-info
      writing /private/var/folders/_7/bccbfn110b501smf76mty7xh0000gp/T/pip-pip-egg-info-lhjbyb35/psycopg2.egg-info/PKG-INFO
      writing dependency_links to /private/var/folders/_7/bccbfn110b501smf76mty7xh0000gp/T/pip-pip-egg-info-lhjbyb35/psycopg2.egg-info/dependency_links.txt
      writing top-level names to /private/var/folders/_7/bccbfn110b501smf76mty7xh0000gp/T/pip-pip-egg-info-lhjbyb35/psycopg2.egg-info/top_level.txt
      writing manifest file '/private/var/folders/_7/bccbfn110b501smf76mty7xh0000gp/T/pip-pip-egg-info-lhjbyb35/psycopg2.egg-info/SOURCES.txt'

      Error: pg_config executable not found.

      pg_config is required to build psycopg2 from source.  Please add the directory
      containing pg_config to the $PATH or specify the full executable path with the
      option:

          python setup.py build_ext --pg-config /path/to/pg_config build ...

      or with the pg_config option in 'setup.cfg'.

      If you prefer to avoid building psycopg2 from source, please install the PyPI
      'psycopg2-binary' package instead.

      For further information please check the 'doc/src/install.rst' file (also at
      <https://www.psycopg.org/docs/install.html>).

      [end of output]

error: metadata-generation-failed

× Encountered error while generating package metadata.
╰─> See above for output.

note: This is an issue with the package mentioned above, not pip.
hint: See above for details.
```

### Anything else?

This re-opens #20 with a twist!
