### 2024-10-02 IPESE/REHO#32 @willu47
`brew install postresql` solved the issue.  I believe this is really a problem with the documentation.  The 'part time user' could add a note first to install postgres.
