### 2024-10-02 IPESE/REHO#34 @willu47
[Bug]: JOSS REVIEW Running test script with AMPL cloud license results in recursion error
### Is there an existing issue for this?

- [x] I have searched the existing issues

### What happened?

I've install v1.1.4 using pip, installed postgres via brew, and installed AMPL manually via the website (OSX version) and activated the cloud license.
I have a `.env` file in my folder, although the .env file is empty - as I do not have a license file. AMPL provide a cloud license.

Either the script does not recognise my local folder, and the presence of the `.env` file, or I have the wrong value in my .env file. Either way, the execution should stop on being unable to run AMPL (if that is the issue) and not produce this very long error trace.

### Version

1.0.2 (Stable)

### Relevant log output

```shell
% reho-test-import
AMPL_PATH is not defined. Please include a .env file at the project root (e.g., AMPL_PATH='C:/AMPL')
Traceback (most recent call last):
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/bin/reho-test-import", line 8, in <module>
    sys.exit(test_import_reho_modules())
             ^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/reho/test/test_import.py", line 19, in test_import_reho_modules
    __import__(module)
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/reho/model/__init__.py", line 1, in <module>
    from . import reho
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/reho/model/reho.py", line 4, in <module>
    from reho.model.master_problem import *
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/reho/model/master_problem.py", line 11, in <module>
    from reho.model.preprocessing.local_data import *
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/reho/model/preprocessing/local_data.py", line 2, in <module>
    import reho.model.preprocessing.weather as weather
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/reho/model/preprocessing/weather.py", line 9, in <module>
    import pvlib
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/pvlib/__init__.py", line 3, in <module>
    from pvlib import (  # noqa: F401
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/pvlib/spectrum/__init__.py", line 15, in <module>
    from pvlib.spectrum.response import (  # noqa: F401
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/pvlib/spectrum/response.py", line 9, in <module>
    from scipy.interpolate import interp1d
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/scipy/interpolate/__init__.py", line 167, in <module>
    from ._interpolate import *
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/scipy/interpolate/_interpolate.py", line 12, in <module>
    from . import _fitpack_py
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/scipy/interpolate/_fitpack_py.py", line 8, in <module>
    from ._fitpack_impl import bisplrep, bisplev, dblint  # noqa: F401
    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/scipy/interpolate/_fitpack_impl.py", line 103, in <module>
    'iwrk': array([], dfitpack_int), 'u': array([], float),
            ^^^^^^^^^^^^^^^^^^^^^^^

    elif np.issubdtype(dtype, np.number):
         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/numerictypes.py", line 417, in issubdtype
    arg1 = dtype(arg1).type
           ^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 46, in __repr__
    arg_str = _construction_repr(dtype, include_align=False)
              ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 100, in _construction_repr
    return _scalar_str(dtype, short=short)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 143, in _scalar_str
    elif np.issubdtype(dtype, np.number):
         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/numerictypes.py", line 417, in issubdtype
    arg1 = dtype(arg1).type
           ^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 46, in __repr__
    arg_str = _construction_repr(dtype, include_align=False)
              ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 100, in _construction_repr
    return _scalar_str(dtype, short=short)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 143, in _scalar_str
    elif np.issubdtype(dtype, np.number):
         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/numerictypes.py", line 417, in issubdtype
    arg1 = dtype(arg1).type
           ^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 46, in __repr__
    arg_str = _construction_repr(dtype, include_align=False)
              ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 100, in _construction_repr
    return _scalar_str(dtype, short=short)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 143, in _scalar_str
    elif np.issubdtype(dtype, np.number):
         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/numerictypes.py", line 417, in issubdtype
    arg1 = dtype(arg1).type
           ^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 46, in __repr__
    arg_str = _construction_repr(dtype, include_align=False)
              ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 100, in _construction_repr
    return _scalar_str(dtype, short=short)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 143, in _scalar_str
    elif np.issubdtype(dtype, np.number):
         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/numerictypes.py", line 417, in issubdtype
    arg1 = dtype(arg1).type
           ^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 46, in __repr__
    arg_str = _construction_repr(dtype, include_align=False)
              ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 100, in _construction_repr
    return _scalar_str(dtype, short=short)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 143, in _scalar_str
    elif np.issubdtype(dtype, np.number):
         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/numerictypes.py", line 417, in issubdtype
    arg1 = dtype(arg1).type
           ^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 46, in __repr__
    arg_str = _construction_repr(dtype, include_align=False)
              ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 100, in _construction_repr
    return _scalar_str(dtype, short=short)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 143, in _scalar_str
    elif np.issubdtype(dtype, np.number):
         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/numerictypes.py", line 417, in issubdtype
    arg1 = dtype(arg1).type
           ^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 46, in __repr__
    arg_str = _construction_repr(dtype, include_align=False)
              ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 100, in _construction_repr
    return _scalar_str(dtype, short=short)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 143, in _scalar_str
    elif np.issubdtype(dtype, np.number):
         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/numerictypes.py", line 417, in issubdtype
    arg1 = dtype(arg1).type
           ^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 46, in __repr__
    arg_str = _construction_repr(dtype, include_align=False)
              ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 100, in _construction_repr
    return _scalar_str(dtype, short=short)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 143, in _scalar_str
    elif np.issubdtype(dtype, np.number):
         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/numerictypes.py", line 417, in issubdtype
    arg1 = dtype(arg1).type
           ^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 46, in __repr__
    arg_str = _construction_repr(dtype, include_align=False)
              ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 100, in _construction_repr
    return _scalar_str(dtype, short=short)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 143, in _scalar_str
    elif np.issubdtype(dtype, np.number):
         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/numerictypes.py", line 417, in issubdtype
    arg1 = dtype(arg1).type
           ^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 46, in __repr__
    arg_str = _construction_repr(dtype, include_align=False)
              ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 100, in _construction_repr
    return _scalar_str(dtype, short=short)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 143, in _scalar_str
    elif np.issubdtype(dtype, np.number):
         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/numerictypes.py", line 417, in issubdtype
    arg1 = dtype(arg1).type
           ^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 46, in __repr__
    arg_str = _construction_repr(dtype, include_align=False)
              ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 100, in _construction_repr
    return _scalar_str(dtype, short=short)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 143, in _scalar_str
    elif np.issubdtype(dtype, np.number):
         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/numerictypes.py", line 417, in issubdtype
    arg1 = dtype(arg1).type
           ^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 46, in __repr__
    arg_str = _construction_repr(dtype, include_align=False)
              ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 100, in _construction_repr
    return _scalar_str(dtype, short=short)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 143, in _scalar_str
    elif np.issubdtype(dtype, np.number):
         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/numerictypes.py", line 417, in issubdtype
    arg1 = dtype(arg1).type
           ^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 46, in __repr__
    arg_str = _construction_repr(dtype, include_align=False)
              ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 100, in _construction_repr
    return _scalar_str(dtype, short=short)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 143, in _scalar_str
    elif np.issubdtype(dtype, np.number):
         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/numerictypes.py", line 417, in issubdtype
    arg1 = dtype(arg1).type
           ^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 46, in __repr__
    arg_str = _construction_repr(dtype, include_align=False)
              ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 100, in _construction_repr
    return _scalar_str(dtype, short=short)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 143, in _scalar_str
    elif np.issubdtype(dtype, np.number):
         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/numerictypes.py", line 417, in issubdtype
    arg1 = dtype(arg1).type
           ^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 46, in __repr__
    arg_str = _construction_repr(dtype, include_align=False)
              ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 100, in _construction_repr
    return _scalar_str(dtype, short=short)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 143, in _scalar_str
    elif np.issubdtype(dtype, np.number):
         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/numerictypes.py", line 417, in issubdtype
    arg1 = dtype(arg1).type
           ^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 46, in __repr__
    arg_str = _construction_repr(dtype, include_align=False)
              ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 100, in _construction_repr
    return _scalar_str(dtype, short=short)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 143, in _scalar_str
    elif np.issubdtype(dtype, np.number):
         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/numerictypes.py", line 417, in issubdtype
    arg1 = dtype(arg1).type
           ^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 46, in __repr__
    arg_str = _construction_repr(dtype, include_align=False)
              ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 100, in _construction_repr
    return _scalar_str(dtype, short=short)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 143, in _scalar_str
    elif np.issubdtype(dtype, np.number):
         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/numerictypes.py", line 417, in issubdtype
    arg1 = dtype(arg1).type
           ^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 46, in __repr__
    arg_str = _construction_repr(dtype, include_align=False)
              ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 100, in _construction_repr
    return _scalar_str(dtype, short=short)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/_dtype.py", line 143, in _scalar_str
    elif np.issubdtype(dtype, np.number):
         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/Caskroom/miniforge/base/envs/reho_review/lib/python3.12/site-packages/numpy/core/numerictypes.py", line 417, in issubdtype
    arg1 = dtype(arg1).type
           ^^^^^^^^^^^
[…]
