### 2024-10-02 IPESE/REHO#35 @willu47
[Feature]: JOSS REVIEW Editable install deprecation
### Summary

I cloned and performed an editable install using pip:
```bash
git clone https://github.com/IPESE/REHO.git
cd REHO
pip install -e . --extra-index-url https://pypi.ampl.com
```
and received the following warning:
```python
DEPRECATION: Legacy editable install of REHO==1.1.4 from file:///Users/wusher_1/repository/REHO (setup.py develop) is deprecated. pip 25.0 will enforce this behaviour change. A possible replacement is to add a pyproject.toml or enable --use-pep517, and use setuptools >= 64. If the resulting installation is not behaving as expected, try using --config-settings editable_mode=compat. Please consult the setuptools documentation for more information. Discussion can be found at https://github.com/pypa/pip/issues/11457
```

### Detailed design

Your design

### Notes and comments

Your notes
