### 2024-10-02 1847189
solved error : "Cannot invoke gurobi: no such program."
