### 2024-10-07 IPESE/REHO#36 @willu47
JOSS REVIEW: Paper
I spotted a few typos in the [paper](https://github.com/openjournals/joss-papers/blob/joss.06734/joss.06734/10.21105.joss.06734.pdf):

- line 17 add "**The** documentation website"
- line 26 add "**The** intensification of..."
