### 2024-10-07 #6734 @willu47
I've run into several issues again installing and running the REHO package on my Macbook Pro (M3 Pro chip).

- [ ] Issue IPESE/REHO#32 flags the issue with psycopg2.  I'd suggest clearly stating in the installation section of the docs that users must first install PostgreSQL rather than stating its as simple is `pip install`. Also link to the section in the docs which provides support for this specific issue.
- [ ] Issue IPESE/REHO#34 I'm unable to get REHO to recognise my cloud licensed AMPL installation.  Running AMPL via Python seems to work fine, but the check REHO makes for the presence of the .env file seems to create an issue.
- [ ] Issue IPESE/REHO#35 The packaging adopted using setup.py is now outdated - consider using a .toml file instead.
- [ ] Issue IPESE/REHO#36 lists very minor issues in the paper
