+++ scripts/examples/6a_Actors_problem.py
@@ -12,7 +12,7 @@ if __name__ == '__main__':
     # Set building parameters
     reader = QBuildingsReader()
     reader.establish_connection('Geneva')
-    qbuildings_data = reader.read_db(234, nb_buildings=2)
+    qbuildings_data = reader.read_db(district_id=234, nb_buildings=2)
 
     # Set specific parameters
     parameters = {}
