+++ scripts/examples/0_Compact_formulation.py
@@ -7,7 +7,7 @@ if __name__ == '__main__':
     # Set building parameters
     reader = QBuildingsReader()  # load QBuildingsReader class
     reader.establish_connection('Geneva')  # connect to QBuildings database
-    qbuildings_data = reader.read_db(5, egid=['2034144/2034143/2749579/2034146/2034145'])  # read data
+    qbuildings_data = reader.read_db(district_id=5, egid=['2034144/2034143/2749579/2034146/2034145'])  # read data
 
     # Select clustering options for weather data
     #  - I refers to Irradiance, T to Temperature, and W to Weekday
