### 2024-10-09 82b76a4
documentation (local vs cloud license)
