+++ joss/paper.md
@@ -32,14 +32,14 @@ bibliography: paper.bib
 
 The transition to sustainable energy systems in the face of growing renewable energy adoption and electrification is a complex and critical challenge. The _Renewable Energy Hub Optimizer_ (REHO) emerges as a powerful decision support tool designed to investigate the deployment of energy conversion and storage technologies in this evolving landscape. REHO leverages a Mixed-Integer Linear Programming (MILP) framework combined with a Dantzig-Wolfe decomposition algorithm to simultaneously address the optimal design and operation of district energy systems, catering to multi-objective considerations across economic, environmental, and efficiency criteria.
 
-REHO is deployed as an open-source and collaborative Python library, available as a [PyPI package](https://pypi.org/project/REHO/) and supported by comprehensive [documentation](https://reho.readthedocs.io/en/main/). Documentation website includes step-by-step instructions, details about the mathematical background and model foundations, as well as a list of academic publications, conference proceedings, research projects, and other works related to REHO.
+REHO is deployed as an open-source and collaborative Python library, available as a [PyPI package](https://pypi.org/project/REHO/) and supported by comprehensive [documentation](https://reho.readthedocs.io/en/main/). The documentation website includes step-by-step instructions, details about the mathematical background and model foundations, as well as a list of academic publications, conference proceedings, research projects, and other works related to REHO.
 
 This paper introduces REHO and highlights its key features and contributions to the field of sustainable energy system planning.
 
 
 # Statement of need
 
-Cities around the world are moving towards increasing the penetration of local energy harvesting and storage capacities to render their energy consumption more sustainable and less dependent on a geopolitical context. Intensification of renewables deployment is witnessed in the past decade and keeps continuing, leading to important techno-economic-social trade-offs in energy strategy. This transition blurs the boundaries between demand and supply and creates new types of stakeholders. Adopting a district-level approach for energy system planning seems thus particularly relevant, as it promotes the valorization of endogenous resources and enables economies of scale while preserving local governance [@heldewegRenewableEnergyCommunities2020]. The emergence of the concept of energy communities is a clear example of this growing interest for energy planning at the neighborhood scale [@dociExploringTransitionPotential2014]. Energy communities are expected to play a pivotal role in the ongoing energy transition by fostering decentralized, sustainable, and community-driven approaches to energy production and consumption. Through the collaborative efforts of residents, utilities, and institutions, energy communities offer a techno-economic framework to support the paradigm shift from centralized to distributed and district-level energy systems [@caramizaruEnergyCommunitiesOverview2020].
+Cities around the world are moving towards increasing the penetration of local energy harvesting and storage capacities to render their energy consumption more sustainable and less dependent on a geopolitical context. The intensification of renewables deployment is witnessed in the past decade and keeps continuing, leading to important techno-economic-social trade-offs in energy strategy. This transition blurs the boundaries between demand and supply and creates new types of stakeholders. Adopting a district-level approach for energy system planning seems thus particularly relevant, as it promotes the valorization of endogenous resources and enables economies of scale while preserving local governance [@heldewegRenewableEnergyCommunities2020]. The emergence of the concept of energy communities is a clear example of this growing interest for energy planning at the neighborhood scale [@dociExploringTransitionPotential2014]. Energy communities are expected to play a pivotal role in the ongoing energy transition by fostering decentralized, sustainable, and community-driven approaches to energy production and consumption. Through the collaborative efforts of residents, utilities, and institutions, energy communities offer a techno-economic framework to support the paradigm shift from centralized to distributed and district-level energy systems [@caramizaruEnergyCommunitiesOverview2020].
 
[…]
