+++ .github/workflows/draft-pdf.yml
@@ -1,7 +1,7 @@
 on:
   push:
-    branches:
-      - documentation
+    tags:
+      - '*JOSS*'  # Matches any tag that contains "JOSS"
 
 jobs:
   paper:
@@ -17,7 +17,7 @@ jobs:
           # This should be the path to the paper within your repo.
           paper-path: joss/paper.md
       - name: Upload
-        uses: actions/upload-artifact@v1
+        uses: actions/upload-artifact@v4
         with:
           name: paper
           # This is the output path where Pandoc will write the compiled PDF.

[…]
