### 2024-10-09 IPESE/REHO#36 @DorsanL
Thank you, it has been corrected (https://github.com/IPESE/REHO/commit/82b76a4f09038fc0a27d8d9ba4af41ad9c189edb)

The updated paper can be downloaded here: https://github.com/IPESE/REHO/actions/runs/11259187496/artifacts/2036040498
