### 2024-10-09 IPESE/REHO#32 @DorsanL
Indeed, thank you for this accurate feedback.

The "Getting started" section has been adapted to reflect the PostgreSQL dependency. This issue and the suggested fixes are now stated before the installation from PyPI or from source: https://reho.readthedocs.io/en/main/sections/5_Getting_started.html#postgresql
