### 2024-10-09 IPESE/REHO#34 @DorsanL
Thank you for reporting this problem.

The relevant parts of the code have been adapted: an .env file is now only required when using a standard local license (its path has to be specified as an environment variable). For the Community Edition cloud-based license, the user simply has to activate it before running his first optimization. (https://github.com/IPESE/REHO/commit/82b76a4f09038fc0a27d8d9ba4af41ad9c189edb)

The "AMPL license" paragraph of "Getting started" section has been redesigned to clarify these instructions: https://reho.readthedocs.io/en/main/sections/5_Getting_started.html#ampl-license
