### 2024-10-09 #6734 @DorsanL
Dear @willu47,
Thank you for these accurate and pertinent findings. I answered each of them, implementing the proposed fixes. I hope you will be able to launch the optimizations and test the tool's functionality.

I'll keep an eye out for any further comments.
