### 2024-10-10 IPESE/REHO#34 @willu47
Upon testing, while the recursion error has gone, the fundamental issue of not recognising that the `.env` file is in my current working folder (so relative to where I call the test script from) and not in a location relative to where the REHO package is kept, is the main issue here.  See note in #13 again.
