### 2024-10-10 IPESE/REHO#32 @willu47
Could you clarify if postgreSQL is an optional install, or is it a required dependency for REHO?  If it is optional, it may be worth excluding it from the normal install of REHO.  For example `pip install reho` or `pip install reho[postgresql]` (see [here](https://packaging.python.org/en/latest/guides/writing-pyproject-toml/#dependencies-optional-dependencies) for how to do that).
