### 2024-10-10 c843655
Demand by categories of distances  +working on sascha data
