### 2024-10-11 b0ba53c
load .env from root or from cwd
