### 2024-10-11 IPESE/REHO#34 @DorsanL
NB: For your convenience, I deployed a [REHO 1.1.5](https://pypi.org/project/REHO/) in PyPI. I hope it'll be fine with these errors 🤞
