### 2024-10-14 IPESE/REHO#34 @willu47
I installed the latest version from Github using `pip install git+https://github.com/IPESE/REHO.git@main#egg=REHO`.  It is a good idea to deploy pre-release versions to PyPI, which is automatic with a tag such as `v1.5.2-alpha` (see [this](https://pythonpackaging.info/07-Package-Release.html#Versioning-your-code) for examples of valid tags).
