### 2024-10-14 #6734 @willu47
Hi. I have now completed my review. It took me a long time to get my AMPL license to be recognised by REHO, which caused delays in the review process. While we have made progress with REHO, and I think it is adequately documented and structured for a JOSS publication, AMPL could be substituted by an open-source package with no loss of functionality.
