### 2024-10-14 f850ffe
bug distance corrected + import parameters
