### 2024-10-19 #6734 @hgandhi2411
Hello, I have completed my review and agree with @willu47 about getting around AMPL license. Everything else looks good and works as intended.
