### 2024-10-22 bef3cc7
Interperiod, Hydrogen, rSOC, at building scale, with updated Sankey
