+++ reho/plotting/plotting.py
@@ -1476,3 +1476,183 @@ def plot_composite_curve(df_Results, cluster, periods=["Yearly"], filename=None,
         return plt, data.fillna(0)
     else:
         return plt
+
+def plot_storage_profile(df_Results, resolution='daily'):
+    if resolution == 'monthly':
+        items_average = 730
+    elif resolution == 'weekly':
+        items_average = 168
+    elif resolution == 'daily':
+        items_average = 24
+    else:
+        items_average = 1
+
+    SOC = df_Results["df_storage"]["BAT_E_stored_IP"]
+    time_index = np.arange(0, 8760)
+
+    SOC_average = moving_average(SOC, items_average)
+    time_index_average = moving_average(time_index, items_average)
+
+    fig = go.Figure()
+    fig.add_trace(go.Scatter(
+        x=time_index_average,
+        y=SOC_average,
+        mode='lines',
+        name='SOC',
+        line=dict(color='blue')
+    ))
+
+    fig.update_layout(
+        title='State of Charge Over Time',
+        xaxis_title='T',
+        yaxis_title='SOC (kWh)',
+        template='plotly',
+        showlegend=True
+    )
+
+    return fig
+
+
+def plot_energy_balance_for_battery(df_Results, units_to_plot, color='ColorPastel', day_of_the_year = '' ,time_range='week', label='EN_long'):
+    if time_range == 'week':
+        period = 7
+    elif time_range == 'month':
+        period = 30
+    elif time_range == '2weeks':
+        period = 14
+    elif time_range == '3days':
+        period = 3
+    else:
+        period = 1
+
+    starting_hour = ((day_of_the_year - 1) * 24) + 1
+
+    # selection of the period
+    ending_hour = starting_hour + period * 24
+
+    SOC = df_Results["df_storage"]["BAT_E_stored_IP"][starting_hour:ending_hour]
+
+    units_demand = []
+    units_supply = []
+    for unit in units_to_plot:
+        if unit == "BESS_IP":
+            continue
+        if unit == "PV":
+            units_supply.append(unit)
+        elif unit in ["Battery", "EV_district"]:
+            units_demand.append(unit)
+            units_supply.append(unit)
+        else:
+            units_demand.append(unit)
+
+    # Grids
+    imports = {}
+    exports = {}
+    layers = df_Results['df_Grid_t'].index.get_level_values("Layer").unique()
+    for layer in layers:
+        imports[layer] = df_Results['df_Grid_t'].xs((layer, 'Network'), level=('Layer', 'Hub')).Grid_supply[:-2]
+        exports[layer] = df_Results['df_Grid_t'].xs((layer, 'Network'), level=('Layer', 'Hub')).Grid_demand[:-2]
+
+    import_profile = {}
+    export_profile = {}
+    for layer in layers:
+        import_profile[layer] = np.array([])
+        export_profile[layer] = np.array([])
+        for i in range(day_of_the_year, day_of_the_year + period):
+            id = df_Results['df_Index'].PeriodOfYear[i * 24]
+            import_profile[layer] = np.concatenate((import_profile[layer], imports[layer].xs(id)))
+            export_profile[layer] = np.concatenate((export_profile[layer], exports[layer].xs(id)))
+
+    # Units
+    demands = dict()
+    supplies = dict()
+    for unit in units_to_plot:
+        df_aggregated = df_Results['df_Unit_t'][
+            df_Results['df_Unit_t'].index.get_level_values('Unit').str.contains(unit)]
+        if unit in units_demand:
+            demand = df_aggregated.droplevel('Layer').Units_demand[:-2].groupby(['Period', 'Time']).sum()
+            demands[unit] = np.array([])
[…]
