### 2024-10-22 #6734 @DorsanL
Thank you to the three of you for all the time you invested in this review for REHO!

I can understand your concerns about the non open-source nature of AMPL, and I've been thinking a lot over the last few days to stimulate a constructive exchange.

For sure, I understand the importance of open-source tools in fostering transparency and collaboration, especially in the context of an open-source Python package. However, I would like to offer several points for consideration in maintaining the current AMPL-based implementation of REHO.

Indeed, after a few attempts and explorations of other possibilities (I looked at several alternatives such as GNU MathProg and GLPK, or Python optimisation libraries such as Pyomo or PuLP), and here are my impressions:

1. Efficiency: AMPL is designed specifically for mathematical optimization, which allows for more intuitive modeling of complex mathematical structures. Its concise and expressive syntax simplifies the representation of large-scale optimization problems and complex constraint systems. In the case of REHO, the decomposition with the sub-problems and the master problem is thus readily implemented by taking advantages of some AMPL advanced features (typically data structures and indexing capabilities, not all available in GNU MathProg). Such implementation of the decomposition algorithm would require more complex workarounds in alternative framework (I gave a few tries with Pyomo, and it was quite complex).

2. Solver flexibility: AMPL in itself is « solver agnostic », meaning users can still opt for open-source solvers. A wide range of solvers are supported, and AMPL provides a highly optimized and efficient way to communicate with all of them. This includes both open-source and commercial solvers, ensuring flexibility in solver choice without compromising on the ability to scale or handle large problem instances efficiently. As a demonstration of this, all the REHO examples work fine with HiGHS or Gurobi.

3. Existing model complexity and development investment: As of now, the core optimization model is described using many .mod files, which define an extensive structure of constraints, variables, and objectives. Rewriting this model using a different framework, would require a massive amount of time and effort. Moreover, translating the model manually would introduce additional risk of errors during conversion, potentially impacting the reliability and accuracy of the final implementation.

4. Interfacing with input and output data: The current implementation makes use of the amplpy Python library for data input and output, which simplifies integration between Python and AMPL. This has enabled REHO to have a streamlined workflow for managing data directly with Python. Such high-level control of the optimization model directly with Python code, without needing to delve into AMPL code, is an important benefit compared to the past when the tool used to be solely based on .mod, .dat, and .run files.

5. Performance: While GLPK, Pyomo and other open-source optimization frameworks are powerful, their performance, especially in terms of speed and solver interaction, may not always match AMPL highly optimized backend (C-based implementation). This difference in performance may be significant for large-scale problems, where even small inefficiencies can have an outsized impact on computation time. REHO has already been applied to the entire Swiss building stock; performance has therefore become an essential element in mind.

____________________

In summary, while I fully acknowledge the preference for open-source tools in the Python ecosystem, the current use of AMPL provides significant advantages in terms of development effort, model expressiveness, and computational efficiency. Rewriting the entire package with an alternative would be a gigantic task, that would introduce complexity and reduce reliability. Given the current maturity of this AMPL implementation and its tangible benefits listed below, I believe maintaining it as the default setup will be beneficial in terms of stability and functionality.

From my research, I saw it is common in scientific computing and engineering to balance the use of proprietary software with open-source tools, especially when the proprietary tool offers superior performance or usability for specific tasks. Since REHO is already open-source, the openness of the overall tool is preserved, while the use of AMPL is isolated to the model description and optimization backend.

Finally, I think it's important to note that AMPL has decided to move towards greater accessibility with its Community Edition licence released last year. In fact, it was this event that encouraged me to make some effort for the REHO documentation and give the tool a greater visibility now that it could be used by anyone without restriction. In this context, I am confident that the availability of REHO as an open-source tool is compatible with its use of AMPL.

I hope I've convinced you of my commitment to this project and my desire of contribution to the open-source community.
