### 2024-10-22 #6734 @mbarzegary
Thank you @willu47 and @hgandhi2411 for finalizing your reviews, and also @DorsanL for a thorough clarification on AMPL. This was indeed an issue that we were discussing about in the pre-review phase as well.

@willu47 can you please let me know if the above explanation is acceptable so that you can mark the "Installation instruction" as completed in your review checklist? After that, I can proceed to accepting the submission.
