### 2024-11-04 aca9f91
read_geometry, debug extreme weather for HP
