+++ reho/plotting/plotting.py
@@ -1477,7 +1477,40 @@ def plot_composite_curve(df_Results, cluster, periods=["Yearly"], filename=None,
     else:
         return plt
 
-def plot_storage_profile(df_Results, resolution='daily'):
+def plot_storage_profile(df_Results, resolution='daily',storage_ID = "all"):
+
+    def plot_storage_sep(storage_SOC_tot,counter, fig, stor_var, items_average):
+        cm = {
+            "Electricity":"#a7a5a5",
+            "H2":"#0063a6",
+            "CH4":"#d7b652",
+            "CO2":"#D2363E"}
+
+
+        time_index = np.arange(0, 8760)
+
+        SOC_average = moving_average(storage_SOC_tot[stor_var], items_average)
+        time_index_average = moving_average(time_index, items_average)
+
+        mol = stor_var.split("_")[0]
+        if mol == "BAT":
+            mol = "Electricity"
+
+        fig.add_trace(go.Scatter(
+            x=time_index_average,
+            y=SOC_average,
+            mode='lines',
+            name=mol+" storage",
+            line=dict(color=cm[mol], width=0),
+            fill='tozeroy',
+            fillcolor="#0063a6"
+        ),
+            row = counter,
+            col = 1
+        )
+
+        return fig
+
     if resolution == 'monthly':
         items_average = 730
     elif resolution == 'weekly':
@@ -1487,33 +1520,52 @@ def plot_storage_profile(df_Results, resolution='daily'):
     else:
         items_average = 1
 
-    SOC = df_Results["df_storage"]["BAT_E_stored_IP"]
-    time_index = np.arange(0, 8760)
+    counter = 1
+    if storage_ID == 'all':
+        df_storage = df_Results["df_storage"].loc[:, (df_Results["df_storage"] != 0).any(axis=0)]
+        list_stor = list(df_storage.keys())
+        if len(list_stor) > 0:
+            storage_SOC_tot = df_storage.groupby(level=1).sum()
+            fig = make_subplots(rows=len(list_stor), cols=1, shared_xaxes=True, vertical_spacing=0.02)
+            for storage in list_stor:
+                fig = plot_storage_sep(storage_SOC_tot,counter,fig, storage, items_average)
+                if "CO2" in storage:
+                    fig.update_yaxes(title_text="SOC, mol", row=counter, col=1)
+                else:
+                    fig.update_yaxes(title_text="SOC, kWh", row=counter, col=1)
 
-    SOC_average = moving_average(SOC, items_average)
-    time_index_average = moving_average(time_index, items_average)
+                counter += 1
+    else:
+        if type(storage_ID) is list:
+            list_stor = storage_ID
+        else:
+            list_stor = [storage_ID]
+        df_storage = df_Results["df_storage"].loc[:, (df_Results["df_storage"] != 0).any(axis=0)]
+        if len(list_stor) > 0:
+            storage_SOC_tot = df_storage.groupby(level=1).sum()
+            fig = make_subplots(rows=len(list_stor), cols=1, shared_xaxes=True, vertical_spacing=0.02)
+            for storage in list_stor:
+                fig = plot_storage_sep(storage_SOC_tot,counter,fig, storage, items_average)
+                if "CO2" in storage:
+                    fig.update_yaxes(title_text="SOC, mol", row=counter, col=1)
+                else:
+                    fig.update_yaxes(title_text="SOC, kWh", row=counter, col=1)
 
-    fig = go.Figure()
-    fig.add_trace(go.Scatter(
-        x=time_index_average,
-        y=SOC_average,
-        mode='lines',
-        name='SOC',
-        line=dict(color='blue')
-    ))
+                counter+=1
 
-    fig.update_layout(
-        title='State of Charge Over Time',
[…]
