+++ reho/model/reho.py
@@ -446,9 +446,10 @@ class REHO(MasterProblem):
         df_Performance.loc['Network', 'ANN_factor'] = df_Performance['ANN_factor'][0]
 
         if self.method["actors_problem"]:
+            #TODO: Add variables
             df_actor = self.results_MP[Scn_ID][Pareto_ID][ids['Iter']]["df_District"][
                 ['C_op_renters_to_utility', 'C_op_renters_to_owners', 'C_op_utility_to_owners', 'owner_inv',
-                 'owner_portfolio']]
+                 'owner_portfolio', 'C_rent_fix', 'renter_expense']]
             df_Performance = pd.concat([df_Performance, df_actor], axis=1)
             df_Results["df_Actors_tariff"] = self.results_MP[Scn_ID][Pareto_ID][ids['Iter']]["df_Actors_tariff"]
             df_Results["df_Actors"] = self.results_MP[Scn_ID][Pareto_ID][ids['Iter']]["df_Actors"]
