+++ reho/model/actors_problem.py
@@ -25,11 +25,12 @@ class ActorsProblem(REHO):
     def __init__(self, qbuildings_data, units, grids, parameters=None, set_indexed=None, cluster=None, method=None, scenario=None, solver="highs", DW_params=None):
 
         super().__init__(qbuildings_data, units, grids, parameters, set_indexed, cluster, method, scenario, solver, DW_params)
-    
+
     def execute_actors_problem(self, n_sample=10, bounds=None, actor="Owners"):
         self.method["actors_problem"] = True
         self.method["include_all_solutions"] = True
         self.method['building-scale'] = True
+        # self.method['district-scale'] = True
         self.scenario["Objective"] = "TOTEX_bui"
         self.set_indexed["ActorObjective"] = np.array([actor])
         if bounds is not None:
@@ -41,7 +42,7 @@ class ActorsProblem(REHO):
             self.samples = samples.round(4)
         else:
             self.samples = pd.DataFrame([[None, None]] * n_sample, columns=['utility_portfolio', 'owner_portfolio'])
-    
+
         Scn_ID = self.scenario['name']
         self.pool = mp.Pool(mp.cpu_count())
         results = {ids: self.pool.apply_async(self.run_actors_optimization, args=(self.samples, ids)) for ids in self.samples.index}
@@ -49,6 +50,7 @@ class ActorsProblem(REHO):
             time.sleep(1)
         for ids in self.samples.index:
             df_Results, df_Results_MP, solver_attributes = results[ids].get()
+            solver_attributes = solver_attributes.droplevel('Iter').iloc[[-1]].copy()
             self.add_df_Results_MP(Scn_ID, ids, self.iter, df_Results_MP, solver_attributes)    # store results MP (pool.apply_async don't store it)
             self.add_df_Results(None, Scn_ID, ids, self.scenario)   # process results based on results MP
             self.get_KPIs(Scn_ID, ids)
@@ -57,11 +59,10 @@ class ActorsProblem(REHO):
         for i in self.results_MP[self.scenario["name"]]:
             if self.results_MP[self.scenario["name"]][i] is not None:
                 self.samples.loc[i, "objective"] = self.results_MP[self.scenario["name"]][i][0]["df_District"]["Objective"]["Network"]
-    
+
         self.pool.close()
-    
         gc.collect()  # free memory
-    
+
     def run_actors_optimization(self, samples, ids):
         if any([samples[col][0] for col in samples]):  # if samples contain values
             param = samples.iloc[ids]
@@ -74,7 +75,7 @@ class ActorsProblem(REHO):
             return self.results[scn][0], self.results_MP[scn][0][self.iter], self.solver_attributes_MP
         except:
             return None, None
-    
+
     def read_configurations(self, path=None):
         if path is None:
             filename = 'tr_' + str(self.buildings_data["Building1"]["transformer"]) + '_' + str(len(self.buildings_data)) + '.pickle'
@@ -132,3 +133,42 @@ class ActorsProblem(REHO):
         filename = 'tr_' + str(self.buildings_data["Building1"]["transformer"]) + '_' + str(len(self.buildings_data)) + '.pickle'
         writer = open(os.path.join(path_to_configurations, filename), 'wb')
         pickle.dump([self.results_SP, self.feasible_solutions, self.number_SP_solutions], writer)
+
+    def set_actors_boundary(self, bounds, n_sample=1):
+        sampler = qmc.LatinHypercube(d=2)
[…]
