+++ reho/paths.py
@@ -47,6 +47,10 @@ path_to_cenpts = os.path.join(path_to_skydome, 'skyPatchesCenPts.txt')  # locati
 
 path_to_weather = os.path.join(path_to_data, 'weather')
 
+path_to_actor = os.path.join(path_to_data, 'actor')
+path_to_income = os.path.join(path_to_actor, 'income_percentile.csv')
+path_to_rent = os.path.join(path_to_actor, 'rent_proportion.csv')
+
 # scripts specific paths
 path_to_clustering = os.path.join(os.getcwd(), 'data', 'clustering')
 path_to_configurations = os.path.join(os.getcwd(), 'results', 'configurations')

[…]
