+++ reho/model/sub_problem.py
@@ -169,13 +169,15 @@ class SubProblem:
                 ampl.read('pv_orientation.mod')
             else:
                 ampl.read('pv.mod')
-
         if 'rSOC' in self.infrastructure_sp.UnitTypes:
             # ampl.read('rSOC_district.mod')
             ampl.read('rSOC.mod')
-
-        if "Methanizer" in self.infrastructure_sp.UnitTypes:
-            ampl.read('methanizer.mod')
+        if "Methanator" in self.infrastructure_sp.UnitTypes:
+            ampl.read('methanator.mod')
+        if 'FuelCell' in self.infrastructure_sp.UnitTypes:
+            ampl.read('fuel_cell.mod')
+        if 'Electrolyzer' in self.infrastructure_sp.UnitTypes:
+            ampl.read('electrolyzer.mod')
 
         # district Units
         if 'EV' in self.infrastructure_sp.UnitTypes:
@@ -187,55 +189,34 @@ class SubProblem:
             ampl.read('heatstorage.mod')
         if 'WaterTankDHW' in self.infrastructure_sp.UnitTypes:
             ampl.read('dhwstorage.mod')
-        if 'BESS_IP_district' in self.infrastructure_sp.UnitTypes:
-            ampl.read('battery_interperiod_district.mod')
         if 'Battery' in self.infrastructure_sp.UnitTypes:
             ampl.read('battery.mod')
-        ampl.cd(path_to_ampl_model)
 
+        ampl.cd(path_to_ampl_model)
         # Objectives, epsilon constraints and specific constraints
         ampl.read('scenario.mod')
 
-        # TODO: integrate all storage units into infrastructure (avoid using ampl eval)
-
         if self.method_sp['use_Storage_Interperiod']:
-            """
-            ampl.eval('set UnitsOfStorage := setof{u in UnitsOfType["BESS_IP"]} u;')
-            ampl.eval('set UnitsOfStorage := setof{u in UnitsOfType["Battery_interperiod"]} u;')
-            'union UnitsOfType["PTES_storage"]'
-            'union UnitsOfType["PTES_conversion"] union UnitsOfType["CH4storage"]'
-            'union UnitsOfType["H2storage"] union UnitsOfType["SOEFC"]'
-            'union UnitsOfType["Methanizer"] union UnitsOfType["FuelCell"]'
-            'union UnitsOfType["Electrolyzer"] union UnitsOfType["WaterTankSH_interperiod"]'
-            'union UnitsOfType["SolidLiquidLHS"]'
-            '} u;')
-            """
 
             # Storage Units
             ampl.cd(path_to_units_storage)
 
             if 'Battery_interperiod' in self.infrastructure_sp.UnitTypes:
-                ampl.read('battery_interperiod_v1.mod')
+                ampl.read('battery_interperiod.mod')
 
             # ampl.read('h2_storage.mod')
             if 'H2storage' in self.infrastructure_sp.UnitTypes:
-                ampl.read('STORAGE_H2.mod')
+                ampl.read('H2storage.mod')
             if 'CH4storage' in self.infrastructure_sp.UnitTypes:
-                ampl.read('STORAGE_CH4.mod')
+                ampl.read('CH4storage.mod')
             if 'CO2storage' in self.infrastructure_sp.UnitTypes:
-                ampl.read('STORAGE_CO2.mod')
+                ampl.read('CO2storage.mod')
+
             #if 'WaterTankSH_interperiod' in self.infrastructure_sp.UnitTypes:
             #    ampl.read('heatstorage_interperiod.mod')
 
-
             # ampl.read('heatstorage_interperiod.mod')
             # ampl.read('LHS_storage.mod')
-
[…]
