### 2024-11-07 146c3c4
SA with DEM, %cars and % EV
