### 2024-11-07 5b2fbcd
update SA and parameters split for SPs
