### 2024-11-07 d5294f5
remove actors example (still under construction)
