### 2024-11-07 ce3f99f
diagrams update, paper typos
