+++ docs/sections/3_Model.rst
@@ -8,9 +8,9 @@ diverse end use demands through building-level equipment and district-level infr
 For a delimited perimeter of buildings, REHO selects the optimal energy system configuration minimizing the specified objective function.
 All the energy flows at building-level and district-level are then fully characterized by the model decision variables.
 
-.. figure:: ../images/district_documentation.svg
+.. figure:: ../images/district.svg
    :align: center
-   :name: district_documentation
+   :name: district
 
    District-level energy hub model in REHO
 
@@ -26,16 +26,16 @@ Energy can be stored in installed equipment (such as a battery or a water tank),
 Photovoltaic panels act as a renewable energy source.
 The building-level energy system is interconnected to the energy distribution infrastructure of the district (electrical grid, natural gas grid, ...).
 
-.. figure:: ../images/diagram_model.svg
+.. figure:: ../images/model.svg
    :align: center
-   :name: fig-diagram_model
+   :name: fig-model
 
    REHO model architecture
 
-.. figure:: ../images/building_energy_hub.png
+.. figure:: ../images/building.png
    :width: 450
    :align: center
-   :name: building_energy_hub
+   :name: building
 
    Building-level energy hub in REHO
 
@@ -315,7 +315,7 @@ List of symbols
 Inputs
 ===========================
 
-For the application of REHO, the energy hub description needs to contain - as highlighted by :ref:`fig-diagram_model` :
+For the application of REHO, the energy hub description needs to contain - as highlighted by :ref:`fig-model` :
 
 - the *End Use Demands (EUDs)*, from the meteorological data and the buildings characteristics,
 - the resources to which it has access to provide those *EUDs*, namely the grids,
@@ -750,15 +750,15 @@ Grid capacity
 The maximum capacity of the local low-voltage transformer is considered.
 The electricity export and the import is constrained within the feasibility range of the transformer.
 
-.. _network_diagram:
+.. _network:
 
-.. figure:: ../images/network_diagram.svg
+.. figure:: ../images/network.svg
    :align: center
 
    Energy flows and network constraints in REHO
 
 
-:ref:`network_diagram` distinguishes the:
+:ref:`network` distinguishes the:
 
 - Grid = energy flows within the district boundary
 - Network = exchanges with the district exterior, through the interface (transformer perspective)
