### 2024-11-07 #6734 @DorsanL
Thank you to the 3 reviewers for completing their reviews. 🤝

@mbarzegary thank you for your suggestions, I merged your PR. ✅

I have carefully addressed each point requested:
- the latest version 1.1.6 has been deployed on PyPI: https://pypi.org/project/REHO/
- the GitHub release is visible here: https://github.com/IPESE/REHO/releases/tag/v1.1.6
- a Zenodo archive has been published: https://zenodo.org/records/14052179 (DOI: 10.5281/zenodo.14052179)

Thank you for all the follow-up!
