### 2024-11-13 1f123b8
SA script for mobility paper
