+++ scripts/solene/functions.py
@@ -1,0 +1,332 @@
+import pandas as pd
+from reho.paths import *
+import matplotlib.pyplot as plt
+from plotly.subplots import make_subplots
+import plotly.express as px
+import plotly.graph_objects as go
+
+
+def linear_split_bin_table(df,col,lowerbound = None, upperbound = None):
+    """
+    Parameters
+    col : the columns to apply the change upon
+
+    """
+    if upperbound:
+        if upperbound in df.index:
+            df_up = df.loc[df.index <= upperbound,:].copy()
+        else:
+            df_up = df.loc[df.index < upperbound,:].copy()
+            last_bin = df.loc[df.index > upperbound,:].iloc[0]
+            last_bin[col] = (upperbound - df_up.index[-1]) / (last_bin.name - df_up.index[-1] ) * last_bin[col]
+            last_bin.name = upperbound
+            df_up = pd.concat([df_up,last_bin.to_frame().T])
+    
+    if lowerbound:
+        if upperbound:
+            df_up = df_up.loc[df_up.index > lowerbound,:].copy()
+        else:
+            df_up = df.loc[df.index > lowerbound,:].copy()
+
+        if not lowerbound in df.index:
+            prev_bin = df.loc[df.index < lowerbound,:].iloc[-1]
+            weight = (df_up.index[0] - lowerbound) / (df_up.index[0] - prev_bin.name) * df_up.iloc[0][col]
+            df_up.at[df_up.index[0],col] = weight
+
+    return df_up
+
+
+def mobility_demand_from_WP1data_modes(DailyDist,max_dist = 70 ,nbins = 1,modalwindow = 0.01,  share_cars = None,share_EV_infleet = None ):
+    """
+    Donne les shares pour les modes MD/PT/cars. Voir twin function mobility_demand_from_WP1data_units qui donne pour les units et pas pour les modes
+
+    Parameters
+    -----------
+    DailyDist : DailyDist total
+    max_dist :
+
+    """
+    parameters = {"DailyDist" : {},
+                }
+    modalshares_csv = pd.DataFrame()
+    modalshares_custom = pd.DataFrame()
+
+    df_dist = pd.read_csv(r"C:\Users\Public\REHO2024\scripts\solene\data\mobility\WP1_distributionofdistances.csv",index_col = 0)
+    df_modalshares = pd.read_csv(r"C:\Users\Public\REHO2024\scripts\solene\data\mobility\WP1_modalsharesbydistances_work.tsv",sep = "\t",index_col = 0)
+    
+    # mapping_modesunits = {
+    #     "Marche"   :  "Bike_district",
+    #     "Vélo"        : "Bike_district",
+    #     "moto"        : "ICE_district",
+    #     "voiture"     : "ICE_district",
+    #     "tram/bus"    : "PT_bus",
+    #     "train" : "PT_train"
+    # }
+
+    mapping_modesunits = {
+        "Marche"   :  "MD",
+        "Vélo"        : "MD",
+        "moto"        : "cars",
+        "voiture"     : "cars",
+        "tram/bus"    : "PT",
+        "train" : "PT"
+    }
+
+    df_dist_inf = linear_split_bin_table(df_dist,"weight",upperbound=max_dist)
+    df_dist_inf.weight = df_dist_inf.weight /df_dist_inf.weight.sum()
+
+    df_dist_inf['up'] = df_dist_inf.index
+    df_dist_inf['low'] =   df_dist_inf.up.shift(1).fillna(0)
+    df_dist_inf['mean'] = (df_dist_inf.up + df_dist_inf.low)/2
[…]
