### 2024-11-19 61dbb40
debug 2 for SA
