+++ reho/model/reho.py
@@ -694,8 +694,9 @@ class REHO(MasterProblem):
         # df_Unit_t
         df_Unit_t = self.get_final_SPs_results(MP_selection, 'df_Unit_t')
         df_Unit_t = df_Unit_t.droplevel(['Scn_ID', 'Pareto_ID', 'Iter', 'FeasibleSolution', 'house'])
-        df_district_units = last_results["df_Unit_t"]
-        df_Unit_t = pd.concat([df_Unit_t, df_district_units])
+        if "df_Unit_t" in last_results.keys():
+            df_district_units = last_results["df_Unit_t"]
+            df_Unit_t = pd.concat([df_Unit_t, df_district_units])
         df_Results["df_Unit_t"] = df_Unit_t
 
         if self.method["save_streams"]:
