+++ scripts/solene/SA_cooptimization.py
@@ -1,8 +1,6 @@
 from reho.model.postprocessing.sensitivity_analysis import *
-from reho.plotting import plotting
 from reho.model.preprocessing.mobility_generator import *
 import datetime
-
 from scripts.solene.functions import *
 
 if __name__ == '__main__':
@@ -10,53 +8,42 @@ if __name__ == '__main__':
 
     districts = [234,3112,837]
     n_buildings = 2
-    district_parameters = { # DUMMY VALUES FOR NOW
-        234 : {"N_house" : 129,"rho" : pd.Series([95,3,2],index=['household','industry','service']), "f" : 1 , "Scluster" : 94322 },
-        3112 : {"N_house" : 200,"rho" : pd.Series([89,2,6],index=['household','industry','service']), "f" : 1 , "Scluster" : 110399 },
-        837 : {"N_house" : 56,"rho" : pd.Series([70,10,20],index=['household','industry','service']), "f" : 1 , "Scluster" : 56540 }
-    }
-
+    district_parameters = {234 : {"N_house": 129, "rho": pd.Series([95,3,2],index=['household','industry','service']), "f" : 1 , "Scluster" : 94322 },
+                           3112 : {"N_house": 200, "rho": pd.Series([89,2,6],index=['household','industry','service']), "f" : 1 , "Scluster" : 110399 },
+                           837 : {"N_house": 56, "rho": pd.Series([70,10,20],index=['household','industry','service']), "f" : 1 , "Scluster" : 56540 }}
     district_parameters = compute_district_parameters(district_parameters)
-
     for k in district_parameters.keys():
         district_parameters[k]['Population'] = n_buildings * district_parameters[k]['PopHouse']
 
-
-    # GETTING THE SAMPLED PARAMETERS ================================================================================
-
     reader = QBuildingsReader()
     reader.establish_connection('Suisse')
-    qbuildings_data = reader.read_db(transformer=234, nb_buildings=2)
+    qbuildings_data = reader.read_db(transformer=234, nb_buildings=n_buildings)
     cluster = {'Location': 'Geneva', 'Attributes': ['T', 'I', 'W'], 'Periods': 10, 'PeriodDuration': 24}
 
     scenario = dict()
     scenario['Objective'] = 'TOTEX'
     scenario['name'] = 'totex'
-    scenario['exclude_units'] = [ 'NG_Cogeneration','Battery']
+    scenario['exclude_units'] = [ 'NG_Cogeneration', 'Battery']
     scenario['enforce_units'] = ['EV_district']
     scenario['specific'] = []
     scenario['EMOO'] = {}
-    grids = infrastructure.initialize_grids({'Electricity': {},
-                                             'NaturalGas': {},
-                                             'FossilFuel': {},
-                                             'Mobility': {},
-                                             })
-    units = infrastructure.initialize_units(scenario, grids,district_data = True)
-    method = {'building-scale': True} #coordination entre batiments ou non 
+    grids = infrastructure.initialize_grids({'Electricity': {}, 'NaturalGas': {}, 'FossilFuel': {}, 'Mobility': {}})
+    units = infrastructure.initialize_units(scenario, grids,district_data=True)
+    method = {'building-scale': True} # coordination entre batiments ou non
 
     # SA object used to generate the sampled values
     reho = REHO(qbuildings_data=qbuildings_data, units=units, grids=grids, cluster=cluster, scenario=scenario, method=method, solver="gurobi")
-    SA  = SensitivityAnalysis(reho, SA_type="Monte_Carlo", sampling_parameters=8)
[…]
